Bernoulli

Bernoulli

Arguments

input
(Tensor) the input tensor of probability values for the Bernoulli distribution

generator
(torch.Generator, optional) a pseudorandom number generator for sampling

out
(Tensor, optional) the output tensor.

bernoulli(input, *, generator=None, out=None) -> Tensor

Draws binary random numbers (0 or 1) from a Bernoulli distribution.

The input tensor should be a tensor containing probabilities to be used for drawing the binary random number. Hence, all values in input have to be in the range: \(0 \leq \text{input}_i \leq 1\).

The \(\text{i}^{th}\) element of the output tensor will draw a value \(1\) according to the \(\text{i}^{th}\) probability value given in input.

$$ \text{out}_{i} \sim \mathrm{Bernoulli}(p = \text{input}_{i}) $$ The returned out tensor only has values 0 or 1 and is of the same shape as input.

out can have integral dtype, but input must have floating point dtype.

Examples

a = torch_empty(c(3, 3))$uniform_(0, 1)  # generate a uniform random matrix with range c(0, 1)
a

#> torch_tensor 
#>  0.4666  0.6223  0.0293
#>  0.3311  0.3295  0.4024
#>  0.4563  0.5072  0.7706
#> [ CPUFloatType{3,3} ]

torch_bernoulli(a)

#> torch_tensor 
#>  0  0  0
#>  0  0  1
#>  1  1  0
#> [ CPUFloatType{3,3} ]

a = torch_ones(c(3, 3)) # probability of drawing "1" is 1
torch_bernoulli(a)

#> torch_tensor 
#>  1  1  1
#>  1  1  1
#>  1  1  1
#> [ CPUFloatType{3,3} ]

a = torch_zeros(c(3, 3)) # probability of drawing "1" is 0
torch_bernoulli(a)

#> torch_tensor 
#>  0  0  0
#>  0  0  0
#>  0  0  0
#> [ CPUFloatType{3,3} ]