Clamp

Clamp

Arguments

input
(Tensor) the input tensor.

min
(Number) lower-bound of the range to be clamped to

max
(Number) upper-bound of the range to be clamped to

out
(Tensor, optional) the output tensor.

value
(Number) minimal value of each element in the output

clamp(input, min, max, out=None) -> Tensor

Clamp all elements in input into the range [ min, max ] and return a resulting tensor:

$$ y_i = \begin{cases} \text{min} & \text{if } x_i < \text{min} \\ x_i & \text{if } \text{min} \leq x_i \leq \text{max} \\ \text{max} & \text{if } x_i > \text{max} \end{cases} $$ If input is of type FloatTensor or DoubleTensor, args min and max must be real numbers, otherwise they should be integers.

clamp(input, *, min, out=None) -> Tensor

Clamps all elements in input to be larger or equal min.

If input is of type FloatTensor or DoubleTensor, value should be a real number, otherwise it should be an integer.

clamp(input, *, max, out=None) -> Tensor

Clamps all elements in input to be smaller or equal max.

If input is of type FloatTensor or DoubleTensor, value should be a real number, otherwise it should be an integer.

Examples

a = torch_randn(c(4))
a

#> torch_tensor 
#> -0.8773
#> -0.5994
#> -0.3199
#> -0.1252
#> [ CPUFloatType{4} ]

torch_clamp(a, min=-0.5, max=0.5)

#> torch_tensor 
#> -0.5000
#> -0.5000
#> -0.3199
#> -0.1252
#> [ CPUFloatType{4} ]

a = torch_randn(c(4))
a

#> torch_tensor 
#>  0.3435
#>  0.2023
#>  1.2560
#>  0.8784
#> [ CPUFloatType{4} ]

torch_clamp(a, min=0.5)

#> torch_tensor 
#>  0.5000
#>  0.5000
#>  1.2560
#>  0.8784
#> [ CPUFloatType{4} ]

a = torch_randn(c(4))
a

#> torch_tensor 
#> -0.0530
#> -1.0853
#>  1.8960
#> -1.0185
#> [ CPUFloatType{4} ]

torch_clamp(a, max=0.5)

#> torch_tensor 
#> -0.0530
#> -1.0853
#>  0.5000
#> -1.0185
#> [ CPUFloatType{4} ]