Div

Div

Arguments

input
(Tensor) the input tensor.

other
(Number) the number to be divided to each element of input

div(input, other, out=None) -> Tensor

Divides each element of the input input with the scalar other and returns a new resulting tensor.

$$ \text{out}_i = \frac{\text{input}_i}{\text{other}} $$ If the torch_dtype of input and other differ, the torch_dtype of the result tensor is determined following rules described in the type promotion documentation . If out is specified, the result must be castable to the torch_dtype of the specified output tensor. Integral division by zero leads to undefined behavior.

Each element of the tensor input is divided by each element of the tensor other. The resulting tensor is returned.

$$ \text{out}_i = \frac{\text{input}_i}{\text{other}_i} $$ The shapes of input and other must be broadcastable . If the torch_dtype of input and other differ, the torch_dtype of the result tensor is determined following rules described in the type promotion documentation . If out is specified, the result must be castable to the torch_dtype of the specified output tensor. Integral division by zero leads to undefined behavior.

Examples

a = torch_randn(c(5))
a

#> torch_tensor 
#> -0.8682
#>  1.4319
#>  0.0284
#>  0.3992
#> -0.8106
#> [ CPUFloatType{5} ]

torch_div(a, 0.5)

#> torch_tensor 
#> -1.7364
#>  2.8638
#>  0.0567
#>  0.7984
#> -1.6213
#> [ CPUFloatType{5} ]

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#> -1.1357  0.5563  0.6955 -1.2786
#>  0.4258 -1.0868  1.3424 -0.2401
#>  1.4080 -0.5536 -0.3271 -1.1902
#> -0.1756  0.6139 -0.7103  0.8725
#> [ CPUFloatType{4,4} ]

b = torch_randn(c(4))
b

#> torch_tensor 
#> -0.4679
#>  0.2458
#> -0.2811
#> -0.6740
#> [ CPUFloatType{4} ]

torch_div(a, b)

#> torch_tensor 
#>  2.4274  2.2636 -2.4742  1.8969
#> -0.9101 -4.4221 -4.7757  0.3562
#> -3.0093 -2.2527  1.1638  1.7658
#>  0.3754  2.4982  2.5272 -1.2945
#> [ CPUFloatType{4,4} ]