Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.4047  1.3043
#>   5.8192 -4.6152
#>  -2.3986 -3.7090
#> 
#> (2,.,.) = 
#>  -1.2640 -1.4408
#>   4.4458  1.3136
#>   3.5468  1.1900
#> 
#> (3,.,.) = 
#>  -2.5775  1.8687
#>   4.2696 -2.6635
#>   1.7024  2.1248
#> 
#> (4,.,.) = 
#>  -6.2194  0.9395
#>  -6.2945 -3.8405
#>  -7.7372  0.2664
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.6164  0.6679
#>   2.0600 -2.4514
#>  -1.2216 -0.0320
#> 
#> (2,.,.) = 
#>   1.1383  1.0977
#>  -0.9011  2.1971
#>  -1.2561  1.3625
#> 
#> (3,.,.) = 
#>   1.1253  0.9186
#>   2.9844 -1.1880
#>   0.8736 -0.7602
#> 
#> (4,.,.) = 
#>  -0.0519 -1.3799
#>   1.6759 -3.1730
#>  -0.7944 -4.2794
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   -4.1434   7.0701
#>    4.1283  10.2848
#>    4.6852   1.7205
#>   -9.0529   9.7753
#>   -3.4198  -3.9566
#> 
#> (2,1,1,.,.) = 
#>   0.3713 -8.4854
#>  -0.7390  1.2517
#>  -6.5925  5.4276
#>  -0.5410 -5.2995
#>  -2.1654  5.5068
#> 
#> (3,1,1,.,.) = 
#>   -3.6679   2.2298
#>    1.8170   8.4524
#>    3.7806  -4.1507
#>   -1.6525  11.0216
#>   -5.9397  11.5484
#> 
#> (1,2,1,.,.) = 
#>   0.5332  3.7269
#>  -2.9550  0.5212
#>   0.1108 -0.0696
#>  -1.1644 -4.3266
#>   4.9107  0.3265
#> 
#> (2,2,1,.,.) = 
#>   2.7827 -0.5611
#>   4.6510  4.7550
#>  -6.6286 -2.4118
#>  -9.6534  2.3991
#>  -3.4302 -3.6442
#> 
#> (3,2,1,.,.) = 
#>   9.0893 -10.3659
#>   0.6097  0.4505
#>   5.6303 -4.2214
#>  -7.3809 -8.5110
#>   0.4500  0.0070
#> 
#> (1,3,1,.,.) = 
#>  -4.2899 -6.2850
#>  -5.2426 -2.0440
#>  -7.4942  3.9191
#>   5.9185  5.7432
#>   7.2063 -1.3311
#> 
#> (2,3,1,.,.) = 
#>   4.9742 -5.8157
#>  -2.2930  9.5255
#>  -3.7235 -7.4687
#>   0.2767 -5.9362
#>  -10.5351  2.6712
#> 
#> (3,3,1,.,.) = 
#>   0.1449  0.6357
#>   4.4132  3.2738
#>  -8.4976 -10.8212
#>   1.9415  5.7853
#>   4.1209  6.6877
#> 
#> (1,1,2,.,.) = 
#>    2.7944  10.2502
#>   -2.6382   2.5856
#>    0.4772  -1.5225
#>    0.5480  -1.0840
#>    0.1325  -2.7438
#> 
#> (2,1,2,.,.) = 
#>    6.2578  -0.6948
#>   -2.2649   2.0072
#>   11.0929   7.8163
#>   -1.9270  -0.2111
#>   12.1356  -1.8140
#> 
#> (3,1,2,.,.) = 
#>   -5.7680  -3.2596
#>    3.1769   1.2258
#>   -0.1269  -7.1658
#>   11.6463  -1.1832
#>    4.4600  -7.2732
#> 
#> (1,2,2,.,.) = 
#>  -1.3298 -2.8651
#>   2.4784 -9.4233
#>  -0.1348  0.0955
#>  -1.6568  4.2689
#>  -1.3159  0.5445
#> 
#> (2,2,2,.,.) = 
#>   -1.5045  -3.3074
#>   -1.6387   1.3416
#>   -1.3955   9.5188
#>    3.3928   4.7425
#>   10.0942   0.2618
#> 
#> (3,2,2,.,.) = 
#>  -1.4006 -4.4934
#>  -4.3818 -1.9495
#>   3.7270 -0.1141
#>  -1.9584 -3.8504
#>  -4.5234 -3.8022
#> 
#> (1,3,2,.,.) = 
#>   4.3224  0.9749
#>  -3.6135  0.8805
#>  -5.7271 -3.2207
#>   6.3817  4.1217
#>  -1.2512 -14.5058
#> 
#> (2,3,2,.,.) = 
#>  -1.5442 -0.7721
#>  -1.3867 -3.2204
#>   0.3348  3.7756
#>  -3.9844 -8.9147
#>  -6.2723  4.9882
#> 
#> (3,3,2,.,.) = 
#>  -2.7711 -4.7008
#>   0.7715 -2.2512
#>   3.5882 -6.2813
#>   3.8047 -0.0679
#>  -3.3465  0.1180
#> 
#> (1,1,3,.,.) = 
#>    2.5197   0.5154
#>   10.1644   1.8316
#>   -2.4873  -4.2751
#>   -5.7731  -0.4932
#>   -4.3769  -3.4219
#> 
#> (2,1,3,.,.) = 
#>  -1.8959  0.1350
#>   5.4780 -9.0621
#>   2.0695 -1.6297
#>  -2.2507 -2.6596
#>  -7.7881 -7.6886
#> 
#> (3,1,3,.,.) = 
#>   -7.9515   1.4164
#>   12.0891   4.3152
#>   -8.2487   6.0854
#>   -0.4767  -5.9527
#>    9.6458   1.8879
#> 
#> (1,2,3,.,.) = 
#>   1.0717 -0.5952
#>   4.2871 -4.3763
#>  -7.9670 -8.7686
#>  -0.6896  8.4010
#>  -5.6339  3.1826
#> 
#> (2,2,3,.,.) = 
#>   -1.9312  -7.9856
#>    9.6076   0.2346
#>   -5.7809  -0.5108
#>    5.2225  11.2391
#>    1.5432   6.1599
#> 
#> (3,2,3,.,.) = 
#>  -3.3317 -0.7868
#>  -4.6704 -1.7096
#>  -5.6420  0.0205
#>   3.2218  4.6299
#>  -3.8106  1.0523
#> 
#> (1,3,3,.,.) = 
#>   6.3116 -2.9499
#>   1.9619  1.1135
#>  -4.0135 -4.6918
#>   6.7471  9.0695
#>   2.7886  2.0852
#> 
#> (2,3,3,.,.) = 
#>  -4.7534 -8.8724
#>   1.4729  2.2041
#>   4.5054  0.8199
#>  -4.8471 -2.5222
#>  -4.8468 -3.5371
#> 
#> (3,3,3,.,.) = 
#>   0.8117 -1.4793
#>   1.8113  5.0108
#>  -4.2890  7.8753
#>  -7.9320 -4.8054
#>   1.0105  1.3181
#> 
#> (1,1,4,.,.) = 
#>    3.2043   0.8774
#>    4.5022   3.6977
#>    0.6659  14.1211
#>    3.8720  -7.0694
#>   -0.5504  -1.2779
#> 
#> (2,1,4,.,.) = 
#>   0.7320  3.4057
#>  -2.8226  3.9954
#>  -0.8798 -0.1450
#>  -1.0296  2.2530
#>   2.1595 -1.4476
#> 
#> (3,1,4,.,.) = 
#>  -5.2558  6.2523
#>   4.3333 -2.4675
#>   0.8648  1.0994
#>  -5.1346  1.5157
#>   0.3120  5.8920
#> 
#> (1,2,4,.,.) = 
#>   12.3784  -0.0506
#>  -10.5742   2.1197
#>    5.0325   2.7410
#>   -0.3141  -8.2443
#>   -4.1050  -1.7742
#> 
#> (2,2,4,.,.) = 
#>  -1.0073 -6.4397
#>  -4.8384 -3.7420
#>   6.0159 -4.4275
#>   2.5899  2.5436
#>   6.3318  4.3016
#> 
#> (3,2,4,.,.) = 
#>  -7.6409 -4.1411
#>   2.2491  2.2985
#>   4.2850 -7.6930
#>   7.9590 -3.9549
#>   4.7552  5.8662
#> 
#> (1,3,4,.,.) = 
#>  -9.0599  1.4179
#>   0.5921 -2.3927
#>  -3.1475  1.4323
#>  -0.3698  7.6861
#>   1.0474  3.7092
#> 
#> (2,3,4,.,.) = 
#>   2.2169 -1.1596
#>  -2.1680  0.9122
#>  -7.7383  5.5874
#>  -3.7544  4.8578
#>  -1.6724 -7.2719
#> 
#> (3,3,4,.,.) = 
#>    2.7489  -0.9676
#>   10.6979   1.5089
#>   -4.4474 -12.5104
#>   -9.2893  -1.6349
#>    1.2395   3.1945
#> 
#> (1,1,5,.,.) = 
#>   5.1813  2.1716
#>  -11.1962 -2.5415
#>   3.7535 -3.0715
#>   9.0859 -4.7127
#>   5.4954 -13.6242
#> 
#> (2,1,5,.,.) = 
#>  -2.2058 -3.5267
#>  -1.7985  7.2834
#>  -4.1551  0.0798
#>   3.3272  5.3555
#>  -5.1084  5.5254
#> 
#> (3,1,5,.,.) = 
#>  -6.0070 -0.0692
#>   0.1380  6.5265
#>   1.1382  5.8019
#>   9.1053  1.2792
#>  -4.7555 -1.0759
#> 
#> (1,2,5,.,.) = 
#>  -4.9907 -7.6187
#>   5.0010  6.6394
#>   6.2825 -5.5056
#>   2.7132  3.6739
#>  -1.9792  7.3512
#> 
#> (2,2,5,.,.) = 
#>  -1.3670  1.4575
#>  -1.8807 -1.0125
#>   0.2462 -1.2225
#>   2.3256 -8.0305
#>   2.0745  5.2149
#> 
#> (3,2,5,.,.) = 
#>  -1.4578 -2.1325
#>   1.8978 -2.4055
#>  -9.5748 -0.4570
#>   3.6611  2.9278
#>  -1.5334  1.4253
#> 
#> (1,3,5,.,.) = 
#>  -2.3917 -1.8407
#>  -0.2516  1.1492
#>  -2.1970  3.7725
#>  -3.3630  1.9469
#>  -1.2893  0.9953
#> 
#> (2,3,5,.,.) = 
#>   4.6252 -2.5233
#>  -0.3591  3.0850
#>   8.0371  6.2925
#>   4.3740  5.0486
#>   8.8902  1.4243
#> 
#> (3,3,5,.,.) = 
#>    9.4911  -7.1369
#>   -2.5986  10.2831
#>    0.1577   0.4197
#>    0.9673   3.4087
#>   -7.6229  -9.2692
#> [ CPUFloatType{3,3,5,5,2} ]