Create a Generator object

A torch_generator is an object which manages the state of the algorithm that produces pseudo random numbers. Used as a keyword argument in many In-place random sampling functions.

torch_generator()

Examples

# Via string
generator <- torch_generator()
generator$current_seed()

#> Loading required package: bit64

#> Loading required package: bit

#> Attaching package bit

#> package:bit (c) 2008-2012 Jens Oehlschlaegel (GPL-2)

#> creators: bit bitwhich

#> coercion: as.logical as.integer as.bit as.bitwhich which

#> operator: ! & | xor != ==

#> querying: print length any all min max range sum summary

#> bit access: length<- [ [<- [[ [[<-

#> for more help type ?bit

#> 
#> Attaching package: ‘bit’

#> The following object is masked from ‘package:base’:
#> 
#>     xor

#> Attaching package bit64

#> package:bit64 (c) 2011-2012 Jens Oehlschlaegel

#> creators: integer64 seq :

#> coercion: as.integer64 as.vector as.logical as.integer as.double as.character as.bin

#> logical operator: ! & | xor != == < <= >= >

#> arithmetic operator: + - * / %/% %% ^

#> math: sign abs sqrt log log2 log10

#> math: floor ceiling trunc round

#> querying: is.integer64 is.vector [is.atomic} [length] format print str

#> values: is.na is.nan is.finite is.infinite

#> aggregation: any all min max range sum prod

#> cumulation: diff cummin cummax cumsum cumprod

#> access: length<- [ [<- [[ [[<-

#> combine: c rep cbind rbind as.data.frame

#> WARNING don't use as subscripts

#> WARNING semantics differ from integer

#> for more help type ?bit64

#> 
#> Attaching package: ‘bit64’

#> The following object is masked from ‘package:bit’:
#> 
#>     still.identical

#> The following objects are masked from ‘package:base’:
#> 
#>     :, %in%, is.double, match, order, rank

#> integer64
#> [1] 67280421310721

generator$set_current_seed(1234567L)
generator$current_seed()

#> integer64
#> [1] 1234567