Irfft

Irfft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

onesided
(bool, optional) controls whether input was halfed to avoid redundancy, e.g., by rfft. Default: True

signal_sizes
(list or torch.Size, optional) the size of the original signal (without batch dimension). Default: None

Note

Due to the conjugate symmetry, `input` do not need to contain the full
complex frequency values. Roughly half of the values will be sufficient, as
is the case when `input` is given by [`~torch.rfft`] with
``rfft(signal, onesided=True)``. In such case, set the `onesided`
argument of this method to ``True``. Moreover, the original signal shape
information can sometimes be lost, optionally set `signal_sizes` to be
the size of the original signal (without the batch dimensions if in batched
mode) to recover it with correct shape.

Therefore, to invert an [`~torch.rfft`], the `normalized` and
`onesided` arguments should be set identically for [`~torch.irfft`],
and preferably a `signal_sizes` is given to avoid size mismatch. See the
example below for a case of size mismatch.

See [`~torch.rfft`] for details on conjugate symmetry.

The inverse of this function is ~torch.rfft.

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

irfft(input, signal_ndim, normalized=False, onesided=True, signal_sizes=None) -> Tensor

Complex-to-real Inverse Discrete Fourier Transform

This method computes the complex-to-real inverse discrete Fourier transform. It is mathematically equivalent with ifft with differences only in formats of the input and output.

The argument specifications are almost identical with ~torch.ifft. Similar to ~torch.ifft, if normalized is set to True, this normalizes the result by multiplying it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary, where \(N_i\) is the size of signal dimension \(i\).

Warning

Generally speaking, input to this function should contain values following conjugate symmetry. Note that even if onesided is True, often symmetry on some part is still needed. When this requirement is not satisfied, the behavior of ~torch.irfft is undefined. Since torch_autograd.gradcheck estimates numerical Jacobian with point perturbations, ~torch.irfft will almost certainly fail the check.

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

x = torch_randn(c(4, 4))
torch_rfft(x, 2, onesided=TRUE)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.3574  0.0000
#>  -2.2024  1.8518
#>  -3.0432  0.0000
#> 
#> (2,.,.) = 
#>  -0.5529  2.3965
#>  -1.8436 -2.1012
#>  -2.8422  0.0755
#> 
#> (3,.,.) = 
#>  -0.8523  0.0000
#>   1.4132 -1.8156
#>   3.4963  0.0000
#> 
#> (4,.,.) = 
#>  -0.5529 -2.3965
#>   2.3302  2.2482
#>  -2.8422 -0.0755
#> [ CPUFloatType{4,3,2} ]

x = torch_randn(c(4, 5))
torch_rfft(x, 2, onesided=TRUE)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.9500  0.0000
#>   1.8160 -0.4082
#>  -1.9780  4.3058
#> 
#> (2,.,.) = 
#>   1.8885  1.7305
#>   5.6626 -10.3351
#>   0.3566  1.4718
#> 
#> (3,.,.) = 
#>   7.3281  0.0000
#>  -1.2425  0.5933
#>   6.8105 -2.9378
#> 
#> (4,.,.) = 
#>   1.8885 -1.7305
#>  -1.3759  4.3716
#>  -2.4013 -0.0660
#> [ CPUFloatType{4,3,2} ]

y = torch_rfft(x, 2, onesided=TRUE)
torch_irfft(y, 2, onesided=TRUE, signal_sizes=c(4,5))  # recover x

#> torch_tensor 
#>  1.2726  0.8189  0.8042 -0.4027  0.0459
#>  0.1569 -0.3646 -1.7469 -2.9396  1.9595
#>  0.4465 -1.1236  0.2780  1.0080  0.0415
#> -2.1308  0.4868  1.3766 -0.4218 -0.5152
#> [ CPUFloatType{4,5} ]