Lstsq

Lstsq

Arguments

input
(Tensor) the matrix \(B\)

A
(Tensor) the \(m\) by \(n\) matrix \(A\)

out
(tuple, optional) the optional destination tensor

Note

The case when \eqn{m < n} is not supported on the GPU.

lstsq(input, A, out=None) -> Tensor

Computes the solution to the least squares and least norm problems for a full rank matrix \(A\) of size \((m \times n)\) and a matrix \(B\) of size \((m \times k)\).

If \(m \geq n\), lstsq solves the least-squares problem:

$$ \begin{array}{ll} \min_X & \|AX-B\|_2. \end{array} $$ If \(m < n\), lstsq solves the least-norm problem:

$$ \begin{array}{ll} \min_X & \|X\|_2 & \text{subject to} & AX = B. \end{array} $$ Returned tensor \(X\) has shape \((\max(m, n) \times k)\). The first \(n\) rows of \(X\) contains the solution. If \(m \geq n\), the residual sum of squares for the solution in each column is given by the sum of squares of elements in the remaining \(m - n\) rows of that column.

Examples

A = torch_tensor(rbind(
 c(1,1,1),
 c(2,3,4),
 c(3,5,2),
 c(4,2,5),
 c(5,4,3)
))
B = torch_tensor(rbind(
 c(-10, -3),
 c(12, 14),
 c(14, 12),
 c(16, 16),
 c(18, 16)
))
out = torch_lstsq(B, A)
out[[1]]

#> torch_tensor 
#>   2.0000   1.0000
#>   1.0000   1.0000
#>   1.0000   2.0000
#>  10.9635   4.8501
#>   8.9332   5.2418
#> [ CPUFloatType{5,2} ]