Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.04269
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.4629
#>  0.8272
#>  3.3230
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.6677  0.1788  0.1100
#> -0.9483 -0.6320 -0.2384
#>  0.4043  0.4897  1.1888
#> -2.9265  1.2784 -3.7981
#>  1.2350 -2.0582 -0.8987
#>  0.2674 -0.2361  1.9445
#>  2.6379  1.2707 -3.2039
#>  0.8476 -0.2293  0.1852
#> -0.5818 -1.5260 -1.1090
#>  0.0889 -0.1561  1.6968
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.3208 -2.0049  0.0031  1.4821  0.7776
#>  -1.3912 -1.3747 -1.7962  1.1029  0.2305
#>  -0.2800  2.1352 -7.8007  1.9999  3.5190
#> 
#> (2,.,.) = 
#>  -0.5397  0.6611  0.2352  0.9327 -0.3681
#>   3.5489  1.0260 -1.6102  1.5173 -2.8862
#>   1.4242  0.6877 -0.9000  0.1969 -1.6317
#> 
#> (3,.,.) = 
#>  -1.9569 -2.1040 -0.4612  0.1028 -0.9828
#>   1.5872  1.6451 -0.7184 -0.5117 -1.2345
#>   2.3066  3.6352  3.6576 -3.0932 -1.5319
#> 
#> (4,.,.) = 
#>   1.3860  0.5633 -2.2218 -0.9974 -0.0042
#>  -2.8575  2.5016 -0.1102 -2.9560  5.4582
#>   0.2938  0.6981 -2.1610 -0.4241 -0.2839
#> 
#> (5,.,.) = 
#>   0.8168 -0.6799  0.5988  0.1536 -0.0093
#>   3.1275 -1.7761  1.0287  0.0538 -0.3631
#>   3.2511 -0.9495  2.6802  0.3178  2.1521
#> 
#> (6,.,.) = 
#>   1.6689  3.7958 -2.9695  1.6781  4.6761
#>  -0.2899  0.9416  0.4293  1.4347 -0.3611
#>   0.9907  2.1351 -1.8957  1.4407  1.7778
#> 
#> (7,.,.) = 
#>  -0.5542 -0.3126  0.8788 -1.0230  0.6356
#>   0.2203 -0.7804  2.8507 -3.7470  0.3183
#>   1.1327  1.7075 -0.1281 -1.8078  0.0235
#> 
#> (8,.,.) = 
#>  -0.4351 -0.8151  0.3316 -0.6766  0.3662
#>  -0.0989  2.4965  3.1309  0.2347 -0.8686
#>  -0.9504  0.7816  3.3810 -1.2918 -0.9614
#> 
#> (9,.,.) = 
#>   2.1006  0.9474  1.7417 -2.9480  1.7784
#>  -0.4342 -0.3652  0.0694  0.5269 -0.5427
#>   0.3339 -4.6863  2.8291  4.8941 -3.0781
#> 
#> (10,.,.) = 
#>   2.3710  4.6403 -4.4895 -0.1243 -0.6281
#>   0.4697  1.0462 -1.6761  0.5185  0.4500
#>  -1.3095 -3.9811  2.3301 -1.2038 -0.5334
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.6487  0.7231  1.0536  0.0180  0.2486
#>   1.4458  1.6342  0.3699 -3.6597 -2.4003
#>   0.0668  0.3077 -0.4088 -1.0802  0.0592
#> 
#> (2,.,.) = 
#>  -0.1601  0.9107  0.0530 -2.5999 -2.0699
#>  -2.5280 -0.3227 -0.4990 -0.7435 -1.7564
#>   0.2254 -0.3270  0.4326  1.7600  2.6848
#> 
#> (3,.,.) = 
#>   1.1077 -0.2651 -0.1701  1.2637  2.3495
#>  -0.0929 -2.0234  2.4172  6.8473 -0.9843
#>   1.7069  1.4915  1.2432 -2.5685 -2.9917
#> 
#> (4,.,.) = 
#>   3.7518 -1.0557  2.6951  5.9052 -0.8544
#>   2.0284 -0.0328  1.6261  1.3372 -4.1401
#>   4.0313  0.5814  1.8204  1.3860 -0.1009
#> 
#> (5,.,.) = 
#>  -4.6061 -0.8816 -1.1351  0.1281  1.8817
#>  -0.1058  0.4131 -0.1894 -1.5474 -1.8103
#>  -2.0629 -2.0016 -0.8285  3.8739  1.9679
#> 
#> (6,.,.) = 
#>   0.1004  0.5376  2.7644  0.0454 -6.0572
#>   2.4445  0.6556 -0.8564 -1.5156  0.6425
#>   2.8848 -0.6509  1.9059  3.6418 -2.5365
#> 
#> (7,.,.) = 
#>  -0.7900 -1.1659 -1.8692  1.4381  2.7339
#>  -1.0159 -0.2644  0.0643  0.6920  1.5223
#>  -1.7030 -1.5408 -1.0526  3.0257  3.8570
#> 
#> (8,.,.) = 
#>  -0.4043  0.5885 -0.5919 -2.4073 -1.7434
#>  -1.2636  0.8417 -1.5456 -3.3914  2.0911
#>   3.0882 -0.6701  2.5458  4.4963 -1.6251
#> 
#> (9,.,.) = 
#>  -4.7235 -0.9884 -3.0721 -0.7362  5.2630
#>   0.3262 -0.0970  0.8119  1.2951  1.2358
#>  -3.3271 -0.8457 -0.9994  0.3460  0.8850
#> 
#> (10,.,.) = 
#>   2.6787 -0.0138  0.0866  1.2510  1.8553
#>   0.3136  0.2252  0.8376 -0.0204 -1.6592
#>   1.8933  0.5544  1.2222  0.1226 -0.9690
#> [ CPUFloatType{10,3,5} ]