Max

Max

Arguments

input
NA

dim
NA

keepdim
NA False.

out
(tuple, optional) the result tuple of two output tensors (max, max_indices)

input
(Tensor) the input tensor.

other
(Tensor) the second input tensor

Note

When the shapes do not match, the shape of the returned output tensor follows the broadcasting rules .

max(input) -> Tensor

Returns the maximum value of all elements in the input tensor.

max(input, dim, keepdim=False, out=None) -> (Tensor, LongTensor)

Returns a namedtuple (values, indices) where values is the maximum value of each row of the input tensor in the given dimension dim. And indices is the index location of each maximum value found (argmax).

If keepdim is True, the output tensors are of the same size as input except in the dimension dim where they are of size 1. Otherwise, dim is squeezed (see torch_squeeze), resulting in the output tensors having 1 fewer dimension than input.

max(input, other, out=None) -> Tensor

Each element of the tensor input is compared with the corresponding element of the tensor other and an element-wise maximum is taken.

The shapes of input and other don't need to match, but they must be broadcastable .

$$ \text{out}_i = \max(\text{tensor}_i, \text{other}_i) $$

Examples

a = torch_randn(c(1, 3))
a

#> torch_tensor 
#> -0.4569 -1.2771 -0.1725
#> [ CPUFloatType{1,3} ]

torch_max(a)

#> torch_tensor 
#> -0.172526
#> [ CPUFloatType{} ]

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#>  1.9436 -0.0701  0.4748 -0.5550
#>  0.5765 -0.7701 -0.5446 -0.1677
#> -1.0555 -1.1980 -1.2923 -0.8152
#>  0.2854 -0.1207 -0.2205 -1.5142
#> [ CPUFloatType{4,4} ]

torch_max(a, dim = 1)

#> [[1]]
#> torch_tensor 
#>  1.9436
#>  0.5765
#> -0.8152
#>  0.2854
#> [ CPUFloatType{4} ]
#> 
#> [[2]]
#> torch_tensor 
#>  0
#>  0
#>  3
#>  0
#> [ CPULongType{4} ]
#> 

a = torch_randn(c(4))
a

#> torch_tensor 
#> -0.7105
#>  0.6665
#>  0.2212
#> -0.5153
#> [ CPUFloatType{4} ]

b = torch_randn(c(4))
b

#> torch_tensor 
#> -0.0055
#>  0.0338
#> -0.0641
#>  1.2357
#> [ CPUFloatType{4} ]

torch_max(a, other = b)

#> torch_tensor 
#> -0.0055
#>  0.6665
#>  0.2212
#>  1.2357
#> [ CPUFloatType{4} ]