Pinverse

Pinverse

Arguments

input
(Tensor) The input tensor of size \((*, m, n)\) where \(*\) is zero or more batch dimensions

rcond
(float) A floating point value to determine the cutoff for small singular values. Default: 1e-15

Note

This method is implemented using the Singular Value Decomposition.

The pseudo-inverse is not necessarily a continuous function in the elements of the matrix `[1]`_.
Therefore, derivatives are not always existent, and exist for a constant rank only `[2]`_.
However, this method is backprop-able due to the implementation by using SVD results, and
could be unstable. Double-backward will also be unstable due to the usage of SVD internally.
See `~torch.svd` for more details.

pinverse(input, rcond=1e-15) -> Tensor

Calculates the pseudo-inverse (also known as the Moore-Penrose inverse) of a 2D tensor. Please look at Moore-Penrose inverse_ for more details

Examples

input = torch_randn(c(3, 5))
input

#> torch_tensor 
#> -0.7229 -1.8883  0.0426  0.8194 -1.2712
#>  0.3499  0.8774  2.0622  0.7425  0.0649
#>  0.3178 -0.7913 -0.5177  0.8580  0.4745
#> [ CPUFloatType{3,5} ]

torch_pinverse(input)

#> torch_tensor 
#> -0.1635  0.0783  0.3140
#> -0.2393  0.0636 -0.2083
#>  0.1139  0.3597 -0.1621
#>  0.0689  0.2336  0.5086
#> -0.2899  0.0236  0.4533
#> [ CPUFloatType{5,3} ]

# Batched pinverse example
a = torch_randn(c(2,6,3))
b = torch_pinverse(a)
torch_matmul(b, a)

#> torch_tensor 
#> (1,.,.) = 
#>   1.0000e+00 -1.5646e-07 -5.1782e-07
#>   2.8312e-07  1.0000e+00 -5.6997e-07
#>   3.2783e-07 -5.9605e-08  1.0000e+00
#> 
#> (2,.,.) = 
#>   1.0000  0.0000  0.0000
#>   0.0000  1.0000 -0.0000
#>  -0.0000  0.0000  1.0000
#> [ CPUFloatType{2,3,3} ]