Randn

Randn

Arguments

size
(int...) a sequence of integers defining the shape of the output tensor. Can be a variable number of arguments or a collection like a list or tuple.

out
(Tensor, optional) the output tensor.

dtype
(torch.dtype, optional) the desired data type of returned tensor. Default: if None, uses a global default (see torch_set_default_tensor_type).

layout
(torch.layout, optional) the desired layout of returned Tensor. Default: torch_strided.

device
(torch.device, optional) the desired device of returned tensor. Default: if None, uses the current device for the default tensor type (see torch_set_default_tensor_type). device will be the CPU for CPU tensor types and the current CUDA device for CUDA tensor types.

requires_grad
(bool, optional) If autograd should record operations on the returned tensor. Default: False.

randn(*size, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor

Returns a tensor filled with random numbers from a normal distribution with mean 0 and variance 1 (also called the standard normal distribution).

$$ \text{out}_{i} \sim \mathcal{N}(0, 1) $$ The shape of the tensor is defined by the variable argument size.

Examples

torch_randn(c(4))

#> torch_tensor 
#>  0.3291
#>  0.9967
#>  0.2446
#>  0.5778
#> [ CPUFloatType{4} ]

torch_randn(c(2, 3))

#> torch_tensor 
#> -0.5993 -0.8420  0.1534
#> -0.1845 -0.4167 -0.4723
#> [ CPUFloatType{2,3} ]