Sort

Sort

Arguments

input
(Tensor) the input tensor.

dim
(int, optional) the dimension to sort along

descending
(bool, optional) controls the sorting order (ascending or descending)

out
(tuple, optional) the output tuple of (Tensor, LongTensor) that can be optionally given to be used as output buffers

sort(input, dim=-1, descending=False, out=None) -> (Tensor, LongTensor)

Sorts the elements of the input tensor along a given dimension in ascending order by value.

If dim is not given, the last dimension of the input is chosen.

If descending is True then the elements are sorted in descending order by value.

A namedtuple of (values, indices) is returned, where the values are the sorted values and indices are the indices of the elements in the original input tensor.

Examples

x = torch_randn(c(3, 4))
out = torch_sort(x)
out

#> [[1]]
#> torch_tensor 
#> -1.2154  0.7292  1.0504  1.8189
#> -1.2647 -0.0016  0.4116  0.9378
#> -1.8727 -0.2049  0.3253  1.2256
#> [ CPUFloatType{3,4} ]
#> 
#> [[2]]
#> torch_tensor 
#>  0  1  2  3
#>  1  3  2  0
#>  1  0  2  3
#> [ CPULongType{3,4} ]
#> 

out = torch_sort(x, 0)
out

#> [[1]]
#> torch_tensor 
#> -1.2154 -1.8727  0.3253 -0.0016
#> -0.2049 -1.2647  0.4116  1.2256
#>  0.9378  0.7292  1.0504  1.8189
#> [ CPUFloatType{3,4} ]
#> 
#> [[2]]
#> torch_tensor 
#>  0  2  2  1
#>  2  1  1  2
#>  1  0  0  0
#> [ CPULongType{3,4} ]
#>