Topk

Topk

Arguments

input
(Tensor) the input tensor.

k
(int) the k in "top-k"

dim
(int, optional) the dimension to sort along

largest
(bool, optional) controls whether to return largest or smallest elements

sorted
(bool, optional) controls whether to return the elements in sorted order

out
(tuple, optional) the output tuple of (Tensor, LongTensor) that can be optionally given to be used as output buffers

topk(input, k, dim=None, largest=True, sorted=True, out=None) -> (Tensor, LongTensor)

Returns the k largest elements of the given input tensor along a given dimension.

If dim is not given, the last dimension of the input is chosen.

If largest is False then the k smallest elements are returned.

A namedtuple of (values, indices) is returned, where the indices are the indices of the elements in the original input tensor.

The boolean option sorted if True, will make sure that the returned k elements are themselves sorted

Examples

x = torch_arange(1., 6.)
x

#> torch_tensor 
#>  1
#>  2
#>  3
#>  4
#>  5
#> [ CPUFloatType{5} ]

torch_topk(x, 3)

#> [[1]]
#> torch_tensor 
#>  5
#>  4
#>  3
#> [ CPUFloatType{3} ]
#> 
#> [[2]]
#> torch_tensor 
#>  4
#>  3
#>  2
#> [ CPULongType{3} ]
#>