Var_mean

Var_mean

Arguments

input
(Tensor) the input tensor.

unbiased
(bool) whether to use the unbiased estimation or not

dim
(int or tuple of ints) the dimension or dimensions to reduce.

keepdim
(bool) whether the output tensor has dim retained or not.

var_mean(input, unbiased=True) -> (Tensor, Tensor)

Returns the variance and mean of all elements in the input tensor.

If unbiased is False, then the variance will be calculated via the biased estimator. Otherwise, Bessel's correction will be used.

var_mean(input, dim, keepdim=False, unbiased=True) -> (Tensor, Tensor)

Returns the variance and mean of each row of the input tensor in the given dimension dim.

If keepdim is True, the output tensor is of the same size as input except in the dimension(s) dim where it is of size 1. Otherwise, dim is squeezed (see torch_squeeze), resulting in the output tensor having 1 (or len(dim)) fewer dimension(s).

If unbiased is False, then the variance will be calculated via the biased estimator. Otherwise, Bessel's correction will be used.

Examples

a = torch_randn(c(1, 3))
a

#> torch_tensor 
#>  0.4086 -0.2189 -0.6886
#> [ CPUFloatType{1,3} ]

torch_var_mean(a)

#> [[1]]
#> torch_tensor 
#> 0.303025
#> [ CPUFloatType{} ]
#> 
#> [[2]]
#> torch_tensor 
#> -0.166313
#> [ CPUFloatType{} ]
#> 

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#> -0.8073  0.8183 -0.8192  0.0773
#> -0.1744 -1.2846  0.4797  0.0278
#>  0.0588 -0.6418 -0.2783  0.8315
#> -0.1949  2.0302  1.2109  0.1404
#> [ CPUFloatType{4,4} ]

torch_var_mean(a, 1)

#> [[1]]
#> torch_tensor 
#>  0.6217
#>  0.5617
#>  0.3947
#>  1.0357
#> [ CPUFloatType{4} ]
#> 
#> [[2]]
#> torch_tensor 
#> -0.1827
#> -0.2379
#> -0.0074
#>  0.7966
#> [ CPUFloatType{4} ]
#>