Baddbmm

Baddbmm

Arguments

beta
(Number, optional) multiplier for input (\(\beta\))

input
(Tensor) the tensor to be added

alpha
(Number, optional) multiplier for \(\text{batch1} \mathbin{@} \text{batch2}\) (\(\alpha\))

batch1
(Tensor) the first batch of matrices to be multiplied

batch2
(Tensor) the second batch of matrices to be multiplied

out
(Tensor, optional) the output tensor.

baddbmm(beta=1, input, alpha=1, batch1, batch2, out=None) -> Tensor

Performs a batch matrix-matrix product of matrices in batch1 and batch2. input is added to the final result.

batch1 and batch2 must be 3-D tensors each containing the same number of matrices.

If batch1 is a \((b \times n \times m)\) tensor, batch2 is a \((b \times m \times p)\) tensor, then input must be broadcastable with a \((b \times n \times p)\) tensor and out will be a \((b \times n \times p)\) tensor. Both alpha and beta mean the same as the scaling factors used in torch_addbmm.

$$ \text{out}_i = \beta\ \text{input}_i + \alpha\ (\text{batch1}_i \mathbin{@} \text{batch2}_i) $$ For inputs of type FloatTensor or DoubleTensor, arguments beta and alpha must be real numbers, otherwise they should be integers.

Examples

M = torch_randn(c(10, 3, 5))
batch1 = torch_randn(c(10, 3, 4))
batch2 = torch_randn(c(10, 4, 5))
torch_baddbmm(M, batch1, batch2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.5532  0.5630  1.0450 -1.6589 -2.8716
#>   0.7226  0.9734  0.8653  1.7153 -0.2198
#>  -4.5768 -0.8717 -3.2530 -4.2584  0.2332
#> 
#> (2,.,.) = 
#>   1.2190 -0.7162 -2.6078 -0.0106  2.2166
#>  -1.4823 -1.1109  2.4768  1.5547  0.4615
#>  -3.8128 -2.4184 -0.0166  2.2174  2.4476
#> 
#> (3,.,.) = 
#>  -0.3630  0.3426 -2.0910  0.0219  1.8372
#>  -0.1157 -2.3000 -0.3902  2.6760  2.5820
#>  -0.7991 -0.8670 -1.5908  1.6476  0.4549
#> 
#> (4,.,.) = 
#>   2.6230  0.2776 -3.5993  1.2619  2.0755
#>  -1.5890  2.9232  1.6261 -0.5379  0.1033
#>   0.0418  2.3737  0.9034  3.8939 -0.0550
#> 
#> (5,.,.) = 
#>  -6.4528 -2.1301  2.6023 -4.2910  0.3549
#>  -0.5229 -1.6363 -1.8690  3.4745 -1.3358
#>  -1.7701  4.0583  2.7307 -3.4307  0.3482
#> 
#> (6,.,.) = 
#>   0.5250  2.6148  7.3087 -2.2597 -0.8844
#>  -0.6999  0.4086 -3.9714 -0.3649 -1.1583
#>  -1.6487 -0.8045 -3.4602 -0.5155 -1.1977
#> 
#> (7,.,.) = 
#>   3.2993  1.1186 -0.7007 -1.6347 -1.7633
#>  -2.3803  1.0018 -1.4944  0.6228 -2.8728
#>   0.5978 -3.2697  0.6900 -0.0148  0.7469
#> 
#> (8,.,.) = 
#>   0.8415 -3.9278  2.9679 -0.7600 -0.5658
#>   2.7138 -0.1678  1.3842 -1.1456 -2.1884
#>  -0.0082  4.9714  0.0048 -1.8658 -2.5015
#> 
#> (9,.,.) = 
#>  -4.7181  4.0568 -0.4870 -1.4846 -0.8356
#>   0.8999  2.7887 -0.4951 -0.2899 -0.7583
#>  -3.4310  3.9947 -1.3769 -0.1857 -1.6050
#> 
#> (10,.,.) = 
#>  -4.4024 -1.7683 -0.9539 -0.2584  2.4847
#>  -3.2632 -3.0065 -0.6278  2.3107  0.6357
#>  -1.7312  0.7416  0.0108 -0.9520 -4.2298
#> [ CPUFloatType{10,3,5} ]