Baddbmm

Baddbmm

Arguments

beta
(Number, optional) multiplier for input (\(\beta\))

input
(Tensor) the tensor to be added

alpha
(Number, optional) multiplier for \(\text{batch1} \mathbin{@} \text{batch2}\) (\(\alpha\))

batch1
(Tensor) the first batch of matrices to be multiplied

batch2
(Tensor) the second batch of matrices to be multiplied

out
(Tensor, optional) the output tensor.

baddbmm(beta=1, input, alpha=1, batch1, batch2, out=None) -> Tensor

Performs a batch matrix-matrix product of matrices in batch1 and batch2. input is added to the final result.

batch1 and batch2 must be 3-D tensors each containing the same number of matrices.

If batch1 is a \((b \times n \times m)\) tensor, batch2 is a \((b \times m \times p)\) tensor, then input must be broadcastable with a \((b \times n \times p)\) tensor and out will be a \((b \times n \times p)\) tensor. Both alpha and beta mean the same as the scaling factors used in torch_addbmm.

$$ \text{out}_i = \beta\ \text{input}_i + \alpha\ (\text{batch1}_i \mathbin{@} \text{batch2}_i) $$ For inputs of type FloatTensor or DoubleTensor, arguments beta and alpha must be real numbers, otherwise they should be integers.

Examples

M = torch_randn(c(10, 3, 5))
batch1 = torch_randn(c(10, 3, 4))
batch2 = torch_randn(c(10, 4, 5))
torch_baddbmm(M, batch1, batch2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.3288  0.7881 -0.6697 -0.6873  0.5031
#>   4.2197  0.2838 -3.6574 -1.9255 -1.7399
#>   2.0451 -0.4215 -0.5694 -1.1075 -2.7478
#> 
#> (2,.,.) = 
#>   0.4905 -1.0518 -2.9275  5.7947 -4.5032
#>   0.5375 -0.2184  0.8906 -4.7661  3.3000
#>   0.7636 -2.4083  0.7322 -2.2700  2.3588
#> 
#> (3,.,.) = 
#>  -0.4258 -0.2806 -0.0727 -0.6614 -2.7713
#>   1.0894  2.7829  2.8252 -3.0689  4.2668
#>   0.4234 -0.5574 -2.1175 -0.8862 -1.2929
#> 
#> (4,.,.) = 
#>   1.4579  1.1742  1.3728  0.7324 -0.7000
#>  -0.7872  0.8846  0.6496 -1.7996 -0.3044
#>  -5.7071 -3.6463 -0.0782  1.9616 -3.0899
#> 
#> (5,.,.) = 
#>  -2.7536 -3.9917 -4.3073 -4.2225  0.8132
#>  -1.5330 -8.7609 -1.2027  0.8929 -2.3657
#>  -0.9386 -0.9599  2.0581  1.3073 -2.3903
#> 
#> (6,.,.) = 
#>   4.8846  0.6678  0.0668  0.1522 -4.2534
#>  -3.4191  1.6740  0.0213 -1.6462  0.6259
#>   2.9710  2.4275 -1.3012 -3.7981 -3.8904
#> 
#> (7,.,.) = 
#>   1.1282 -0.2703  5.1412 -0.4345 -1.8980
#>  -0.0126 -2.8765 -4.1867 -1.7262 -2.5079
#>   1.9308 -2.1071 -0.4913 -1.8134  1.0691
#> 
#> (8,.,.) = 
#>  -1.0049  2.2572 -0.3031  7.0174  2.3467
#>   1.2134 -0.0413  1.2479  4.6231  0.6420
#>  -0.5382 -1.7336 -0.9825  1.1547 -1.2571
#> 
#> (9,.,.) = 
#>  -2.0869 -1.6888  1.0194  0.3671 -1.0374
#>  -0.4037  1.1579  2.5488  0.1028 -0.8327
#>   2.3918 -2.4271 -0.9960 -1.9017  0.0574
#> 
#> (10,.,.) = 
#>   2.1154  1.4580  1.5650 -4.5504  4.2094
#>   0.5166 -1.3591  0.7578 -1.3742  0.1063
#>  -1.5573 -0.0852  1.4115 -0.9602 -1.6002
#> [ CPUFloatType{10,3,5} ]