Baddbmm

Baddbmm

Arguments

beta
(Number, optional) multiplier for input (\(\beta\))

input
(Tensor) the tensor to be added

alpha
(Number, optional) multiplier for \(\text{batch1} \mathbin{@} \text{batch2}\) (\(\alpha\))

batch1
(Tensor) the first batch of matrices to be multiplied

batch2
(Tensor) the second batch of matrices to be multiplied

out
(Tensor, optional) the output tensor.

baddbmm(beta=1, input, alpha=1, batch1, batch2, out=None) -> Tensor

Performs a batch matrix-matrix product of matrices in batch1 and batch2. input is added to the final result.

batch1 and batch2 must be 3-D tensors each containing the same number of matrices.

If batch1 is a \((b \times n \times m)\) tensor, batch2 is a \((b \times m \times p)\) tensor, then input must be broadcastable with a \((b \times n \times p)\) tensor and out will be a \((b \times n \times p)\) tensor. Both alpha and beta mean the same as the scaling factors used in torch_addbmm.

$$ \text{out}_i = \beta\ \text{input}_i + \alpha\ (\text{batch1}_i \mathbin{@} \text{batch2}_i) $$ For inputs of type FloatTensor or DoubleTensor, arguments beta and alpha must be real numbers, otherwise they should be integers.

Examples

M = torch_randn(c(10, 3, 5))
batch1 = torch_randn(c(10, 3, 4))
batch2 = torch_randn(c(10, 4, 5))
torch_baddbmm(M, batch1, batch2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.2672  2.3495  1.7255 -0.4168  2.7060
#>   0.7317  4.6747 -0.7024 -2.4117  2.3289
#>  -2.8159  1.4670 -0.5527  2.6698  1.7398
#> 
#> (2,.,.) = 
#>  -1.0968 -1.5057  2.8696  2.9538 -2.9999
#>   1.2589  1.1616 -2.2464  0.1686 -1.6669
#>   0.9481  1.0259  1.5784  1.2510 -1.6563
#> 
#> (3,.,.) = 
#>   0.6881 -1.9041 -0.2464 -0.3903  1.2476
#>  -0.9259  1.5030 -1.8747 -3.2425 -1.7505
#>   0.7762 -5.1459  1.6853  1.7827 -0.1539
#> 
#> (4,.,.) = 
#>  -0.6931 -2.5698  0.9840 -0.5518 -0.1979
#>   0.2289 -1.3985 -7.4090  3.4724 -3.3305
#>   2.0277 -3.4175 -1.0659 -0.8327  1.7181
#> 
#> (5,.,.) = 
#>  -1.4623 -2.5475  1.9947  1.7479 -1.2926
#>   3.7046  0.4089  0.6779  0.0375  1.2279
#>  -0.2505 -0.4367 -1.2189  1.3016  1.2527
#> 
#> (6,.,.) = 
#>   1.4873 -0.7159 -0.2832 -1.5998  0.8029
#>   0.7408 -1.1798 -0.2593  0.0927  0.7279
#>  -1.9743 -3.2111  0.9555 -0.9231 -0.9317
#> 
#> (7,.,.) = 
#>  -0.5620  4.4030  1.6184 -0.2764  0.1145
#>   1.1740  5.6473 -1.3085 -0.7302  1.6017
#>  -0.2521 -3.5346 -1.1871 -0.1003 -0.9406
#> 
#> (8,.,.) = 
#>  -0.0912 -0.1360  1.4268  3.8176  0.6553
#>   3.4799 -1.7063  3.6496 -1.5905  1.1965
#>  -3.5381 -0.4613 -1.2974 -0.2281 -1.4277
#> 
#> (9,.,.) = 
#>   1.7590  0.6722 -2.0162  3.9190 -2.4814
#>   0.1460  0.2438  2.1092 -4.7314  0.0893
#>   0.0041  0.6536 -0.3636 -2.1134  4.1528
#> 
#> (10,.,.) = 
#>  -1.4380 -0.2908  0.7458  2.6174  2.5340
#>  -2.3755  1.0130 -1.0847 -0.0704  2.3968
#>  -0.2163 -2.3204  0.4743 -0.3845  0.0219
#> [ CPUFloatType{10,3,5} ]