Bmm

Bmm

Arguments

input
(Tensor) the first batch of matrices to be multiplied

mat2
(Tensor) the second batch of matrices to be multiplied

out
(Tensor, optional) the output tensor.

Note

This function does not broadcast . For broadcasting matrix products, see torch_matmul.

bmm(input, mat2, out=None) -> Tensor

Performs a batch matrix-matrix product of matrices stored in input and mat2.

input and mat2 must be 3-D tensors each containing the same number of matrices.

If input is a \((b \times n \times m)\) tensor, mat2 is a \((b \times m \times p)\) tensor, out will be a \((b \times n \times p)\) tensor.

$$ \text{out}_i = \text{input}_i \mathbin{@} \text{mat2}_i $$

Examples

input = torch_randn(c(10, 3, 4))
mat2 = torch_randn(c(10, 4, 5))
res = torch_bmm(input, mat2)
res

#> torch_tensor 
#> (1,.,.) = 
#>   0.8623 -0.2242 -0.8064  1.1810 -0.5275
#>  -0.0808  1.8987 -0.7477  1.2060 -0.1614
#>  -1.8882  0.7827  1.7203  0.3344 -0.0279
#> 
#> (2,.,.) = 
#>  -1.8141 -0.5727  2.9210 -0.2726  1.2037
#>   2.3634  0.7298 -0.9221 -0.8270  0.3927
#>  -1.5131 -0.4826 -0.5272 -0.1389  0.6595
#> 
#> (3,.,.) = 
#>   1.7836 -0.7125 -2.0643  3.6928  2.9212
#>  -0.6630 -2.4628 -2.2335  0.0114 -0.1745
#>  -2.2239 -3.0290  0.4260 -2.9730 -3.6834
#> 
#> (4,.,.) = 
#>   0.4645  1.8824 -0.2316  0.2473 -0.8421
#>   0.8401  0.0298 -1.6641  1.7717 -1.6389
#>  -0.9565  1.3660 -0.9363 -1.5906  0.1058
#> 
#> (5,.,.) = 
#>   1.2401  1.2447 -1.8137 -2.6481  1.9014
#>  -0.3962 -2.7717  0.0306 -1.0062 -0.0594
#>  -0.4308  0.2950  3.1510  2.3015 -3.4588
#> 
#> (6,.,.) = 
#>   4.5447 -1.8161  1.0129 -0.9367 -0.5296
#>  -1.5121  1.3863  0.0426 -0.8785  0.2536
#>  -2.0843  1.3254 -1.3412  0.1500 -0.6935
#> 
#> (7,.,.) = 
#>  -0.1501 -0.1975  0.3488  3.4630  0.8495
#>   0.2830 -0.8376  1.2090  3.4837  0.9487
#>  -0.3958  2.4595 -1.2667 -1.5102 -0.3447
#> 
#> (8,.,.) = 
#>   2.8103  0.0743  2.4884  1.4043 -2.6162
#>  -0.3316  0.4362  2.1075 -1.4711 -1.0260
#>   3.1159 -0.8605  0.7050  2.0135 -1.4318
#> 
#> (9,.,.) = 
#>  -1.8200 -2.7651 -2.2669 -0.5292  3.7412
#>   0.2482  2.2253  0.7739  0.1200 -1.8829
#>  -1.4773  0.5268 -0.8546 -0.3524  1.6574
#> 
#> (10,.,.) = 
#>  -1.8993 -0.9521  1.7775  1.5515  1.1468
#>   1.7589  0.1589  2.4372 -0.6778 -0.8335
#>  -4.7758 -1.6642  0.2488  3.2651  3.1034
#> [ CPUFloatType{10,3,5} ]