Div

Div

Arguments

input
(Tensor) the input tensor.

other
(Number) the number to be divided to each element of input

div(input, other, out=None) -> Tensor

Divides each element of the input input with the scalar other and returns a new resulting tensor.

Each element of the tensor input is divided by each element of the tensor other. The resulting tensor is returned.

$$ \text{out}_i = \frac{\text{input}_i}{\text{other}_i} $$ The shapes of input and other must be broadcastable . If the torch_dtype of input and other differ, the torch_dtype of the result tensor is determined following rules described in the type promotion documentation . If out is specified, the result must be castable to the torch_dtype of the specified output tensor. Integral division by zero leads to undefined behavior.

Warning

Integer division using div is deprecated, and in a future release div will perform true division like torch_true_divide. Use torch_floor_divide (// in Python) to perform integer division, instead.

$$ \text{out}_i = \frac{\text{input}_i}{\text{other}} $$ If the torch_dtype of input and other differ, the torch_dtype of the result tensor is determined following rules described in the type promotion documentation . If out is specified, the result must be castable to the torch_dtype of the specified output tensor. Integral division by zero leads to undefined behavior.

Examples

a = torch_randn(c(5))
a

#> torch_tensor 
#>  1.7491
#>  1.0545
#> -0.9421
#>  1.4904
#> -1.2488
#> [ CPUFloatType{5} ]

torch_div(a, 0.5)

#> torch_tensor 
#>  3.4982
#>  2.1090
#> -1.8843
#>  2.9809
#> -2.4975
#> [ CPUFloatType{5} ]

a = torch_randn(c(4, 4))
a

#> torch_tensor 
#> -0.8645 -0.0644  1.6837 -0.2799
#>  0.8651  0.3207  0.9701 -0.7295
#> -0.3726 -1.2875 -3.4855  1.6616
#>  0.5742 -1.1932 -0.3622  1.1164
#> [ CPUFloatType{4,4} ]

b = torch_randn(c(4))
b

#> torch_tensor 
#> -1.5981
#>  1.4460
#> -1.3709
#>  0.5575
#> [ CPUFloatType{4} ]

torch_div(a, b)

#> torch_tensor 
#>  0.5410 -0.0446 -1.2281 -0.5020
#> -0.5413  0.2218 -0.7076 -1.3086
#>  0.2331 -0.8904  2.5424  2.9806
#> -0.3593 -0.8252  0.2642  2.0026
#> [ CPUFloatType{4,4} ]