Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -8.3773  3.6248
#>  -2.2446  1.0897
#>  -2.3596 -0.7066
#> 
#> (2,.,.) = 
#>   4.9731  0.7604
#>  -4.1126  2.8574
#>  -0.6452 -0.6609
#> 
#> (3,.,.) = 
#>  -5.6495  6.5738
#>  -7.0227  2.6824
#>   0.5881 -4.3484
#> 
#> (4,.,.) = 
#>  -0.6931  1.0474
#>   4.3026  4.5640
#>   1.0206 -0.2850
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.4367  3.0016
#>  -2.2693  2.7984
#>  -0.3490 -1.5002
#> 
#> (2,.,.) = 
#>  -0.6102  0.6793
#>   1.6211 -2.5020
#>  -0.6429  0.4940
#> 
#> (3,.,.) = 
#>  -4.5767  2.0977
#>  -2.3643 -0.9123
#>  -0.5367 -1.0273
#> 
#> (4,.,.) = 
#>  -0.7537 -2.1538
#>   0.7679  1.7056
#>  -0.8309  1.3269
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   0.1497 -5.2737
#>  -0.8659  6.2359
#>   4.8456  6.9425
#>  -1.5652 -6.9051
#>  -0.3928 -2.7554
#> 
#> (2,1,1,.,.) = 
#>  -3.5402  0.6524
#>   0.2328  2.7048
#>  -5.3443  4.7451
#>  -1.2418 -5.3425
#>  -0.7583  1.9579
#> 
#> (3,1,1,.,.) = 
#>  -1.0218 -5.8703
#>   3.1747  6.4986
#>  -3.8691 -0.9668
#>   0.1334  0.9270
#>   3.6588 -2.5037
#> 
#> (1,2,1,.,.) = 
#>   1.7162 -9.2945
#>  -2.4229 -2.6208
#>   0.7110 -7.1646
#>   0.9543  4.8382
#>   7.4707 -6.6734
#> 
#> (2,2,1,.,.) = 
#>  -10.3594  -9.1982
#>    2.0475  -7.3252
#>    1.9476   3.7941
#>    1.8448  -5.6072
#>    9.9672  -3.1990
#> 
#> (3,2,1,.,.) = 
#>   0.2643 -3.9095
#>  -0.4682  1.5421
#>  -2.7733 -4.6781
#>   3.5991 -14.6299
#>  -3.5847  0.1011
#> 
#> (1,3,1,.,.) = 
#>    3.9932   1.5141
#>   -3.5176  11.7890
#>   -2.6101   6.2770
#>    3.0546   0.9125
#>   -4.1437  -7.9950
#> 
#> (2,3,1,.,.) = 
#>   2.3308 -12.2663
#>  -4.0821  2.5402
#>  -0.0144 -2.8482
#>  -2.7828  0.3687
#>   1.3227 -2.7881
#> 
#> (3,3,1,.,.) = 
#>  -0.6572  0.6916
#>   1.4381  8.1525
#>  -0.7392 -2.5997
#>   2.7102 -1.3684
#>  -9.0063  7.5199
#> 
#> (1,1,2,.,.) = 
#>   9.1469 -0.6164
#>  -1.9742  4.9940
#>   3.3270  0.9512
#>   8.1288  1.9495
#>  -1.1288  0.1444
#> 
#> (2,1,2,.,.) = 
#>    3.7373   2.2503
#>   11.8966   6.8293
#>    4.2869  -2.6413
#>    9.2146  -5.9252
#>    4.0497 -11.9365
#> 
#> (3,1,2,.,.) = 
#>   0.7221  2.8049
#>  -0.6862  2.5125
#>  -8.3855 -4.9299
#>   2.6521 -1.6775
#>  -3.1159 -14.8524
#> 
#> (1,2,2,.,.) = 
#>  -2.7680  0.3613
#>  -1.7940  0.2003
#>  -0.3858 -14.0861
#>  -1.8480  0.3957
#>  -3.5513 -7.3426
#> 
#> (2,2,2,.,.) = 
#>  -5.8451  2.1999
#>   3.6516 -10.6216
#>  -6.3918 -10.3027
#>   2.9500 -7.3934
#>  -4.2273  2.6692
#> 
#> (3,2,2,.,.) = 
#>   9.1972  3.2468
#>   1.2524  4.4693
#>  -4.5480 -5.6438
#>  -0.4244  2.2104
#>   4.4541 -10.6059
#> 
#> (1,3,2,.,.) = 
#>   5.9636  2.7061
#>  -6.5543  0.5175
#>   0.1469 -5.9422
#>  -0.8321  4.9598
#>   2.8476 -3.0244
#> 
#> (2,3,2,.,.) = 
#>  -2.2153  3.8122
#>   1.7482  5.0721
#>   8.7466 -7.1298
#>   1.0762  5.3046
#>   4.7562  2.4666
#> 
#> (3,3,2,.,.) = 
#>  -7.2244  0.5161
#>   0.3935  3.5232
#>  -4.1514  6.0483
#>  -2.6387  2.3631
#>  -3.5593  0.9383
#> 
#> (1,1,3,.,.) = 
#>   7.3541 -0.9243
#>  -2.6365 -5.1086
#>   1.1539  0.2161
#>   2.1611 -4.1282
#>  -2.0103 -2.2734
#> 
#> (2,1,3,.,.) = 
#>   1.3206  3.5045
#>  -1.2365 -4.1385
#>   1.4804  6.7293
#>  -1.1555 -3.6060
#>   6.3744  1.1155
#> 
#> (3,1,3,.,.) = 
#>  -3.8844  0.2593
#>   3.8237 -3.7728
#>  -1.9480  3.1775
#>  -2.6294  5.3144
#>   5.9141  3.8174
#> 
#> (1,2,3,.,.) = 
#>   5.4622 -2.6380
#>  -7.5544 -3.5887
#>  -10.1589 -5.6119
#>   4.8294  1.2380
#>  -3.2457  1.0986
#> 
#> (2,2,3,.,.) = 
#>  -7.4792 -2.5944
#>  -1.8481  1.7707
#>   1.7556  5.3231
#>  -0.6476 -0.2167
#>   4.9237  1.6517
#> 
#> (3,2,3,.,.) = 
#>    5.6405   6.0987
#>   -1.1050  -0.2973
#>   10.0811   8.1068
#>    3.8626  -0.2483
#>   -5.6683   2.1662
#> 
#> (1,3,3,.,.) = 
#>   1.6579  0.2144
#>   1.9793 -2.6460
#>   1.6087  3.2644
#>   4.1787  4.5892
#>  -4.6566 -1.9983
#> 
#> (2,3,3,.,.) = 
#>  -1.4826  4.5254
#>  -0.2554 -4.1838
#>   3.0260  1.7187
#>   4.4734  5.1358
#>  -5.2988  3.6497
#> 
#> (3,3,3,.,.) = 
#>  -8.5312  4.8679
#>  -0.8771  8.6313
#>   1.5174  2.5782
#>  -2.2378  2.0586
#>   1.7086 -6.6722
#> 
#> (1,1,4,.,.) = 
#>   4.8943 -5.9531
#>   1.7191 -2.7513
#>   1.3912  2.1271
#>   1.1094 -1.9250
#>   1.4548 -2.1335
#> 
#> (2,1,4,.,.) = 
#>   2.5656 -3.0079
#>  -0.6895  0.6236
#>  -6.7338  1.1161
#>   5.2974 -2.3114
#>  -3.6148  1.7230
#> 
#> (3,1,4,.,.) = 
#>  -1.5434  0.6199
#>  -3.7408  3.2200
#>  -5.2425  2.6570
#>   0.1944  4.6676
#>  -3.3026 -4.7999
#> 
#> (1,2,4,.,.) = 
#>  -2.7219  3.7696
#>  -3.8073  6.4382
#>   1.1059 -4.6141
#>   2.9597  1.9934
#>   7.5096  0.6929
#> 
#> (2,2,4,.,.) = 
#>  -2.3183  3.9271
#>  -0.9715 -0.1074
#>  -3.4303  2.2953
#>  -4.4589 -8.8122
#>  -0.3168  0.8183
#> 
#> (3,2,4,.,.) = 
#>   -4.1219   6.3235
#>   -1.0465  10.6843
#>    8.3320   3.3496
#>    0.5548  -4.3541
#>   -5.6917  -1.7722
#> 
#> (1,3,4,.,.) = 
#>   0.5991  0.8539
#>  -0.4692 -1.5159
#>   9.8136  6.6449
#>  -4.7116  2.8549
#>   0.6419 -0.8453
#> 
#> (2,3,4,.,.) = 
#>  -8.6777  9.7957
#>  -11.1989 -2.1682
#>   0.6252  0.0337
#>   3.2240 -3.9806
#>  -4.7467  1.4897
#> 
#> (3,3,4,.,.) = 
#>  -7.5803  5.5696
#>  -1.6161 -2.3529
#>  -7.8575 -7.4299
#>   6.6697 -9.3616
#>   4.5923 -5.4753
#> 
#> (1,1,5,.,.) = 
#>   4.1660 -4.1456
#>   2.5714  3.7100
#>  -2.2861  2.2283
#>   3.9918 -1.4056
#>   1.7836 -0.7868
#> 
#> (2,1,5,.,.) = 
#>   2.0240 -1.9511
#>  -6.0039  2.7040
#>   1.2700 -2.4832
#>  -0.6153  8.2124
#>   0.2672 -0.1410
#> 
#> (3,1,5,.,.) = 
#>   -2.5188  -6.8951
#>   -3.0202   0.3309
#>    4.4590  -4.5421
#>   12.6693   2.8063
#>  -10.8426  -6.1748
#> 
#> (1,2,5,.,.) = 
#>  -6.0880  6.7758
#>   1.1400  2.8938
#>  -7.4309 -0.7559
#>  -11.5777 -13.0110
#>  -0.7723 -6.5439
#> 
#> (2,2,5,.,.) = 
#>   -4.3603   1.3324
#>    6.0514   4.8407
#>   -2.1223  12.1690
#>   -7.9164  -1.6713
#>   -3.9912   2.2989
#> 
#> (3,2,5,.,.) = 
#>   0.0233 -7.0779
#>  -6.2843 -3.6845
#>   5.7933 -6.2497
#>   2.3509 -3.5344
#>  -2.9579 -2.2815
#> 
#> (1,3,5,.,.) = 
#>   2.2648 -0.4917
#>  -5.4885 -1.6845
#>   0.3042 -6.2279
#>  -9.4157  2.4254
#>  -0.1173 -1.9619
#> 
#> (2,3,5,.,.) = 
#>   0.6721 -5.2125
#>  -3.8439  6.0852
#>  -2.6664 -2.0646
#>   2.7986 -3.6396
#>  -1.3686 -2.0803
#> 
#> (3,3,5,.,.) = 
#>  -7.7942  8.2927
#>  -0.4961 -5.0846
#>  -2.5209  7.7243
#>  -0.1730 -5.5018
#>   1.0804  4.1413
#> [ CPUFloatType{3,3,5,5,2} ]