Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.4365  3.4257
#>  -2.1468 -0.7571
#>   2.4442 -5.6768
#> 
#> (2,.,.) = 
#>   4.4267  3.3580
#>   3.6414 -3.9885
#>   3.7286  1.2518
#> 
#> (3,.,.) = 
#>  -1.3633  2.1518
#>   0.4865 -4.1595
#>   2.1825 -3.4434
#> 
#> (4,.,.) = 
#>  -1.2572  6.0711
#>   1.4414 -4.1612
#>   7.8725 -1.2227
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   0.3424  3.7517
#>   0.8556 -3.2666
#>   4.0570 -2.2728
#> 
#> (2,.,.) = 
#>   0.9100  1.7395
#>  -0.7015  1.4006
#>  -0.5532 -1.5943
#> 
#> (3,.,.) = 
#>  -1.2424 -0.9629
#>  -1.6858  0.8083
#>  -1.7436 -2.2873
#> 
#> (4,.,.) = 
#>  -0.4466 -1.1025
#>  -0.6151  0.3006
#>   0.6840  0.4776
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   1.7914  0.0494
#>  -6.9617 -2.5954
#>   6.0873 -2.9136
#>  -1.9202  4.4099
#>   1.8064 -0.0484
#> 
#> (2,1,1,.,.) = 
#>   -5.1231  -0.9675
#>    1.1492  -2.9110
#>    4.6119  13.9854
#>    0.9269   0.4197
#>   -0.2615  -3.1461
#> 
#> (3,1,1,.,.) = 
#>   5.1586  3.7633
#>   7.8248  1.4496
#>   6.0585 -4.2275
#>   7.0316  2.1766
#>  -0.8565 -1.2412
#> 
#> (1,2,1,.,.) = 
#>   -7.8691  -5.9228
#>    1.2059  -2.5000
#>    2.9389   1.0273
#>   11.6509   3.4548
#>    2.7779  -3.9932
#> 
#> (2,2,1,.,.) = 
#>  -6.3236  0.0377
#>   0.6549  3.1918
#>   1.8550  6.9059
#>   2.9504 -2.5803
#>  -2.3291 -7.9165
#> 
#> (3,2,1,.,.) = 
#>   8.4127 -4.7213
#>  -5.0675 -1.9068
#>  -1.8540  6.8363
#>   1.0410 -0.2078
#>   5.1210  0.0998
#> 
#> (1,3,1,.,.) = 
#>    3.6121  10.0412
#>   -6.5362  -3.3009
#>    0.5797   3.5292
#>    6.9416   3.1653
#>   -3.2000   0.9658
#> 
#> (2,3,1,.,.) = 
#>    3.1583   2.6123
#>    2.0810   4.7433
#>  -14.1109  -0.5452
#>   11.8515  -2.9695
#>   -7.7374   5.7996
#> 
#> (3,3,1,.,.) = 
#>   9.3771 -8.4643
#>   0.1982  0.1959
#>   0.0150 -2.9204
#>  -3.7229  0.8058
#>  -7.0771  3.3220
#> 
#> (1,1,2,.,.) = 
#>   2.1309  6.8131
#>   0.8327 -1.7792
#>  -7.0408  3.9540
#>   4.1741 -3.5532
#>   1.1412  2.4012
#> 
#> (2,1,2,.,.) = 
#>  -3.1196 -14.7415
#>  -0.1739  4.7643
#>  -0.7597  7.3621
#>  -2.3053 -2.7710
#>  -3.2653 -5.1693
#> 
#> (3,1,2,.,.) = 
#>   3.7773  0.2954
#>   2.7119  1.8015
#>   8.3893 -1.5412
#>  -4.7292 -1.4925
#>   1.2008 -1.8165
#> 
#> (1,2,2,.,.) = 
#>  -2.8957 -8.3602
#>   8.3446  3.4167
#>  -0.0531  9.5941
#>   8.1054 -0.2681
#>   2.2313 -9.0906
#> 
#> (2,2,2,.,.) = 
#>  -3.8885 -0.0056
#>  -0.4637  2.0038
#>   4.6340  2.0994
#>   2.0780  1.7342
#>  -3.7732 -6.1493
#> 
#> (3,2,2,.,.) = 
#>  -5.2378 -4.1367
#>  -7.3681 -0.1327
#>   1.9957 -5.0910
#>  -2.2003  2.9797
#>  -9.4620 -6.0832
#> 
#> (1,3,2,.,.) = 
#>   -6.6876   0.7137
#>   -6.5470  -4.0295
#>    2.9950  -2.8351
#>    0.2575  -4.1144
#>    0.7616  12.7860
#> 
#> (2,3,2,.,.) = 
#>  -3.2243 -0.0775
#>  -4.3391  6.9062
#>  -13.3096 -1.3628
#>  -4.0850  6.0997
#>   1.6014 -5.6656
#> 
#> (3,3,2,.,.) = 
#>  -5.6679 -1.3614
#>   0.4791  3.0829
#>   9.3927 -5.2862
#>  -1.2413  5.4951
#>  -0.5686 -5.3596
#> 
#> (1,1,3,.,.) = 
#>  -3.2258  0.0504
#>   2.0990 -5.5614
#>  -0.5096  2.0396
#>  -3.3375  1.3989
#>  -7.5730  4.4817
#> 
#> (2,1,3,.,.) = 
#>  -3.0976 -3.2800
#>   0.9195 -1.4917
#>  -7.4508 -2.1797
#>   5.0095 -0.0722
#>  -9.6178 -0.0334
#> 
#> (3,1,3,.,.) = 
#>    5.0162   2.9260
#>    0.6187  -0.2743
#>   -1.0439  11.7827
#>    9.4404   2.4685
#>   -1.8315   1.7939
#> 
#> (1,2,3,.,.) = 
#>   -0.9911   1.1596
#>    1.5463  -1.4200
#>   -5.8116  13.4196
#>    1.1199   6.1745
#>   -3.7591   8.3417
#> 
#> (2,2,3,.,.) = 
#>    2.2045  -0.1597
#>   -1.4650   5.4761
#>    7.4548   5.4064
#>    5.6223  10.8325
#>    0.3534  -0.5921
#> 
#> (3,2,3,.,.) = 
#>  -1.2975  6.7886
#>   0.6126  0.6473
#>  -5.5057 -0.5179
#>   3.3744  0.1658
#>  -2.7777 -1.1498
#> 
#> (1,3,3,.,.) = 
#>  -1.8583  0.1468
#>   5.7776 -2.3319
#>   2.1546 -6.0905
#>  -0.7755  1.3671
#>   3.8284 -5.2547
#> 
#> (2,3,3,.,.) = 
#>   2.5011  7.3438
#>  -2.0770  0.4962
#>  -2.3133 -6.0247
#>   4.2450  5.4053
#>   3.1738  2.2681
#> 
#> (3,3,3,.,.) = 
#>  -4.7131  6.3970
#>   4.6130 -9.1317
#>  -0.6388  3.7863
#>   2.0019 -4.5394
#>  -0.0107  5.7149
#> 
#> (1,1,4,.,.) = 
#>   5.5761  3.6421
#>  -3.0796  2.7086
#>  -1.1480 -0.4948
#>   1.5135 -5.3372
#>   3.6694  6.1210
#> 
#> (2,1,4,.,.) = 
#>  -0.7363  8.6532
#>   3.0686  3.3354
#>  -0.4524  3.5807
#>   0.8130  5.1460
#>   2.0133  8.3245
#> 
#> (3,1,4,.,.) = 
#>   7.3698 -1.1825
#>   0.3399 -0.0499
#>   3.5985  0.7421
#>  -0.7956 -2.0406
#>  -2.1240  1.5348
#> 
#> (1,2,4,.,.) = 
#>   8.3271  0.4809
#>  -10.1951  2.2802
#>  -2.8250  1.9880
#>   3.1374 -1.6912
#>  -5.9945  3.9003
#> 
#> (2,2,4,.,.) = 
#>   9.0681  3.0603
#>   2.3707 -7.7409
#>  -11.5891  1.6186
#>   7.5529 -8.6253
#>  -8.9638  1.1477
#> 
#> (3,2,4,.,.) = 
#>  -2.3949 -3.4461
#>  -3.3956 -0.6699
#>  -2.7433 -11.7738
#>  -6.0127 -3.3732
#>  -6.1611 -5.8590
#> 
#> (1,3,4,.,.) = 
#>   5.3269  6.7895
#>  -0.1581  2.6589
#>  -3.9082 -6.0479
#>  -8.2684 -7.0439
#>  -9.8153  2.0760
#> 
#> (2,3,4,.,.) = 
#>   8.4479  1.3011
#>   8.9855 -5.5462
#>  -4.9285 -0.1178
#>  -1.6060 -0.0666
#>   5.7092 -4.2187
#> 
#> (3,3,4,.,.) = 
#>  -4.9860 -3.7521
#>  -2.6478 -5.6869
#>  -1.8261 -3.5155
#>   7.5024  3.2605
#>  -4.2290  3.8692
#> 
#> (1,1,5,.,.) = 
#>    2.0113  -7.8474
#>   10.2617  -2.7270
#>   -4.4988   0.4907
#>   -1.3462   4.1472
#>    4.5302   0.7823
#> 
#> (2,1,5,.,.) = 
#>  -2.2762  0.1232
#>  -4.5210 -5.2763
#>   2.7434 -4.7291
#>  -0.7510  1.9246
#>   9.2671  0.1395
#> 
#> (3,1,5,.,.) = 
#>   6.0449 -1.4001
#>   5.9936  0.4113
#>   2.2850  3.4918
#>   2.8770 -13.6269
#>   1.4134 -3.3539
#> 
#> (1,2,5,.,.) = 
#>  -4.1268 -4.9198
#>   2.7066  5.8932
#>  -7.3517 -1.8770
#>   1.5174  5.2309
#>   2.1090 -3.0130
#> 
#> (2,2,5,.,.) = 
#>   7.8372  6.2648
#>   6.3700 -2.5002
#>   6.1166  1.6275
#>  -1.0665  0.1052
#>  -6.2330  4.5141
#> 
#> (3,2,5,.,.) = 
#>  -4.5939 -1.1604
#>   4.4330 -0.9048
#>  -0.2325  3.9605
#>   4.6745 -3.2163
#>  -4.5515 -1.0745
#> 
#> (1,3,5,.,.) = 
#>  -1.0086  1.4084
#>   0.3681  0.5429
#>  -7.6002  8.6141
#>   1.6903 -2.8305
#>  -3.5208  2.0310
#> 
#> (2,3,5,.,.) = 
#>   8.9745 -3.7305
#>   9.0560  3.5521
#>   1.6310 -2.9861
#>   1.7669 -1.4121
#>   6.1454  1.8183
#> 
#> (3,3,5,.,.) = 
#>   -5.9584   6.2336
#>   -5.3925   2.5361
#>   -5.9519  -3.4034
#>    0.1401  11.0285
#>    4.7713   5.1519
#> [ CPUFloatType{3,3,5,5,2} ]