Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   3.1005  3.2041
#>   1.3027  5.3388
#>   2.9356  4.3950
#> 
#> (2,.,.) = 
#>  -5.6914 -3.5430
#>  -4.1740 -0.4785
#>  -0.8523 -9.4284
#> 
#> (3,.,.) = 
#>  -4.9994  0.8356
#>  -6.4940  1.6126
#>  -1.0322 -2.9614
#> 
#> (4,.,.) = 
#>  -8.3065  0.1208
#>   4.1247  4.5167
#>   3.0443 -5.0909
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -3.9742  0.1544
#>  -1.3102  2.7474
#>   1.0239 -3.2714
#> 
#> (2,.,.) = 
#>   2.9409  1.2459
#>   3.1980 -1.1432
#>   2.0763  0.8650
#> 
#> (3,.,.) = 
#>   3.0247  1.8655
#>  -1.2855  0.7283
#>  -0.0721  3.9882
#> 
#> (4,.,.) = 
#>   1.1090 -0.0617
#>   0.7004  3.0062
#>  -0.0924  2.8133
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -4.0652  7.8355
#>  -3.5424  1.4277
#>  -0.4713  5.9662
#>  -0.1573 -2.4718
#>  -5.9831 -7.6047
#> 
#> (2,1,1,.,.) = 
#>  -3.2879  2.0056
#>  -4.1304 -0.9281
#>   0.4877  1.8639
#>   9.3288  3.5154
#>  -1.7281  3.9431
#> 
#> (3,1,1,.,.) = 
#>   7.1226  1.9599
#>  -6.6945 -8.1547
#>  -1.9358  0.8171
#>   2.1278  2.5075
#>   3.3095  7.7045
#> 
#> (1,2,1,.,.) = 
#>   -8.0118   0.1788
#>   -2.2510   3.5099
#>   -6.7712  -1.9273
#>   -6.5440  10.0852
#>    4.6726   4.6771
#> 
#> (2,2,1,.,.) = 
#>  -1.6646 -1.3536
#>  -4.2981  1.0034
#>  -5.8923 -7.1970
#>  -1.0785 -4.1113
#>   6.5998 -6.8959
#> 
#> (3,2,1,.,.) = 
#>  -2.8295  6.0270
#>   5.3658  0.8249
#>  -5.4763  4.5189
#>   5.3305  9.4575
#>  -16.3867 -5.3005
#> 
#> (1,3,1,.,.) = 
#>  -1.2479 -5.4708
#>   6.6328 -4.7731
#>  -8.1953  1.1222
#>   7.8974  1.6298
#>   1.6167  3.9841
#> 
#> (2,3,1,.,.) = 
#>    3.7163   5.5667
#>   11.2345   5.2150
#>   -0.2454  -0.1170
#>    0.6909   3.2995
#>   -4.4744  -8.9865
#> 
#> (3,3,1,.,.) = 
#>   3.6110 -5.9100
#>   3.5500 -2.7090
#>  -6.9089  0.8915
#>  -0.3260  1.2351
#>   4.5273 -3.5934
#> 
#> (1,1,2,.,.) = 
#>   1.4482  0.3604
#>  -6.4091 -1.2250
#>  -1.6957 -7.5606
#>  -5.6688 -5.1901
#>  -3.7850  1.2892
#> 
#> (2,1,2,.,.) = 
#>  -4.0918  0.0221
#>  -0.8773 -5.1786
#>   0.6084 -3.6117
#>   5.6749 -1.2527
#>   3.6758  2.4225
#> 
#> (3,1,2,.,.) = 
#>  -0.2075 -2.6710
#>   4.1239 -0.2151
#>  -2.8177  0.5034
#>   4.3780  3.4107
#>  -1.0015 -2.6534
#> 
#> (1,2,2,.,.) = 
#>    1.0789   3.4573
#>   10.7834   0.7439
#>   -3.7427   3.1540
#>   -5.5350   4.7759
#>    4.8342   1.9537
#> 
#> (2,2,2,.,.) = 
#>  -0.4569  3.9635
#>  -12.4766  5.8551
#>   6.7845 -2.4719
#>   2.3759  0.7331
#>   0.3192 -1.0458
#> 
#> (3,2,2,.,.) = 
#>    0.6892  -7.7319
#>   -2.7892  16.8190
#>   -9.9950   2.9050
#>    1.5946   8.8552
#>   -0.1195  -0.4136
#> 
#> (1,3,2,.,.) = 
#>  -0.8426 -4.4367
#>  -0.9459 -4.5130
#>   7.5149  4.3272
#>  -0.4385  0.9148
#>   5.7765 -4.9552
#> 
#> (2,3,2,.,.) = 
#>  -2.4725  8.2023
#>  -10.1944 -5.0348
#>  -8.3336 -0.6324
#>  -3.1571  2.0303
#>   6.8296  4.1812
#> 
#> (3,3,2,.,.) = 
#>  -5.4066 -1.3802
#>   1.5373 -1.5903
#>   5.8244  1.9931
#>   6.5489  9.8210
#>   0.0610  3.0866
#> 
#> (1,1,3,.,.) = 
#>   0.2532  8.9893
#>   7.1315 -3.0948
#>   3.1377  0.1112
#>   4.4710 -1.2016
#>  -0.7683 -1.8460
#> 
#> (2,1,3,.,.) = 
#>  -6.0291 -3.0306
#>   2.6190 -5.7992
#>  -4.9348 -12.6836
#>  -6.4902  6.2449
#>   0.7171  2.3597
#> 
#> (3,1,3,.,.) = 
#>  -8.0247 -7.7369
#>  -4.5541  4.2686
#>  -1.8020 -2.7610
#>  -6.6145  8.0896
#>  -3.4703  4.7999
#> 
#> (1,2,3,.,.) = 
#>  -3.1619 -2.7609
#>   0.7105 -6.9211
#>  -1.6062 -2.5353
#>  -9.3931  1.6277
#>  -4.9052  3.7678
#> 
#> (2,2,3,.,.) = 
#>   6.7848  5.1247
#>   3.7982  3.1943
#>  -7.3590 -1.0131
#>   1.3404 -5.3975
#>   0.4438 -1.1979
#> 
#> (3,2,3,.,.) = 
#>   4.1927  3.5833
#>  -0.4821  8.1513
#>  -2.7147 -5.4096
#>   1.1562 -1.7988
#>  -4.0328 -2.9055
#> 
#> (1,3,3,.,.) = 
#>   2.5222 -0.7584
#>   4.0793 -0.6436
#>  -7.9806  3.9971
#>   0.5909 -4.1206
#>   0.4717  1.4096
#> 
#> (2,3,3,.,.) = 
#>   11.1922  -3.1255
#>    4.4572   0.1601
#>   -1.1980  -3.0637
#>   -4.3888   1.0556
#>   -6.6816  -7.8316
#> 
#> (3,3,3,.,.) = 
#>   8.2427 -3.0724
#>   0.4508 -2.0524
#>   1.7459  1.4590
#>  -3.5520  1.0305
#>  -14.9216 -3.3830
#> 
#> (1,1,4,.,.) = 
#>  -6.1277  3.4207
#>   3.3732  4.1625
#>  -5.7838 -1.3798
#>  -5.6210  0.2567
#>   8.5005  0.4356
#> 
#> (2,1,4,.,.) = 
#>  -3.9729  3.4417
#>  -0.2571  4.3771
#>   1.2790 -2.8964
#>  -1.1773  1.2567
#>   9.8072 -0.8487
#> 
#> (3,1,4,.,.) = 
#>  -9.0689  4.6948
#>   0.2896 -3.3769
#>   3.4208  6.0595
#>   3.1519  6.6605
#>   3.7552  1.8423
#> 
#> (1,2,4,.,.) = 
#>  -7.0038 -2.4444
#>   3.8934 -5.2337
#>  -1.9018  0.5507
#>   2.9012  3.9023
#>  -1.9298 -2.2849
#> 
#> (2,2,4,.,.) = 
#>   3.9520 -5.1728
#>   3.5540 -6.4249
#>  -4.9821 -5.9957
#>   3.9270  2.7600
#>  -0.8047 -0.6466
#> 
#> (3,2,4,.,.) = 
#>   1.4696 -0.5358
#>  -4.3046 -7.2740
#>  -1.1971 -9.8205
#>  -0.3723  2.0167
#>  -0.3459 -1.3791
#> 
#> (1,3,4,.,.) = 
#>   2.5386  7.5665
#>  -7.2466  0.5193
#>  -8.5542  0.2432
#>  -7.5853  0.9623
#>   2.8908  0.7314
#> 
#> (2,3,4,.,.) = 
#>  -4.7621 -0.1809
#>  -3.5229  1.6899
#>  -0.9608  2.3530
#>   8.1351 -4.8530
#>  -8.3305 -5.1711
#> 
#> (3,3,4,.,.) = 
#>   2.4879  6.8847
#>   2.6485 -0.6922
#>  -6.6538  2.5043
#>  -4.3979  8.0412
#>  -3.1829 -4.5154
#> 
#> (1,1,5,.,.) = 
#>  -6.8359 -1.1840
#>   5.1752  2.4042
#>  -1.3005 -5.8934
#>   2.1548 -4.2759
#>  -1.3816 -7.5355
#> 
#> (2,1,5,.,.) = 
#>  -3.0806  0.2377
#>  -2.8652  2.6075
#>   4.1835  1.8181
#>   2.3370  0.5633
#>  -0.0289 -5.6344
#> 
#> (3,1,5,.,.) = 
#>   1.3274 -0.1342
#>   2.0934 -1.7572
#>  -1.4719  0.0475
#>  -0.5094 -1.3602
#>   1.5240 -2.4890
#> 
#> (1,2,5,.,.) = 
#>  -7.5531  2.9146
#>   0.7348  5.1703
#>   1.1183  1.1718
#>   8.7563  1.4778
#>   6.3868  0.2656
#> 
#> (2,2,5,.,.) = 
#>   -1.0910   2.3091
#>   10.0869   1.1438
#>    2.4395   4.9528
#>    3.7323   3.2165
#>    2.5481   4.5091
#> 
#> (3,2,5,.,.) = 
#>  -2.0743  1.7999
#>  -4.3825  4.2049
#>  -3.8392 -0.5473
#>   2.5288  4.0873
#>  -3.7846 -6.1437
#> 
#> (1,3,5,.,.) = 
#>  -4.0471  6.5153
#>   1.5204  4.9631
#>   0.1587  0.0017
#>  -1.7989 -0.7558
#>  -11.6536 -3.9174
#> 
#> (2,3,5,.,.) = 
#>  -3.0316  1.7210
#>  -2.6083 -1.3017
#>  -0.5551 -0.6676
#>   0.3432  9.7838
#>  -1.6319  3.1394
#> 
#> (3,3,5,.,.) = 
#>  -1.4190 -4.2193
#>  -4.5579  1.9284
#>   5.0404 -2.6589
#>   0.0624 -4.9452
#>   5.6924 -1.0265
#> [ CPUFloatType{3,3,5,5,2} ]