Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -5.6637 -2.1076
#>   0.9547 -7.0030
#>  -4.6167 -2.3238
#> 
#> (2,.,.) = 
#>  -6.4385  1.3334
#>   7.9300 -2.4287
#>   3.6834 -2.8553
#> 
#> (3,.,.) = 
#>   8.1408  2.3418
#>   6.7841 -2.6787
#>   2.4555 -3.4944
#> 
#> (4,.,.) = 
#>   3.6365 -0.5213
#>  -5.1113 -0.9556
#>   6.5734  0.2691
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.0812  0.2616
#>   2.6394 -3.2665
#>   2.0239 -2.1011
#> 
#> (2,.,.) = 
#>  -3.9148 -3.6311
#>  -1.0890  2.1793
#>  -0.9870 -0.4299
#> 
#> (3,.,.) = 
#>   1.3198 -0.1445
#>   1.2300 -1.5743
#>  -3.1045 -0.8080
#> 
#> (4,.,.) = 
#>  -2.9875  1.4064
#>  -1.8256 -4.3414
#>  -2.5492  1.0151
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -4.5525  2.1773
#>   1.8994 -1.6680
#>  -12.4760 -3.1391
#>   5.1533  0.9916
#>   7.4149  3.1399
#> 
#> (2,1,1,.,.) = 
#>  -5.3533 -1.5458
#>  -4.5300 -0.6545
#>  -0.2051 -6.8311
#>   7.3933  0.0389
#>  -7.0205  4.2245
#> 
#> (3,1,1,.,.) = 
#>   10.1423   6.3049
#>    1.2327  -2.5759
#>   -5.9249  -1.7535
#>    0.1570   2.5355
#>    9.4530  -2.8292
#> 
#> (1,2,1,.,.) = 
#>  -10.5461   1.0989
#>   -2.7024  -2.2106
#>   -2.4231  -6.1884
#>    3.8419  -5.8903
#>    2.0524  -1.6502
#> 
#> (2,2,1,.,.) = 
#>   5.8860 -6.7430
#>   5.6067  2.3071
#>  -2.4538 -2.5957
#>  -4.5727 -3.5119
#>   0.0494 -12.7552
#> 
#> (3,2,1,.,.) = 
#>   9.4247 -2.0158
#>   3.9146 -1.8329
#>   0.5730  4.9105
#>  -3.7584 -7.6872
#>  -3.0141 -14.7912
#> 
#> (1,3,1,.,.) = 
#>  -0.1175  0.7461
#>  -3.6812  2.3983
#>  -1.8834 -2.4979
#>   0.6107 -1.8714
#>  -9.1818 -3.6712
#> 
#> (2,3,1,.,.) = 
#>  -5.1770  4.2797
#>  -3.0670 -1.7504
#>   8.7028  3.2123
#>  -8.1884  3.6448
#>   4.9256 -1.4841
#> 
#> (3,3,1,.,.) = 
#>  -3.8231 -3.7175
#>  -1.2035 -4.4494
#>   2.8281  1.4049
#>  -0.0996 -4.2788
#>  -3.0485 -3.2423
#> 
#> (1,1,2,.,.) = 
#>   2.9279 -3.3450
#>   0.8204  5.1580
#>  -4.1785  4.6399
#>   2.7079  1.8296
#>   1.6651  3.8334
#> 
#> (2,1,2,.,.) = 
#>   0.9783  0.4571
#>  -4.7573  3.7389
#>   1.4727 -2.7429
#>   2.7269  4.0189
#>  -4.4311  1.2037
#> 
#> (3,1,2,.,.) = 
#>   0.1485  7.6040
#>   3.4795  4.5886
#>  -1.7142  0.6365
#>  -7.9676  6.5320
#>   6.4592  6.4216
#> 
#> (1,2,2,.,.) = 
#>  -2.6179 -7.8683
#>  -9.8198  1.9502
#>  -1.4846  3.7815
#>  -2.3778  1.4130
#>  -3.9947  0.7845
#> 
#> (2,2,2,.,.) = 
#>   1.2918 -3.3648
#>  -5.9235 -1.0471
#>  -6.7918 -4.2295
#>   6.4402 -0.0234
#>   0.5866  0.8474
#> 
#> (3,2,2,.,.) = 
#>   7.3033  7.8553
#>  -2.7509  2.1777
#>   4.0027 -7.0071
#>  -0.7114  0.7998
#>  -5.3065 -5.8956
#> 
#> (1,3,2,.,.) = 
#>  -1.2065  0.0890
#>  -0.6931 -0.4565
#>   5.7808 -3.4627
#>   7.3028 -2.8341
#>   6.3712  8.3189
#> 
#> (2,3,2,.,.) = 
#>   5.6793 -4.8825
#>  -3.4379 -3.2304
#>  -6.5984  4.5589
#>   0.1343 -0.5841
#>  -4.2133 -9.0516
#> 
#> (3,3,2,.,.) = 
#>    3.0322  -9.6899
#>   -1.7361  -2.1461
#>   -8.4151  15.9352
#>    3.7936   9.7714
#>   -3.6344   2.6097
#> 
#> (1,1,3,.,.) = 
#>   -2.6307  10.6810
#>    1.8843  14.6622
#>    2.2835   3.2567
#>   -3.5046  -1.3614
#>    6.0555  -0.4788
#> 
#> (2,1,3,.,.) = 
#>   0.8562  1.5985
#>  -3.8223 -2.8376
#>   2.7286  1.3915
#>   3.5480 -8.8867
#>   2.0659 -2.5972
#> 
#> (3,1,3,.,.) = 
#>    6.8362   3.7538
#>   11.4146  -0.7612
#>   -2.7447   4.1264
#>   -2.2480  -5.6317
#>    2.3473  -2.0123
#> 
#> (1,2,3,.,.) = 
#>   3.8496  4.9664
#>   0.6597 -5.6139
#>  -3.9334  3.5423
#>  -0.0825  4.5088
#>  -1.1086  2.5114
#> 
#> (2,2,3,.,.) = 
#>  -5.2044  5.7947
#>  -1.4427 -1.1540
#>   3.7128  7.4319
#>  -0.2890 -3.7897
#>  -6.9078 -3.3708
#> 
#> (3,2,3,.,.) = 
#>   3.6492 -0.4633
#>  -3.1714  6.5610
#>   4.2153  0.5012
#>  -11.7703 -0.1013
#>  -4.7784  4.8675
#> 
#> (1,3,3,.,.) = 
#>   0.0522  4.8103
#>  -7.9503  3.9363
#>  -3.7653  2.6827
#>  -2.2454  7.5609
#>  -0.8419  1.3196
#> 
#> (2,3,3,.,.) = 
#>   8.1436 -7.7611
#>   7.5454 -1.7605
#>  -3.5587 -3.1433
#>  -3.9565  6.0329
#>   4.9573 -5.9890
#> 
#> (3,3,3,.,.) = 
#>   2.3643 -4.6361
#>  -1.6484  0.8950
#>  -1.1303  3.9882
#>  -4.1082 -3.2615
#>   2.7415 -0.5491
#> 
#> (1,1,4,.,.) = 
#>  -3.4548  1.4661
#>   4.3341 -0.4220
#>   8.3631  6.6964
#>   9.8601 -2.8944
#>   2.1018 -3.5435
#> 
#> (2,1,4,.,.) = 
#>   0.6837  1.5887
#>  -2.9638  0.9687
#>   1.1870 -5.6933
#>  -2.9973 -4.6332
#>  -0.4390 -7.0192
#> 
#> (3,1,4,.,.) = 
#>   10.3982  -1.4188
#>   -0.3769  -1.7225
#>    6.1651  -5.1107
#>   10.9318  -4.4133
#>    2.7077   4.0205
#> 
#> (1,2,4,.,.) = 
#>  -1.9926  3.0240
#>   0.6286 -3.1455
#>  -4.9613  0.8279
#>  -5.6725 -1.2915
#>  -0.1389  3.7048
#> 
#> (2,2,4,.,.) = 
#>   10.6865  -2.6885
#>   -4.9905   2.2381
#>    2.2129  -3.1043
#>   -1.5625  -0.9077
#>    1.4797   0.9484
#> 
#> (3,2,4,.,.) = 
#>   5.1565 -0.7315
#>  -7.0653 -11.7681
#>  -4.9715 -1.5790
#>  -1.7234 -3.1639
#>  -1.1817 -0.2975
#> 
#> (1,3,4,.,.) = 
#>  -0.1950  5.4184
#>   2.9121  4.4204
#>  -2.3849  0.5550
#>  -4.5078 -1.9613
#>  -0.9835  5.3547
#> 
#> (2,3,4,.,.) = 
#>  -0.1669  2.8306
#>  -3.1221 -3.3290
#>   1.0161  1.7840
#>  -1.9536 -2.3344
#>  -12.2201 -3.2027
#> 
#> (3,3,4,.,.) = 
#>   -1.4115 -11.2720
#>    1.3938   7.7327
#>   -2.9900   7.3670
#>   10.3688   5.3575
#>    2.9307  -3.7862
#> 
#> (1,1,5,.,.) = 
#>  -4.2870 -2.7373
#>   0.5974  1.1386
#>   0.2076  5.2425
#>  -0.9731 -1.0306
#>   5.3132  1.9261
#> 
#> (2,1,5,.,.) = 
#>  -8.5631  6.8058
#>  -1.5390 -10.3426
#>   8.9559  4.2994
#>  -2.4208 -7.8832
#>  -7.1500  2.7781
#> 
#> (3,1,5,.,.) = 
#>   2.8227  6.4259
#>  -0.0112 -2.5651
#>   0.6540 -1.1196
#>  -3.7278  3.1086
#>   1.4334 -3.3960
#> 
#> (1,2,5,.,.) = 
#>  -3.2971  0.3874
#>   6.3399 -7.2229
#>  -1.0153 -7.7427
#>   6.4741  0.3096
#>  -11.2081  3.8212
#> 
#> (2,2,5,.,.) = 
#>    8.1294  -3.2503
#>   -5.1611  -3.1956
#>   -9.6591  11.7951
#>   -3.1094  -1.6623
#>    2.2709 -15.9624
#> 
#> (3,2,5,.,.) = 
#>   8.3558 -6.5961
#>   8.1063 -1.5091
#>  -4.5716  5.7837
#>  -10.7584 -2.2899
#>   1.7811  4.9996
#> 
#> (1,3,5,.,.) = 
#>  -11.1415   1.8413
#>    5.5735   5.7982
#>   -7.8669   8.4188
#>   -4.7714  -5.8866
#>   -2.3149  -4.0908
#> 
#> (2,3,5,.,.) = 
#>    0.1157  10.9325
#>   -1.5922  -2.0848
#>   -2.7198   4.2186
#>    1.4657  -5.1653
#>   -6.1043  -9.7570
#> 
#> (3,3,5,.,.) = 
#>   10.6174   1.9617
#>  -17.2405   8.1538
#>    0.8857  -1.4762
#>   -0.4582  -5.2782
#>   -1.7017   1.2171
#> [ CPUFloatType{3,3,5,5,2} ]