Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.2124 -2.9043
#>   0.6629  3.7074
#>   7.7124  1.8205
#> 
#> (2,.,.) = 
#>  -4.3501 -1.3243
#>   6.0236  1.1549
#>  -0.8779 -1.0770
#> 
#> (3,.,.) = 
#>  -3.8956 -1.5054
#>  -2.6960  8.9512
#>  -1.0268  2.3806
#> 
#> (4,.,.) = 
#>  -4.7197  2.7201
#>  -7.1973 -2.3913
#>   0.3703 -2.4405
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -3.1882 -0.7535
#>  -0.8017  2.8556
#>   1.5445  0.1709
#> 
#> (2,.,.) = 
#>   2.0381 -0.2573
#>  -0.0468  1.9943
#>   1.8439 -0.4521
#> 
#> (3,.,.) = 
#>   1.3467 -1.4514
#>  -0.2148  3.4738
#>   1.7983  1.9297
#> 
#> (4,.,.) = 
#>   0.0159 -0.4421
#>   1.7262 -4.6162
#>   2.5257  0.1720
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   7.0034 -3.3739
#>   2.9026 -4.5238
#>  -0.3015 -8.1013
#>  -5.0326  3.4924
#>   8.2646  7.9422
#> 
#> (2,1,1,.,.) = 
#>   13.9873  -0.1815
#>    0.2027  -3.1822
#>    0.3530  -4.6858
#>   -4.1832  -4.5289
#>    0.2098  -3.1896
#> 
#> (3,1,1,.,.) = 
#>   0.2200  7.2414
#>  -3.1234  1.5075
#>  -2.4388  8.4770
#>  -0.0020  5.5614
#>  -7.2743  1.0887
#> 
#> (1,2,1,.,.) = 
#>   6.9750  6.9759
#>  -5.5142  7.6291
#>  -1.3756  1.3416
#>  -5.4625  3.7384
#>  -0.2157 -0.9100
#> 
#> (2,2,1,.,.) = 
#>   0.4912  0.2432
#>  -0.0037  5.6690
#>   4.0289  5.1035
#>   4.8726  4.7669
#>   9.0925 -2.6799
#> 
#> (3,2,1,.,.) = 
#>   -1.3676  -3.8226
#>   -1.7582   2.6489
#>    2.3188  -5.6156
#>   11.5774   2.7108
#>   -0.1385  -3.8296
#> 
#> (1,3,1,.,.) = 
#>   2.5444  2.5049
#>  -0.6736  2.3402
#>   3.9977 -2.7197
#>   0.9093 -9.3219
#>  -0.5731  2.9002
#> 
#> (2,3,1,.,.) = 
#>   -3.5332  -1.3621
#>    6.1941  -3.3272
#>  -16.4685  -0.9530
#>   10.1082   2.6200
#>  -14.7453  -0.0665
#> 
#> (3,3,1,.,.) = 
#>  -9.0459 -4.0066
#>   5.9781  3.2203
#>  -8.1101  0.3492
#>   3.9486  0.1935
#>  -0.2101  3.0832
#> 
#> (1,1,2,.,.) = 
#>   6.8221 -1.0917
#>   4.2981 -1.6653
#>   1.2459  7.2199
#>   7.5586 -6.0287
#>  -3.3932 -6.3751
#> 
#> (2,1,2,.,.) = 
#>   0.5147 -5.8863
#>  -6.1975 -5.2635
#>  -1.2043 -6.8308
#>  -1.7762 -7.3359
#>   1.5681  5.5212
#> 
#> (3,1,2,.,.) = 
#>   7.1826 -1.1496
#>   5.6576 -1.0362
#>  -3.3273  5.8028
#>   2.6661 -1.1525
#>  -1.5057 -1.7252
#> 
#> (1,2,2,.,.) = 
#>   12.9120  -1.6614
#>   -2.2206   5.4326
#>   -1.9834   5.4489
#>    4.8002   3.7252
#>    0.3921  -5.8176
#> 
#> (2,2,2,.,.) = 
#>   1.6195 -0.3149
#>   2.6310  4.0859
#>  -3.6599 -4.5415
#>   8.4359 -12.8497
#>   4.2812 -4.6817
#> 
#> (3,2,2,.,.) = 
#>   1.4138  5.3463
#>  -1.3473 -7.9508
#>   1.4263 -5.3304
#>   6.2822  5.3611
#>  -4.7661  3.7207
#> 
#> (1,3,2,.,.) = 
#>    3.9539  -6.4945
#>   12.6224  -2.1573
#>    5.0273   3.7716
#>    8.7735  -2.4397
#>    7.6593  -4.7788
#> 
#> (2,3,2,.,.) = 
#>  -9.7344 -0.3576
#>   5.7387  0.4594
#>   5.4515  5.5775
#>   3.6939  4.0443
#>  -3.4020 -1.5889
#> 
#> (3,3,2,.,.) = 
#>   2.7452 -8.0257
#>  -1.5516 -5.8047
#>   6.0922  8.3608
#>   1.0253  9.5067
#>  -8.9769  2.1273
#> 
#> (1,1,3,.,.) = 
#>  -2.9889 -2.5961
#>  -3.1773 -2.0112
#>   0.6871 -0.1214
#>  -3.2890  4.8310
#>   0.9614 -1.4463
#> 
#> (2,1,3,.,.) = 
#>    3.3030   7.8981
#>   -4.0408  -0.5365
#>    0.7454  11.8755
#>    1.5206  -9.1710
#>    1.9281  -1.7626
#> 
#> (3,1,3,.,.) = 
#>  -2.6860 -1.2511
#>   0.4848  2.4394
#>   7.7876 -3.0543
#>   1.2969 -6.7868
#>  -0.3404 -6.7096
#> 
#> (1,2,3,.,.) = 
#>  -1.8890  2.8122
#>  -14.0611  0.5173
#>  -3.1395 -11.9238
#>  -2.0671 -1.2772
#>  -5.3502 -4.3681
#> 
#> (2,2,3,.,.) = 
#>    3.4511   3.7946
#>    7.1307  -4.0754
#>   -0.9505  -2.8545
#>    1.0426  -9.7128
#>   10.4856  -2.1682
#> 
#> (3,2,3,.,.) = 
#>  -1.3000 -1.8614
#>  -1.6181 -2.6144
#>  -5.2889  1.0418
#>  -0.6821  2.8147
#>  -4.3889  0.3855
#> 
#> (1,3,3,.,.) = 
#>    5.0083   2.9325
#>    2.0293  -0.5752
#>   -2.4302   0.6194
#>    0.3069   3.8420
#>   -2.1369  12.3418
#> 
#> (2,3,3,.,.) = 
#>   0.4131 -2.5776
#>  -0.0740 -4.0399
#>   4.6331 -3.1218
#>  -2.6776 -0.8250
#>  -2.1270 -1.5478
#> 
#> (3,3,3,.,.) = 
#>  -0.9269  6.5413
#>  -5.5818 -2.9819
#>   4.1817  1.1501
#>   0.2835  4.3732
#>   5.7136 -2.9817
#> 
#> (1,1,4,.,.) = 
#>  -0.0025  5.8975
#>   0.7282  3.9016
#>  -1.9950 -4.8600
#>   1.9989 -8.0581
#>  -1.3200  0.3174
#> 
#> (2,1,4,.,.) = 
#>  -3.8473  0.4840
#>   2.4654  4.8428
#>   5.2698  9.1460
#>  -2.9662  0.7100
#>   0.9567 -0.1974
#> 
#> (3,1,4,.,.) = 
#>   2.1032  1.4425
#>   0.6540  1.1908
#>  -10.9464  0.0111
#>   4.1792 -2.4738
#>  -4.7132 -7.2973
#> 
#> (1,2,4,.,.) = 
#>   3.5473 -8.3967
#>  -3.2588  7.5647
#>  -4.7182 -4.9149
#>   0.0950 -3.2508
#>  -1.9516 -4.1510
#> 
#> (2,2,4,.,.) = 
#>   0.9884  8.2706
#>   5.4046  0.9698
#>  -4.0728 -2.4917
#>  -1.7438  5.4274
#>   1.8805 -2.6768
#> 
#> (3,2,4,.,.) = 
#>  -0.5912 -0.1012
#>   2.5140 -3.5852
#>  -1.4059  6.9588
#>   7.1158 -3.4079
#>  -0.5703 -4.1705
#> 
#> (1,3,4,.,.) = 
#>    8.3640  13.7956
#>    4.1674  -8.8968
#>   -1.4129   7.1990
#>   -0.0638  -8.3858
#>    0.2497  -2.4443
#> 
#> (2,3,4,.,.) = 
#>  -2.5747 -4.0089
#>  -0.5100 -8.0300
#>   6.7691 -7.7204
#>  -14.6511 -3.6499
#>   1.7411 -11.0932
#> 
#> (3,3,4,.,.) = 
#>   4.4685 -3.3192
#>   2.1725 -5.7364
#>   4.6440 -2.9757
#>   6.9269 -2.7998
#>  -1.7848 -1.7122
#> 
#> (1,1,5,.,.) = 
#>  -4.6235 -7.2794
#>  -5.8728 -4.8818
#>   8.3327  2.9795
#>   5.0832  7.9213
#>   3.6636 -5.1307
#> 
#> (2,1,5,.,.) = 
#>  -4.4955 -1.4163
#>  -4.4546  0.7812
#>  -0.8027 -1.9950
#>   1.1379 -2.4070
#>   1.8446  1.8724
#> 
#> (3,1,5,.,.) = 
#>  -5.2251  3.9684
#>  -2.1141 -1.3469
#>  -1.8861  1.1904
#>   7.3177  3.7804
#>   1.9051 -4.8116
#> 
#> (1,2,5,.,.) = 
#>  -3.6341 -0.2675
#>  -2.4974 -0.7319
#>  -0.5365  8.3733
#>  -2.9215 -6.7778
#>   1.2044 -0.7085
#> 
#> (2,2,5,.,.) = 
#>  -1.4074  1.9528
#>  -1.9597  0.3941
#>  -2.4970  2.3196
#>  -0.0587  7.4883
#>  -3.7092  3.8894
#> 
#> (3,2,5,.,.) = 
#>   1.4631  2.9241
#>   0.6604  1.4797
#>  -4.1251  6.3859
#>  -3.9198  2.0043
#>   8.0232 -0.6009
#> 
#> (1,3,5,.,.) = 
#>   2.5063  4.1150
#>  -4.8869  1.1597
#>   3.0749 -4.3662
#>   2.7937 -0.4277
#>   4.4206  5.2969
#> 
#> (2,3,5,.,.) = 
#>  -5.5998  2.3414
#>  -7.0827 -11.2121
#>   1.4948  0.9766
#>   3.0592  3.8057
#>  -1.2724  4.8231
#> 
#> (3,3,5,.,.) = 
#>  -2.1112  4.3622
#>  -5.4069 -7.2807
#>  -2.0093 -2.2248
#>   0.3550 -5.1748
#>  -7.9049 -0.7205
#> [ CPUFloatType{3,3,5,5,2} ]