Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.4479  1.7247
#>   0.0303  3.8190
#>   2.9478 -6.6245
#> 
#> (2,.,.) = 
#>  -0.8296 -2.5042
#>  -4.1798 -4.6895
#>  -2.4143  0.2780
#> 
#> (3,.,.) = 
#>  -1.8542 -1.3998
#>  -1.9779 -0.0040
#>   2.6931 -4.1032
#> 
#> (4,.,.) = 
#>  -0.0169 -2.3904
#>   5.4750 -6.5047
#>  -2.5330 -4.7945
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.0632 -1.1424
#>  -0.1631 -1.8448
#>   0.1734 -3.8111
#> 
#> (2,.,.) = 
#>   1.1040  0.5780
#>   0.0483 -1.4580
#>  -1.2045 -0.6006
#> 
#> (3,.,.) = 
#>   0.3601  1.3049
#>  -0.8107  3.7523
#>   2.6471 -1.5528
#> 
#> (4,.,.) = 
#>   1.0471  0.9843
#>   0.9559  3.3695
#>   1.3318 -0.6600
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   0.7527  4.7233
#>   1.1107 -1.6034
#>  -1.1288  7.2458
#>  -2.0895  1.2062
#>  -1.9889 -2.2026
#> 
#> (2,1,1,.,.) = 
#>   2.5239 -3.3325
#>  -0.0455 -3.0254
#>   5.7056  5.1097
#>   2.5614 -4.4375
#>  -3.0891  0.0541
#> 
#> (3,1,1,.,.) = 
#>  -4.6930  0.4070
#>   8.9145 -2.7871
#>   1.8351 -1.2042
#>  -8.8374  5.8588
#>   0.4835  5.4560
#> 
#> (1,2,1,.,.) = 
#>   1.4334  4.0562
#>  -5.0231 -1.4061
#>   0.7277 -10.2196
#>  -7.8245 -9.9693
#>   7.5000  4.3706
#> 
#> (2,2,1,.,.) = 
#>  -0.4763 -9.4084
#>  -4.5998 -2.4029
#>   2.5728 -5.9447
#>  -3.4640 -0.4149
#>  -3.0853  5.5600
#> 
#> (3,2,1,.,.) = 
#>  -3.1669 -2.8545
#>   6.4699  7.7968
#>   0.3474  1.7397
#>   1.1069 -2.4688
#>   0.9427 -1.3195
#> 
#> (1,3,1,.,.) = 
#>  -1.5867  7.5420
#>   7.8240  9.6579
#>  -3.8679 -7.4911
#>   0.4405 -0.3507
#>  -0.2370  6.2885
#> 
#> (2,3,1,.,.) = 
#>   0.5082 -5.1379
#>   1.3169 -5.6430
#>  -1.6033  1.0669
#>   1.4336  8.4442
#>   3.5123 -4.3534
#> 
#> (3,3,1,.,.) = 
#>   4.2715 -1.2230
#>  -2.5883  0.0352
#>  -0.8732 -3.1806
#>   1.5521  3.8832
#>  -0.5002  4.1758
#> 
#> (1,1,2,.,.) = 
#>  -15.2861  -2.0739
#>    9.8440   0.7404
#>   -9.2140  -1.5601
#>  -12.4612   0.7624
#>    1.9264  11.9751
#> 
#> (2,1,2,.,.) = 
#>   2.6130  0.0302
#>   1.9686  0.7061
#>  -4.7991  5.6087
#>   0.4950 -0.9630
#>   5.1168 -0.4036
#> 
#> (3,1,2,.,.) = 
#>   2.2228 -6.9407
#>   3.1760  7.5351
#>  -2.7847 -4.5022
#>  -1.7657 -4.6054
#>   0.7721 -0.7540
#> 
#> (1,2,2,.,.) = 
#>  -6.6058 -1.6762
#>  -12.1628  4.4440
#>  -0.9288  2.4407
#>  -4.8412  1.2220
#>   2.5685 -3.3796
#> 
#> (2,2,2,.,.) = 
#>   0.6475  2.3817
#>  -8.2421  3.1329
#>   6.2684 -8.2152
#>  -4.7312  1.0092
#>   3.6508  1.0121
#> 
#> (3,2,2,.,.) = 
#>    0.1284  -9.4264
#>   -6.3953   9.5668
#>    1.1269  -2.0367
#>    0.6462   2.1009
#>   12.2894 -10.2699
#> 
#> (1,3,2,.,.) = 
#>    0.2963   0.7716
#>   -4.6067   4.8959
#>   10.8507  -3.2136
#>   -3.3122   7.4117
#>    2.7439  -2.7347
#> 
#> (2,3,2,.,.) = 
#>    3.3813  10.5136
#>   -5.4176  -0.5910
#>    0.5812   0.1147
#>    1.3848   2.5612
#>    3.4539  -4.1325
#> 
#> (3,3,2,.,.) = 
#>  -1.8952 -1.7729
#>  -8.2241  4.7494
#>   1.3378 -1.2935
#>   4.1577  1.4674
#>  -3.8006 -1.8525
#> 
#> (1,1,3,.,.) = 
#>   3.2716 -0.0844
#>  -0.9288 -5.9343
#>  -0.6564  2.9996
#>  -1.0032 -12.5445
#>  -7.5028 -6.5999
#> 
#> (2,1,3,.,.) = 
#>  -0.3073  4.4502
#>  -4.6778  1.8989
#>  -7.7949  0.0072
#>   0.9479  0.5096
#>   8.5118 -0.0803
#> 
#> (3,1,3,.,.) = 
#>   5.8228 -10.2738
#>  -6.1445  6.4883
#>  -0.7179 -1.2352
#>  -4.2617 -4.9985
#>  -3.9160  0.4372
#> 
#> (1,2,3,.,.) = 
#>  -5.6095  0.7792
#>   2.0590 -1.4150
#>  -2.9470  6.7639
#>  -1.6894  2.9427
#>  -1.5463  6.1121
#> 
#> (2,2,3,.,.) = 
#>  -1.5441 -4.4826
#>   7.7192  0.2779
#>  -4.1749 -2.6230
#>   0.3269 -5.6010
#>   1.8583  3.7782
#> 
#> (3,2,3,.,.) = 
#>   10.5526 -16.0190
#>   11.4771  -3.0946
#>   -0.5745   6.1031
#>    2.5448  -1.5423
#>    4.6085  -1.8969
#> 
#> (1,3,3,.,.) = 
#>   3.6850 -1.3997
#>  -1.7118  3.9654
#>   1.6243  8.4932
#>   1.4250 -0.2749
#>   7.3751  1.9406
#> 
#> (2,3,3,.,.) = 
#>   14.1170   0.3336
#>    0.4044  -8.2961
#>   -2.3835   1.8472
#>    9.8789  -2.2399
#>   -3.2840   1.1951
#> 
#> (3,3,3,.,.) = 
#>   11.7709   4.2859
#>   -1.2228  -1.7863
#>   11.7607  -2.2686
#>    7.1019  -2.5155
#>    1.1271  -3.0592
#> 
#> (1,1,4,.,.) = 
#>  -5.2384 -2.1718
#>   1.3243  5.2300
#>  -14.3625 -9.6325
#>   2.5367 -3.1967
#>  -7.5268 -3.7019
#> 
#> (2,1,4,.,.) = 
#>  -2.6061 -1.0681
#>  -3.4590  2.3695
#>  -1.0337  6.5555
#>   4.9368  3.8680
#>  -2.6583 -6.1064
#> 
#> (3,1,4,.,.) = 
#>   3.6191  5.0019
#>   7.4662 -1.1228
#>   3.7275 -0.5510
#>   4.2716  3.0878
#>  -2.7627  3.5826
#> 
#> (1,2,4,.,.) = 
#>  -6.4852 -2.2399
#>  -1.4136 -2.1355
#>   2.4792 -1.9102
#>   0.2365 -5.5257
#>  -3.8569 -5.0402
#> 
#> (2,2,4,.,.) = 
#>   -0.9495   2.0661
#>   -5.4033  -1.2020
#>    4.3704  -5.6345
#>    7.9201  15.8160
#>    4.0996   3.5328
#> 
#> (3,2,4,.,.) = 
#>   2.3758 -6.9754
#>  -12.6388 -0.5723
#>  -3.5028  7.0670
#>  -9.5424 -8.0029
#>   5.1015 -4.8473
#> 
#> (1,3,4,.,.) = 
#>  -0.2896  0.9728
#>   0.4816  0.8108
#>  -1.3237  8.4196
#>   4.5934  4.3511
#>  -5.0212 -7.8805
#> 
#> (2,3,4,.,.) = 
#>  -11.1272  -0.5699
#>   -1.5478   9.0661
#>    3.0963  -1.6386
#>   -7.4762   8.3321
#>    9.6352   1.2744
#> 
#> (3,3,4,.,.) = 
#>   1.3846  5.8532
#>  -0.1392 -5.5595
#>   1.0024 -0.5335
#>   0.4229 -6.8150
#>   1.5420  2.8378
#> 
#> (1,1,5,.,.) = 
#>   6.2351  5.1958
#>   4.0373  2.3533
#>  -4.8082 -3.8346
#>  -6.1393  3.8619
#>   1.8403  5.0844
#> 
#> (2,1,5,.,.) = 
#>  -2.3618 -4.4392
#>   7.5157  6.0276
#>   0.5744  5.0586
#>   4.0473 -1.0463
#>  -3.0962  6.9731
#> 
#> (3,1,5,.,.) = 
#>  -7.5277 -4.5908
#>  -0.4506  2.0664
#>   8.9169 -0.1145
#>   3.3999 -1.7606
#>  -1.2544 -4.5004
#> 
#> (1,2,5,.,.) = 
#>   4.2890 -4.3465
#>   1.8047 -5.4693
#>  -2.2052  1.6576
#>  -7.5223  6.1956
#>   8.2277 -7.6577
#> 
#> (2,2,5,.,.) = 
#>   4.0439 -3.3719
#>  -1.8503 -2.8222
#>   3.1521  2.2759
#>   3.8508 -0.9597
#>   0.1940 -3.9425
#> 
#> (3,2,5,.,.) = 
#>  -0.2715 -2.6924
#>  -5.6460  3.3038
#>  -0.4296 -3.0039
#>  -5.2324  2.0920
#>   4.6862  1.4800
#> 
#> (1,3,5,.,.) = 
#>  -1.1422 -1.5753
#>  -6.1828  3.5928
#>   4.3618  0.9756
#>  -7.0825 -1.7252
#>  -9.9568  7.6763
#> 
#> (2,3,5,.,.) = 
#>  -1.2346 -3.0968
#>  -5.9099  2.9643
#>   4.2070 -1.8007
#>   4.9160  2.6218
#>   1.1672  1.0549
#> 
#> (3,3,5,.,.) = 
#>   1.1886  0.3977
#>  -3.2889 -0.1124
#>  -4.6003  1.1726
#>   2.9545  0.5235
#>   8.3391 -2.2230
#> [ CPUFloatType{3,3,5,5,2} ]