Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>   3.7597 -1.1886
#>  -1.9001 -5.6069
#>   1.1265  1.4221
#> 
#> (2,.,.) = 
#>   3.8713  0.8145
#>  -0.5266  1.7461
#>   0.3583 -1.9295
#> 
#> (3,.,.) = 
#>   5.2281 -7.4829
#>   0.0981  1.7069
#>   5.2268 -4.2273
#> 
#> (4,.,.) = 
#>  -0.7460  3.9434
#>   0.6337  3.1282
#>   0.6950 -2.0228
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>   3.0283 -0.9784
#>  -0.4237  0.2436
#>   1.8517 -1.6894
#> 
#> (2,.,.) = 
#>   0.4151  2.7279
#>  -0.1540 -2.1185
#>  -1.0484  1.3282
#> 
#> (3,.,.) = 
#>   1.4656 -3.3573
#>  -0.4773 -2.1936
#>   1.3250  0.2868
#> 
#> (4,.,.) = 
#>  -1.1493  0.4193
#>  -0.8451 -1.5383
#>  -1.0017  1.4965
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   3.9048  3.1432
#>   0.0906 -1.4782
#>  -3.1744 -0.6598
#>   4.5718  3.0818
#>   0.3719 -1.1976
#> 
#> (2,1,1,.,.) = 
#>   -7.6779   2.9652
#>   -0.1711  -0.2902
#>   -5.0236   1.5802
#>    8.4835   4.0315
#>   -0.5778  13.5156
#> 
#> (3,1,1,.,.) = 
#>   -9.1267   0.3414
#>   12.6732   2.5624
#>   -1.6024  -1.1598
#>   -2.7183   7.7227
#>   10.5616  -1.0794
#> 
#> (1,2,1,.,.) = 
#>  -6.6029  7.6129
#>  -3.0540 -0.2958
#>  -8.3002  5.1049
#>   7.8164  5.5754
#>   3.0946  0.7063
#> 
#> (2,2,1,.,.) = 
#>  -2.8983  2.7208
#>   0.6131 -3.3917
#>  -4.0286 -5.1291
#>   4.0895  2.4884
#>   5.0024  3.4807
#> 
#> (3,2,1,.,.) = 
#>   6.0266 -2.8156
#>   2.3848  5.1107
#>   1.2038 -1.4473
#>  -0.5940 -3.3553
#>   7.7593  4.5381
#> 
#> (1,3,1,.,.) = 
#>   3.0706  4.2322
#>  -5.8536  0.6576
#>   1.8932 -3.3699
#>   6.9478  9.2783
#>  -0.3539  6.4962
#> 
#> (2,3,1,.,.) = 
#>    1.1079   2.9124
#>   -1.6890   5.6916
#>   -6.1432  -9.3719
#>   -0.6300  15.1821
#>    0.7383   4.2963
#> 
#> (3,3,1,.,.) = 
#>   0.2897 -1.2233
#>   1.0957  3.8539
#>  -3.3942  6.6180
#>   4.9179  1.9841
#>   4.3170  8.6111
#> 
#> (1,1,2,.,.) = 
#>   -1.9210  12.4019
#>    5.4754   0.2391
#>   -3.4419   4.0215
#>    3.6538  -1.8343
#>   -0.9536   5.1797
#> 
#> (2,1,2,.,.) = 
#>  -5.7539 -5.6602
#>   0.2156 -3.7421
#>   7.2037 -4.2497
#>  -4.2686 -0.7439
#>  -3.5559 -6.5582
#> 
#> (3,1,2,.,.) = 
#>  -2.1671 -7.5903
#>   6.0292 -0.6595
#>   0.1952 -2.2188
#>  -6.6567 -0.8834
#>  -1.0638  4.1578
#> 
#> (1,2,2,.,.) = 
#>   8.5629  2.7812
#>  -2.8361 -10.1750
#>  -7.0520 -9.2168
#>  -2.4850 -0.4513
#>  -0.6809 -1.1388
#> 
#> (2,2,2,.,.) = 
#>   5.5307  7.4056
#>   0.5267  4.6140
#>   5.0932 -2.7840
#>   3.6697  3.0834
#>  -2.9730  1.8544
#> 
#> (3,2,2,.,.) = 
#>   5.6465 -4.9537
#>   1.1380 -4.0479
#>   2.9763  5.4800
#>   5.3889 -4.4113
#>   3.5672 -1.3771
#> 
#> (1,3,2,.,.) = 
#>  -5.7418  1.5811
#>  -3.9376 -7.6909
#>  -0.8321 -1.0367
#>  -0.2761  0.0475
#>   0.3542 -5.5879
#> 
#> (2,3,2,.,.) = 
#>   1.8804  1.1318
#>   4.9749 -2.1704
#>   4.0903  1.7678
#>  -8.3534  0.4683
#>   7.3973 -6.7424
#> 
#> (3,3,2,.,.) = 
#>   4.9704 -2.0084
#>  -3.1934 -4.6632
#>  -0.2653  6.4793
#>  -6.6119 -1.4872
#>  -4.6379  9.0453
#> 
#> (1,1,3,.,.) = 
#>   8.8293  5.7999
#>   4.0675  0.0310
#>   3.1614 -2.0808
#>   7.2978 -4.6113
#>  -5.3526  2.6249
#> 
#> (2,1,3,.,.) = 
#>   -3.6297  13.0879
#>    0.4426   5.3592
#>   -3.0188  -3.5272
#>   -6.3864  -3.2856
#>    0.1971   4.6184
#> 
#> (3,1,3,.,.) = 
#>   2.3457 -4.5059
#>  -0.6642  2.9940
#>   3.5531 -1.6692
#>  -6.2980 -4.2022
#>   3.8599 -2.0988
#> 
#> (1,2,3,.,.) = 
#>   -7.7148  -0.9027
#>   -0.2000  12.1602
#>    6.0467  -1.3767
#>    3.5084 -11.1105
#>   -5.1683  11.5905
#> 
#> (2,2,3,.,.) = 
#>    1.5637   2.3064
#>    1.0590  12.7400
#>    0.9627   5.5048
#>   -4.8049  -7.6770
#>   -1.3564   2.9993
#> 
#> (3,2,3,.,.) = 
#>   0.0609 -1.4924
#>  -0.3709 -1.4160
#>  -3.0492  0.7079
#>   4.8611  4.9021
#>   3.3353 -10.5338
#> 
#> (1,3,3,.,.) = 
#>  -0.4755 -2.0368
#>   4.7065 -9.7528
#>  -2.7965  0.2146
#>  -1.2834  3.7712
#>   4.4591  6.8290
#> 
#> (2,3,3,.,.) = 
#>  -5.7518 -2.4945
#>   1.8783  3.8928
#>   4.0760  4.7860
#>   2.5946 -0.8541
#>  -1.8343 -5.6057
#> 
#> (3,3,3,.,.) = 
#>  -6.7338  3.6179
#>  -7.4093  2.7052
#>   0.7138  3.8406
#>   9.5513  2.5706
#>  -1.2045  1.0910
#> 
#> (1,1,4,.,.) = 
#>  -3.9263  7.0106
#>  -2.5195  0.4428
#>  -6.4888  6.5006
#>  -1.4293 -4.1111
#>  -1.6695 -6.3649
#> 
#> (2,1,4,.,.) = 
#>  -13.0799  -0.5649
#>   -0.3013   0.2854
#>    0.5151   3.1697
#>    6.3659   0.5913
#>    0.4552   5.0343
#> 
#> (3,1,4,.,.) = 
#>   1.9795 -0.2726
#>  -4.3545 -0.0908
#>  -3.4449  0.3937
#>  -3.4244 -1.4043
#>  -2.4324 -2.5763
#> 
#> (1,2,4,.,.) = 
#>    0.5603   0.6206
#>    5.6226  -2.7528
#>   -2.9207  10.1461
#>   -0.3291  -0.8305
#>    2.9910  -0.7708
#> 
#> (2,2,4,.,.) = 
#>  -1.2602 -2.4204
#>  -0.3227 -2.0235
#>  -3.2672  6.6582
#>  -8.3305  6.8149
#>   3.9845  5.6631
#> 
#> (3,2,4,.,.) = 
#>  -1.2129  9.8860
#>  -1.5130  0.3129
#>  -4.7812  4.0430
#>   0.3661 -4.0057
#>   5.6167 -4.6773
#> 
#> (1,3,4,.,.) = 
#>   0.7721 -3.8790
#>  -0.0949  2.0325
#>   1.8220  0.2131
#>  -6.4614  2.3202
#>  -1.1680  0.2643
#> 
#> (2,3,4,.,.) = 
#>   9.5614 -0.8108
#>   2.7266  1.6703
#>   0.2408 -3.3653
#>  -6.7654  8.0736
#>  -3.9306 -1.7141
#> 
#> (3,3,4,.,.) = 
#>   -3.5046  -0.5573
#>   -1.6020  -2.8473
#>   10.8374  -3.5995
#>    8.5485 -10.3913
#>   -2.6153   5.8670
#> 
#> (1,1,5,.,.) = 
#>  -2.5564 -5.2922
#>   6.6451 -8.2465
#>   0.5204 -2.3425
#>  -4.9790  8.5256
#>  -3.3916  9.6140
#> 
#> (2,1,5,.,.) = 
#>  -5.4102  5.2655
#>  -1.6151  3.1989
#>   0.0230  2.6894
#>   0.2610  3.8089
#>   7.2187  4.4831
#> 
#> (3,1,5,.,.) = 
#>  -2.1687  1.1452
#>   7.2405  0.0237
#>  -6.2894 -0.4475
#>   6.4199 -2.1653
#>  -0.8690 -0.6281
#> 
#> (1,2,5,.,.) = 
#>   2.3866 -9.5212
#>   1.0424 -0.3695
#>   4.0655 -6.4998
#>   1.7199 -1.8309
#>  -2.6886  1.4756
#> 
#> (2,2,5,.,.) = 
#>   -0.8965   3.9333
#>    2.8927   1.2791
#>    6.1394  -3.6822
#>   -2.0155   7.1541
#>   10.7753   1.6398
#> 
#> (3,2,5,.,.) = 
#>  -4.9257  3.2729
#>   1.1625  2.1176
#>  -2.9741  5.9234
#>   2.4396 -3.4627
#>   0.0933 -3.2734
#> 
#> (1,3,5,.,.) = 
#>    0.1165  -6.3275
#>   -0.6939   2.6154
#>   -2.4516  -4.5182
#>    3.1925  -9.0616
#>    1.5353  12.6257
#> 
#> (2,3,5,.,.) = 
#>  -2.7058 -0.5135
#>   0.3806 -0.7217
#>  -2.0976 -8.6706
#>  -3.5017 -3.3308
#>   0.7527 -5.8317
#> 
#> (3,3,5,.,.) = 
#>    3.2427   5.2477
#>   -1.4603   2.0772
#>    4.5818  -0.3293
#>   -3.7817  -1.5063
#>   11.3401  -4.1142
#> [ CPUFloatType{3,3,5,5,2} ]