Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -5.2333 -3.6009
#>  -5.5714 -3.7408
#>  -1.1763 -1.0917
#> 
#> (2,.,.) = 
#>  -2.8460 -4.9768
#>  -0.8823 -4.6249
#>   3.3081 -0.7081
#> 
#> (3,.,.) = 
#>  -3.4161 -0.7354
#>  -4.7006  7.6776
#>   2.4234 -2.3888
#> 
#> (4,.,.) = 
#>   1.7899  5.2235
#>   2.8744 -0.4023
#>   1.3837  3.5221
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.4264 -1.0224
#>  -2.0700 -0.2726
#>   1.4847 -0.1666
#> 
#> (2,.,.) = 
#>   2.0958 -1.8754
#>   0.8380 -3.7938
#>   0.1576  0.8054
#> 
#> (3,.,.) = 
#>  -1.8983 -1.1457
#>  -3.0661  2.2410
#>  -0.8612 -1.5736
#> 
#> (4,.,.) = 
#>  -3.0044  0.4426
#>  -1.2734 -1.9154
#>  -1.9575 -0.1568
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   0.2485  5.9488
#>  -3.2123 -7.3482
#>   4.6961  0.5858
#>   4.1773  4.2147
#>   4.9874 -0.1161
#> 
#> (2,1,1,.,.) = 
#>   3.6463  1.9528
#>   7.8382  4.3116
#>  -4.3367 -0.6916
#>   2.9359 -3.0441
#>  -3.5400  3.0353
#> 
#> (3,1,1,.,.) = 
#>   -1.9304  -4.9390
#>    3.7176 -12.9317
#>   -0.6820   0.4759
#>    1.5522   2.5316
#>   -8.7367  13.0195
#> 
#> (1,2,1,.,.) = 
#>   1.2543  4.7924
#>   3.9668 -6.0327
#>  -2.8527 -0.5639
#>  -8.7262 -1.4723
#>  -1.0480  3.0258
#> 
#> (2,2,1,.,.) = 
#>  -2.9556 -2.9154
#>  -14.3066 -11.6272
#>  -0.9077 -10.8507
#>  -4.4437 -1.5292
#>  -1.1868  6.0295
#> 
#> (3,2,1,.,.) = 
#>    0.7507   0.0911
#>   -1.3028  -2.5225
#>   -3.2273  -1.4866
#>   12.3143   3.8473
#>    0.6210  -3.4206
#> 
#> (1,3,1,.,.) = 
#>   3.6216 -5.5508
#>  -3.7148  2.9180
#>  -3.5760  4.0612
#>  -0.6228  2.6134
#>   2.4119 -0.9789
#> 
#> (2,3,1,.,.) = 
#>   12.2449  -8.7821
#>    4.9118  -4.4066
#>    2.1239  -1.3302
#>    2.1337  -4.8541
#>   -1.5924   2.5664
#> 
#> (3,3,1,.,.) = 
#>  -5.7616  5.7446
#>   2.3546 -4.9894
#>  -9.6968 -8.3546
#>  -10.6428  1.7647
#>  -4.8541 -5.2038
#> 
#> (1,1,2,.,.) = 
#>  -0.6531  7.9845
#>  -4.6305  6.0734
#>   3.9531 -6.4749
#>   0.0309 -3.9954
#>   2.7030 -1.1414
#> 
#> (2,1,2,.,.) = 
#>  -3.2056  3.2650
#>   2.7863  0.9476
#>   0.8640 -2.6237
#>  -0.5956  0.1087
#>   2.5648  1.8113
#> 
#> (3,1,2,.,.) = 
#>   1.6535 -14.0152
#>   8.5021  1.5519
#>   9.7971 -7.8058
#>  -5.4292  3.1063
#>   0.2689  3.6377
#> 
#> (1,2,2,.,.) = 
#>  -8.8848 -7.3138
#>  -7.0701 -0.1976
#>   1.0285 -0.1855
#>   4.1643 -0.3642
#>  -2.2864 -5.9893
#> 
#> (2,2,2,.,.) = 
#>   4.5299  0.8186
#>  -1.1864  9.7083
#>  -11.5627  0.5579
#>   9.0061  3.4969
#>   4.8006  6.7250
#> 
#> (3,2,2,.,.) = 
#>   7.1887  2.9220
#>  -3.6566  2.4779
#>   7.2338 -1.4421
#>  -4.6896 -2.0893
#>   3.6636 -6.5689
#> 
#> (1,3,2,.,.) = 
#>   11.3877   1.4004
#>   -4.2587   0.7115
#>    5.3725   3.4330
#>    6.5087   2.9944
#>    6.7183  -3.0521
#> 
#> (2,3,2,.,.) = 
#>  -0.1856  1.9796
#>  -3.9134  2.9545
#>  -5.0088 -1.7927
#>   5.1995 -3.0660
#>   2.9816 -0.0648
#> 
#> (3,3,2,.,.) = 
#>  -6.9332 -6.2839
#>  -3.6122 -6.0826
#>   6.5710  2.7982
#>  -1.8983  8.8959
#>  -1.3150  0.7277
#> 
#> (1,1,3,.,.) = 
#>  -2.7095 -6.9467
#>   0.1285 -4.6650
#>  -0.4486 -3.9085
#>  -1.4382 -3.8442
#>   8.7897 -6.9060
#> 
#> (2,1,3,.,.) = 
#>   0.6440  6.0348
#>  -4.2090 -10.1946
#>  -0.5351  3.9960
#>   0.6205 -2.3467
#>  -1.7668 -5.8971
#> 
#> (3,1,3,.,.) = 
#>  -4.1211 -7.0871
#>  -0.4402 -7.9849
#>  -3.6830 -10.1821
#>  -1.5947  0.2513
#>   6.9236  0.2942
#> 
#> (1,2,3,.,.) = 
#>  -0.2841 -4.1174
#>  -0.6260 -2.5351
#>  -4.5839 -8.7808
#>  -1.5922 -5.4802
#>  -5.5545 -8.6004
#> 
#> (2,2,3,.,.) = 
#>    0.6545  -1.7451
#>   -3.5546   1.2694
#>    7.3232  -1.3663
#>   -2.4954   4.2415
#>    1.3934  10.8607
#> 
#> (3,2,3,.,.) = 
#>  -2.2154  3.2453
#>  -14.2474 -2.6248
#>  -1.5140  4.5637
#>   4.1250 -7.4190
#>  -5.0089  4.4902
#> 
#> (1,3,3,.,.) = 
#>   4.6012 -1.5029
#>  -5.2104 -10.3353
#>  -0.5598  1.3152
#>   1.6266 -0.3922
#>   5.7207  1.5725
#> 
#> (2,3,3,.,.) = 
#>  -2.6444 -5.6187
#>  -6.6828 -0.9960
#>  -3.4098  3.9628
#>  -3.8969  1.8039
#>   4.7094 -1.1122
#> 
#> (3,3,3,.,.) = 
#>   3.4960 -2.8046
#>   2.2408 -0.3908
#>   4.1548  0.4749
#>   3.3944 -1.5793
#>  -1.6048  3.0698
#> 
#> (1,1,4,.,.) = 
#>    0.2024  -1.0550
#>    8.1195  15.1078
#>   -4.3066  -5.1663
#>   -8.1194   4.2311
#>    3.3433  -5.2980
#> 
#> (2,1,4,.,.) = 
#>  -5.3075 -0.0446
#>  -2.4435 -5.3125
#>   6.1731  1.9523
#>  -1.3673 -9.4030
#>   3.9001 -4.4316
#> 
#> (3,1,4,.,.) = 
#>   0.9581 -8.4363
#>  -0.2888  4.2156
#>   1.6871 -5.3925
#>   2.8593  7.4625
#>   9.6720 -6.9473
#> 
#> (1,2,4,.,.) = 
#>  -5.5789 -0.8192
#>   1.5291  2.4931
#>   0.2214  1.7610
#>  -4.0283 -7.3545
#>   2.7206  7.1098
#> 
#> (2,2,4,.,.) = 
#>   4.4353  2.2700
#>  -7.1696  9.8770
#>  -0.5385 -2.6729
#>   2.0161 -7.8035
#>  -7.9868 -3.4609
#> 
#> (3,2,4,.,.) = 
#>  -3.0709 -3.8251
#>   2.7085 -0.0781
#>   1.5660  6.4596
#>  -3.0003 -8.6518
#>  -6.6925 -11.0380
#> 
#> (1,3,4,.,.) = 
#>  -1.2006  1.7445
#>   9.1087  0.5871
#>  -4.9939  4.5738
#>   6.3450  4.8126
#>  -0.2189 -7.1458
#> 
#> (2,3,4,.,.) = 
#>  -6.5065  6.7388
#>  -7.3065 -3.0360
#>  -4.6855  5.2161
#>  -0.6740 -12.2688
#>   1.1456 -2.4377
#> 
#> (3,3,4,.,.) = 
#>   3.3685 -4.7376
#>  -9.5112  4.9808
#>  -3.7454  6.8830
#>  -2.1846  4.2319
#>   1.9536  0.4895
#> 
#> (1,1,5,.,.) = 
#>   4.7920 -1.3961
#>   6.8186  0.2859
#>   2.9569  3.2710
#>   5.7327 -1.8928
#>  -4.5133  4.8603
#> 
#> (2,1,5,.,.) = 
#>   5.2428  3.9592
#>  -7.5708  4.6216
#>   2.5250  3.2551
#>   4.1926  2.2933
#>   0.0862  7.4324
#> 
#> (3,1,5,.,.) = 
#>    7.8752  -2.4430
#>   11.2053  -4.0025
#>    6.6731  -1.1894
#>    7.1917   2.6196
#>   -5.9203  -6.9483
#> 
#> (1,2,5,.,.) = 
#>   2.2600 -4.1664
#>  -3.5751 -12.4943
#>   3.1958  0.2739
#>   2.3443 -6.4257
#>  -0.3557 -3.6495
#> 
#> (2,2,5,.,.) = 
#>  -4.7302  1.7129
#>   4.1706 -10.9478
#>  -5.4992 -3.6730
#>  -9.0347  2.7646
#>   9.0547  0.4922
#> 
#> (3,2,5,.,.) = 
#>  -5.6140  4.7499
#>  -6.2301  2.6522
#>   1.1663  2.2321
#>  -7.2997  2.5830
#>   0.3789 -3.2953
#> 
#> (1,3,5,.,.) = 
#>   3.6771 -3.7317
#>   8.6446 -7.3733
#>   2.2266 -1.5244
#>   0.2127 -0.9283
#>  -4.0233 -2.9439
#> 
#> (2,3,5,.,.) = 
#>  -4.0830  2.0631
#>  -0.7739 -3.5438
#>  -0.4220 -1.9701
#>   6.5258  2.6212
#>  -13.1351  3.1437
#> 
#> (3,3,5,.,.) = 
#>  -3.1240 -2.9757
#>  -4.5461  3.8076
#>   1.2038  3.8093
#>  -5.1265  1.3387
#>  -0.6127 -5.6232
#> [ CPUFloatType{3,3,5,5,2} ]