Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -5.6691 -1.8374
#>  -0.5448 -3.0953
#>   2.9018  2.2116
#> 
#> (2,.,.) = 
#>  -2.6575  3.6937
#>   3.8441  0.2339
#>   1.1809  0.8695
#> 
#> (3,.,.) = 
#>   0.9437  0.6862
#>   2.6827  0.0546
#>   1.5473  3.1162
#> 
#> (4,.,.) = 
#>   2.3558 -3.5719
#>  -3.8141  1.9452
#>  -0.1736 -2.5350
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.2568 -0.2573
#>   0.5420 -0.2154
#>   1.3641  0.9156
#> 
#> (2,.,.) = 
#>  -3.4696 -1.8842
#>  -0.3790  1.1271
#>  -0.5125  0.1125
#> 
#> (3,.,.) = 
#>  -1.1059 -0.3182
#>   0.5270 -1.3049
#>   0.8604  1.7483
#> 
#> (4,.,.) = 
#>   0.1632  0.6224
#>  -1.2347 -2.7020
#>   1.1898 -0.5648
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>  -2.6398  4.7514
#>  -2.1554  6.2221
#>  -3.8517 -4.5924
#>   3.9247 -8.3807
#>  -3.2791  8.4525
#> 
#> (2,1,1,.,.) = 
#>  -7.8411 -5.6610
#>  -5.9851  0.1843
#>  -2.4531  0.5268
#>   7.7623 -0.0013
#>  -3.2150  9.2896
#> 
#> (3,1,1,.,.) = 
#>   0.5399  2.7420
#>  -10.2028  3.7960
#>  -2.9287 -8.1338
#>  -1.8878  2.4174
#>   9.1908  1.6778
#> 
#> (1,2,1,.,.) = 
#>  -0.2288 -0.8387
#>  -3.1501 -3.8101
#>  -4.8982  3.3953
#>   0.3355  1.0201
#>  -4.5834 -3.4864
#> 
#> (2,2,1,.,.) = 
#>  -3.7946  1.9300
#>  -6.9187 -7.8888
#>   8.8969 -1.7611
#>   1.1444 -0.3800
#>  -1.0129 -5.8826
#> 
#> (3,2,1,.,.) = 
#>   1.3895  6.7307
#>  -0.8474 -1.0152
#>   4.9237  1.6299
#>   6.7448  5.1607
#>  -1.0570 -1.9520
#> 
#> (1,3,1,.,.) = 
#>   5.2671  9.3256
#>   5.9942 -7.2550
#>   2.3914  7.9033
#>  -1.2186  1.2807
#>   8.8157 -3.3310
#> 
#> (2,3,1,.,.) = 
#>  -5.6754 -1.9603
#>   3.0001 -0.5535
#>  -8.8893 -4.8138
#>  -6.4748  5.9657
#>   1.3358 -4.1068
#> 
#> (3,3,1,.,.) = 
#>  -5.9856 -9.8020
#>  -2.8465  4.3816
#>   2.9905 -0.6271
#>   1.2811 -4.7725
#>   2.9899 -2.2838
#> 
#> (1,1,2,.,.) = 
#>  -0.0533 -1.1502
#>  -4.0163 -1.3972
#>  -2.9828  8.7402
#>  -2.7409  2.0899
#>   1.1632  2.1589
#> 
#> (2,1,2,.,.) = 
#>  -1.1919 -0.4174
#>   1.4358  8.5473
#>   2.3924 -2.8852
#>   4.2842 -2.1418
#>  -7.2843  2.8963
#> 
#> (3,1,2,.,.) = 
#>   4.0490 -3.2905
#>  -4.7691  1.1593
#>   3.0600 -2.0113
#>   0.7577 -0.9925
#>   8.3977 -4.1543
#> 
#> (1,2,2,.,.) = 
#>   12.0720  -2.7246
#>   -1.1026  10.2704
#>   -6.2338   4.1128
#>    4.4242  -4.4658
#>    1.8935   6.7246
#> 
#> (2,2,2,.,.) = 
#>  -3.4260 -0.3590
#>  -3.8701  5.3765
#>   2.4266 -1.5879
#>   6.1980  0.5214
#>   9.7938 -1.5415
#> 
#> (3,2,2,.,.) = 
#>   0.8720 -3.0458
#>   0.8214  0.8766
#>  -1.1457  4.1521
#>   6.0425 -1.8147
#>   1.6198 -3.9862
#> 
#> (1,3,2,.,.) = 
#>   4.2660 -10.5792
#>  -2.2981  5.1796
#>  -0.6551  1.2528
#>   1.2466  3.5724
#>  -2.7106 -0.1261
#> 
#> (2,3,2,.,.) = 
#>   10.1056  -0.4052
#>   -4.3687  -1.1735
#>    5.2590  -7.5784
#>   -7.6221  -4.3367
#>   -7.6471  -0.0542
#> 
#> (3,3,2,.,.) = 
#>   3.1462 -1.4424
#>   2.9669 -6.5099
#>  -6.0497  1.6178
#>  -1.2674  6.1849
#>  -7.4746 -6.3688
#> 
#> (1,1,3,.,.) = 
#>   -5.6341  -6.5467
#>    3.4134  -1.0975
#>   -4.5952  -5.3578
#>   -2.8018  14.5290
#>    0.9895   0.9791
#> 
#> (2,1,3,.,.) = 
#>  -5.8557 -1.9146
#>  -9.4606  6.8006
#>  -5.3056 -10.5235
#>  -0.4007  6.3142
#>   0.8872  3.0344
#> 
#> (3,1,3,.,.) = 
#>   -5.5115  10.0565
#>    0.5681  -5.0608
#>    7.7277  -0.0864
#>    3.3659  -2.9287
#>   -3.0525  -7.8309
#> 
#> (1,2,3,.,.) = 
#>   4.0264 -8.7905
#>   4.9831  4.1200
#>  -5.6901 -10.8480
#>   2.2916 -1.6434
#>   3.1923 -1.0767
#> 
#> (2,2,3,.,.) = 
#>  -3.9926  7.4666
#>  -16.7225 -10.3794
#>   8.2417 -5.4958
#>  -3.8420  2.1621
#>  -2.5573 -1.2767
#> 
#> (3,2,3,.,.) = 
#>   -6.1573  -4.9033
#>   -0.4331   7.7767
#>   -3.9125   0.7704
#>   10.1236   2.7462
#>   -3.9985  10.5473
#> 
#> (1,3,3,.,.) = 
#>   1.9928 -3.8144
#>  -4.8671 -6.5292
#>   7.2624 -8.4548
#>  -4.9358  4.6421
#>  -0.0654 -0.7095
#> 
#> (2,3,3,.,.) = 
#>  -1.7342  5.0718
#>   4.2993 -5.6456
#>  -4.8840  7.2341
#>  -1.8598  4.0872
#>   1.3297 -6.6435
#> 
#> (3,3,3,.,.) = 
#>   3.0649 -4.0142
#>  -3.1263  7.8034
#>  -13.0920 -4.3586
#>   9.9992  5.5414
#>  -2.2651 -1.1982
#> 
#> (1,1,4,.,.) = 
#>  -3.8962  0.6932
#>  -0.0237  4.8750
#>   1.3426 -5.4600
#>   1.7428 -0.7907
#>   0.0012 -6.2903
#> 
#> (2,1,4,.,.) = 
#>    4.5410   1.5208
#>    1.3961   5.2610
#>   -3.3978  -0.2035
#>   -2.1849  10.6939
#>   -2.0375   6.0690
#> 
#> (3,1,4,.,.) = 
#>   3.4173 -0.2168
#>  -4.0016  1.0784
#>  -0.0798 -19.5018
#>   0.1108  0.5519
#>   0.1103  1.1036
#> 
#> (1,2,4,.,.) = 
#>  -2.0796 -1.5945
#>  -2.7532 -9.2946
#>   8.5643 -1.6354
#>   2.6042  1.1971
#>  -0.2598 -9.6974
#> 
#> (2,2,4,.,.) = 
#>   2.9462 -2.2414
#>   1.5732  3.1969
#>   1.1088  0.1659
#>   2.6682 -4.7840
#>  -5.7920  4.1998
#> 
#> (3,2,4,.,.) = 
#>   11.1560  -3.4099
#>   -2.8733   5.6990
#>   -3.0645  -1.5899
#>   -1.4316   6.7917
#>   -3.9889   2.4698
#> 
#> (1,3,4,.,.) = 
#>    5.0339  -4.9447
#>    7.7067   0.9768
#>    5.4781   2.6926
#>   -0.5646  14.9734
#>   -5.3697  -0.7555
#> 
#> (2,3,4,.,.) = 
#>  -7.5945 -2.2694
#>  -2.1156  0.6048
#>  -2.0442  0.1063
#>  -0.7984  7.1558
#>   2.9691 -4.8493
#> 
#> (3,3,4,.,.) = 
#>   3.5320 -0.0246
#>  -2.8542 -3.0934
#>  -4.0553 -0.7514
#>  -4.6472  3.2595
#>  -6.8948  0.7452
#> 
#> (1,1,5,.,.) = 
#>   5.4668 -1.0908
#>   4.5259 -2.3536
#>   1.1369 -6.9722
#>   3.4423  1.6317
#>  -0.0685  2.8230
#> 
#> (2,1,5,.,.) = 
#>    8.1855  10.5358
#>   -5.7481   2.9168
#>   -1.9791   1.6761
#>    1.6257   1.2123
#>   -5.1101  -5.8911
#> 
#> (3,1,5,.,.) = 
#>  -1.9214 -0.1168
#>  -2.4013  4.7145
#>   0.5533  3.2265
#>   2.4820  4.0360
#>  -1.8029 -1.4872
#> 
#> (1,2,5,.,.) = 
#>   6.9680  8.4110
#>  -5.0179 -7.8008
#>  -2.0408  2.6892
#>   0.5194  3.9942
#>  -2.2945 -3.5723
#> 
#> (2,2,5,.,.) = 
#>   8.5430 -10.5835
#>  -1.4563 -3.2299
#>   2.4964 -4.1464
#>   3.1373 -8.1710
#>  -5.7506  0.1640
#> 
#> (3,2,5,.,.) = 
#>  -7.7338 -2.3864
#>  -4.1985  3.4098
#>   2.6301  3.6340
#>  -3.9577 -0.9549
#>  -2.2785  4.3586
#> 
#> (1,3,5,.,.) = 
#>  -4.0638  6.4430
#>  -5.6271 -6.4166
#>   3.5672 -2.5122
#>   1.3131  9.5792
#>  -4.7260  4.2449
#> 
#> (2,3,5,.,.) = 
#>  -3.3971 -2.2329
#>  -3.8572  1.8251
#>   2.1139 -0.6331
#>  -14.9438  4.7205
#>  -0.4181  3.8747
#> 
#> (3,3,5,.,.) = 
#>   4.0785 -0.8731
#>   1.2321 -7.2763
#>   5.8291  9.2522
#>  -4.3013  1.0231
#>  -13.5977 -5.0465
#> [ CPUFloatType{3,3,5,5,2} ]