Fft

Fft

Arguments

input
(Tensor) the input tensor of at least signal_ndim + 1 dimensions

signal_ndim
(int) the number of dimensions in each signal. signal_ndim can only be 1, 2 or 3

normalized
(bool, optional) controls whether to return normalized results. Default: False

Note

For CUDA tensors, an LRU cache is used for cuFFT plans to speed up
repeatedly running FFT methods on tensors of same geometry with same
configuration. See cufft-plan-cache for more details on how to
monitor and control the cache.

fft(input, signal_ndim, normalized=False) -> Tensor

Complex-to-complex Discrete Fourier Transform

This method computes the complex-to-complex discrete Fourier transform. Ignoring the batch dimensions, it computes the following expression:

$$ X[\omega_1, \dots, \omega_d] = \sum_{n_1=0}^{N_1-1} \dots \sum_{n_d=0}^{N_d-1} x[n_1, \dots, n_d] e^{-j\ 2 \pi \sum_{i=0}^d \frac{\omega_i n_i}{N_i}}, $$ where \(d\) = signal_ndim is number of dimensions for the signal, and \(N_i\) is the size of signal dimension \(i\).

This method supports 1D, 2D and 3D complex-to-complex transforms, indicated by signal_ndim. input must be a tensor with last dimension of size 2, representing the real and imaginary components of complex numbers, and should have at least signal_ndim + 1 dimensions with optionally arbitrary number of leading batch dimensions. If normalized is set to True, this normalizes the result by dividing it with \(\sqrt{\prod_{i=1}^K N_i}\) so that the operator is unitary.

Returns the real and the imaginary parts together as one tensor of the same shape of input.

The inverse of this function is ~torch.ifft.

Warning

For CPU tensors, this method is currently only available with MKL. Use torch_backends.mkl.is_available to check if MKL is installed.

Examples

# unbatched 2D FFT
x = torch_randn(c(4, 3, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.9070  4.1614
#>   0.1752 -0.2454
#>  -3.4478 -3.2239
#> 
#> (2,.,.) = 
#>  -2.7429  0.0399
#>  -0.7835  2.7771
#>   1.9189  2.5494
#> 
#> (3,.,.) = 
#>  -4.3826  7.8156
#>   1.3106  1.9157
#>   1.6857  3.6816
#> 
#> (4,.,.) = 
#>  -0.0593  0.0712
#>   1.1442  4.9196
#>   0.4663  3.5909
#> [ CPUFloatType{4,3,2} ]

# batched 1D FFT
torch_fft(x, 1)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.2730  3.0220
#>   0.4616  2.3418
#>   0.1558  1.6495
#> 
#> (2,.,.) = 
#>   0.6267 -1.5845
#>   0.2518 -1.0222
#>  -1.0230 -1.3632
#> 
#> (3,.,.) = 
#>  -0.8718  2.9665
#>   0.2813 -1.5066
#>  -1.0368 -1.4207
#> 
#> (4,.,.) = 
#>   0.6111 -0.2427
#>  -0.8195 -0.0584
#>  -1.5438 -2.0895
#> [ CPUFloatType{4,3,2} ]

# arbitrary number of batch dimensions, 2D FFT
x = torch_randn(c(3, 3, 5, 5, 2))
torch_fft(x, 2)

#> torch_tensor 
#> (1,1,1,.,.) = 
#>   2.1576  4.7794
#>  -1.1779  5.8574
#>  -0.0151  3.3711
#>  -2.1155  1.7340
#>  -1.2804  0.8876
#> 
#> (2,1,1,.,.) = 
#>   1.8122 -3.7688
#>  -8.2168  0.5149
#>  -5.9110 -0.8432
#>  -2.4131 -1.5410
#>  -0.7176 -5.4171
#> 
#> (3,1,1,.,.) = 
#>  -2.1153 -3.1665
#>   7.2941  3.0313
#>   9.7073  1.6778
#>   1.5336  0.9033
#>  -1.6587  1.0901
#> 
#> (1,2,1,.,.) = 
#>   3.8306  0.4437
#>  -6.5783 -2.5805
#>  -4.1340  5.9988
#>   4.8870  2.5230
#>  -3.0316 -4.3071
#> 
#> (2,2,1,.,.) = 
#>  -3.2265 -3.2266
#>   7.5452  9.5176
#>   4.5777 -7.4549
#>   1.2228 -0.1868
#>  -3.2901  1.7710
#> 
#> (3,2,1,.,.) = 
#>  -9.6384 -4.9139
#>  -1.5145  3.8643
#>  -1.9577 -8.4815
#>  -9.8914 -5.5241
#>  -2.0145 -2.5521
#> 
#> (1,3,1,.,.) = 
#>   4.1456  3.0585
#>   7.1002  8.1889
#>   0.0245  0.4116
#>   4.7501  3.2673
#>  -3.7894  0.2630
#> 
#> (2,3,1,.,.) = 
#>   0.6299  4.9447
#>   3.6764 -4.6484
#>  -3.7512 -4.8911
#>  -3.0723 -1.0338
#>   1.0322  1.3167
#> 
#> (3,3,1,.,.) = 
#>   -2.0479  -7.0496
#>   12.7065   5.8181
#>   -3.8523   5.0714
#>    0.5518  -3.3425
#>    4.7991  -2.2446
#> 
#> (1,1,2,.,.) = 
#>   -0.0493  -8.0035
#>   -7.6417  -1.6268
#>    4.2159   9.3828
#>   10.5850   7.6916
#>    3.5166   5.4872
#> 
#> (2,1,2,.,.) = 
#>   0.7376 -8.2381
#>  -10.9257 -7.9943
#>   3.3768 -3.9503
#>  -4.1428 -0.9985
#>  -3.8495  5.6400
#> 
#> (3,1,2,.,.) = 
#>   7.2647  8.7173
#>   1.8824  8.1572
#>  -0.0147  0.1998
#>   2.4362 -5.3253
#>  -7.3914  3.4959
#> 
#> (1,2,2,.,.) = 
#>   3.2679 -4.6945
#>  -2.4089 -1.7914
#>  -3.3898 -2.0918
#>   5.8197  0.4970
#>   0.8817  3.0639
#> 
#> (2,2,2,.,.) = 
#>   3.1070 -3.9736
#>  -7.7929 -1.6350
#>  -7.1634  0.3135
#>   1.6711 -2.0834
#>  -1.9368 -1.6102
#> 
#> (3,2,2,.,.) = 
#>   7.9927 -0.4798
#>   3.2052  2.8879
#>   3.8767 -9.1410
#>  -2.6038 -0.3395
#>   2.6650  5.6962
#> 
#> (1,3,2,.,.) = 
#>   7.8536  1.3211
#>   1.3945 -6.5609
#>  -6.9743 -2.6743
#>  -5.6306 -0.8200
#>   3.3056  8.9819
#> 
#> (2,3,2,.,.) = 
#>  -6.2441 -2.3405
#>   5.2645  7.8093
#>   6.4549  0.5092
#>   5.0894  0.4852
#>  -10.1871 -4.1166
#> 
#> (3,3,2,.,.) = 
#>  -9.6720  6.1524
#>  -8.7133 -6.1879
#>   1.8861  1.1682
#>  -12.3008 -1.2690
#>   5.1660  2.2957
#> 
#> (1,1,3,.,.) = 
#>    1.5208  -1.3175
#>   10.7916  -2.5312
#>    1.3114 -22.0561
#>   -6.8953  -5.9718
#>    2.9576  -1.8125
#> 
#> (2,1,3,.,.) = 
#>   7.6987 -0.6828
#>  -1.3772  0.3261
#>   8.6760 -5.2052
#>   9.9316 -0.3739
#>  -5.8891  4.7569
#> 
#> (3,1,3,.,.) = 
#>  -1.8096 -2.2305
#>  -1.0341  2.8504
#>  -0.7261 -5.0715
#>   7.4348  6.0237
#>   1.3632 -1.7326
#> 
#> (1,2,3,.,.) = 
#>   1.7455  6.3041
#>  -4.5721  5.8352
#>   8.3114  4.8993
#>  -3.6635  8.1981
#>   8.0366  7.8813
#> 
#> (2,2,3,.,.) = 
#>  -4.8480 -3.8866
#>   2.3160  4.8777
#>  -2.8133  1.8966
#>  -1.9008  3.2476
#>   8.9584 -11.4205
#> 
#> (3,2,3,.,.) = 
#>  -3.2252  3.6167
#>  -2.1388  1.9692
#>  -4.4355  9.6845
#>  -4.1553 -2.7230
#>  -1.3846  5.2693
#> 
#> (1,3,3,.,.) = 
#>  -0.5533  2.4919
#>   1.3885 -0.8073
#>   9.0983 -5.3594
#>  -6.0854  4.4442
#>   2.0646 -2.2245
#> 
#> (2,3,3,.,.) = 
#>  -8.6243 -9.9892
#>  -7.6684  1.9651
#>  -2.0006 -5.5051
#>  -5.1959  5.6727
#>   2.2672  2.7789
#> 
#> (3,3,3,.,.) = 
#>   0.8697 -4.9801
#>   4.9055 -0.8627
#>   6.3664  8.5464
#>  -1.2886 -1.3151
#>  -3.3648 -8.7511
#> 
#> (1,1,4,.,.) = 
#>  -10.5834   6.1602
#>   -0.1637   7.0387
#>    3.3418   9.6348
#>   -7.8721   7.8822
#>   -2.8043  -4.6269
#> 
#> (2,1,4,.,.) = 
#>   2.8762  2.1376
#>   3.2955 -2.7017
#>  -0.6065 -1.3243
#>   3.5645  2.2015
#>   6.8862  3.1476
#> 
#> (3,1,4,.,.) = 
#>   3.5102 -5.6550
#>  -5.0718 -3.3268
#>  -4.4959  1.5655
#>   6.3578 -0.2292
#>   3.6787  0.1830
#> 
#> (1,2,4,.,.) = 
#>  -0.2590  2.9731
#>   3.8104 -3.7155
#>   5.1392 -4.0780
#>   4.4517  1.7822
#>   5.8143  2.6202
#> 
#> (2,2,4,.,.) = 
#>    6.1117  -7.4024
#>   13.4053   2.8771
#>   -1.4788   2.8509
#>   -6.7319   4.0374
#>    6.6056  17.1405
#> 
#> (3,2,4,.,.) = 
#>  -0.1758  3.9127
#>  -3.3991  0.7438
#>  -0.0283  4.1501
#>   1.0282 -0.5774
#>   3.1310  1.3773
#> 
#> (1,3,4,.,.) = 
#>  -1.4633  0.4106
#>   5.6460 -0.2511
#>  -7.7689 -2.5307
#>   3.7055 -1.5234
#>  -5.1823 -1.5440
#> 
#> (2,3,4,.,.) = 
#>   8.7564  0.5115
#>  -2.8790 -1.5254
#>  -5.3524  1.2590
#>   6.2532  3.7371
#>  -13.5229  2.8859
#> 
#> (3,3,4,.,.) = 
#>  -0.1829 -4.7760
#>   1.7876  3.5565
#>  -2.7219  7.4508
#>  -2.2076 -3.6213
#>  -2.1542 -10.6455
#> 
#> (1,1,5,.,.) = 
#>  -1.0404 -5.2980
#>  -4.6662 -1.9753
#>   4.4815 -3.0591
#>   0.1659  7.3341
#>   0.2488 -0.5060
#> 
#> (2,1,5,.,.) = 
#>  -5.7198 -1.0715
#>   3.3027 -2.3587
#>   4.7613 -0.7652
#>  -4.1624  6.2655
#>  -1.0220 -0.3202
#> 
#> (3,1,5,.,.) = 
#>   0.1775 -9.4738
#>  -5.8738  8.2268
#>  -9.1462 -1.5695
#>   8.3100  1.1618
#>  -7.5697 -3.5068
#> 
#> (1,2,5,.,.) = 
#>  -1.8718  0.5712
#>  -4.9281 -6.3041
#>   3.8844 -3.3348
#>  -5.7590  2.3827
#>   1.9047 -7.9823
#> 
#> (2,2,5,.,.) = 
#>  -2.3695 -1.1938
#>  -5.0046  7.1043
#>  -0.4051 -0.8098
#>   3.2345 -5.2550
#>  -5.0036 -5.6662
#> 
#> (3,2,5,.,.) = 
#>   2.1188 -2.6290
#>  -0.7651 -5.1114
#>  -8.8643 -4.5300
#>   1.8231  1.5307
#>  -4.6411  5.2229
#> 
#> (1,3,5,.,.) = 
#>   -5.8648  -4.6751
#>    3.1944   3.5723
#>    0.7663  -2.4429
#>   -2.1774  10.7206
#>   -6.5521  -5.6805
#> 
#> (2,3,5,.,.) = 
#>   0.8145  9.0113
#>   8.0480 -5.7945
#>   1.2326 -1.9474
#>  -2.1240  7.0861
#>  -3.6780  3.7573
#> 
#> (3,3,5,.,.) = 
#>  -8.9462  2.4628
#>  -4.9127  1.0667
#>   7.3705 -6.9832
#>  -5.5479  5.3205
#>   6.6308 -0.6810
#> [ CPUFloatType{3,3,5,5,2} ]