Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 1.08842
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.3184
#> -0.9883
#>  3.5438
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  3.6152  0.3527 -2.5642
#>  5.0562  0.1619  1.5707
#> -6.0898 -2.2536 -1.5541
#>  9.9867 -4.8027 -3.0000
#> -8.8326  3.4458  4.4981
#>  2.5339 -2.0904  0.2508
#>  0.6458  1.4523  5.4536
#>  0.8821  0.3423  7.1929
#>  2.8480  5.8044  0.9946
#>  2.1922 -4.4558 -1.3391
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   1.4433  1.7321 -5.4053 -2.1592  0.3680
#>   0.6219  0.1957  0.0587 -0.1896 -0.6341
#>   0.1157 -0.4628  1.0384  0.8038  1.3330
#> 
#> (2,.,.) = 
#>   2.5227 -2.0543  0.0219  0.7205 -1.2306
#>   0.4168 -1.4075 -0.5902  1.3461 -0.0209
#>   2.6911 -2.1659  2.3716 -0.3969 -2.0602
#> 
#> (3,.,.) = 
#>  -1.0949 -2.4596 -3.3889 -3.5299  2.1695
#>   0.4303  2.6655  0.1417 -1.1731  1.0178
#>   1.7174  3.5576  0.7937  4.4983 -2.5521
#> 
#> (4,.,.) = 
#>   2.4863 -0.8623  1.3934  3.0152  2.9410
#>  -5.9538  4.3738 -0.1687 -4.1594 -4.5356
#>   2.2974 -1.3971 -0.9635  1.4645  1.1266
#> 
#> (5,.,.) = 
#>   3.0169  0.8190  0.4195 -4.2733 -3.9465
#>  -0.6206 -0.7786  4.0164 -2.9875 -3.2511
#>  -0.4048 -0.6820  0.2439 -0.4397 -0.3692
#> 
#> (6,.,.) = 
#>   1.7445 -2.0133 -1.1678  3.1079  0.1657
#>  -1.1143  0.0519  0.7016 -0.9048 -0.2384
#>   1.1043  0.3462  0.3019  1.1057 -0.7927
#> 
#> (7,.,.) = 
#>   0.7661 -0.3360 -0.6872 -1.1792  0.1150
#>   0.6922  3.4364 -2.9642 -1.0587 -2.1100
#>   3.0835 -2.4866 -3.0667 -1.3604  2.7514
#> 
#> (8,.,.) = 
#>  -0.6178  2.2506  1.8841  2.0723 -1.3901
#>  -0.0918 -0.8108 -0.0876 -0.4034  2.5686
#>  -2.1332 -1.0852 -1.3373 -0.0644  0.4598
#> 
#> (9,.,.) = 
#>   0.8431 -0.7115 -0.0530 -1.8526 -1.0016
#>  -0.8406 -0.1724  1.8061 -1.7163 -3.0575
#>  -3.3852  0.2416 -0.5714 -0.0184  1.9914
#> 
#> (10,.,.) = 
#>  -0.6083  1.7186 -0.6554  1.8584  1.3510
#>  -2.4859  3.7812 -0.4801 -2.0232  0.3564
#>  -1.0965  1.8055 -0.5213  3.5172  1.7734
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.6487 -1.4387 -2.4518 -1.3052 -2.9767
#>  -0.0629  0.2708  2.4621 -1.7849 -0.5192
#>  -2.0478 -0.5034 -1.8465 -2.2596 -0.4826
#> 
#> (2,.,.) = 
#>   1.4196 -0.6288  3.9691 -3.0780 -2.6359
#>  -0.4557  0.8444 -0.0818  0.6338  3.1763
#>  -0.3662  0.4365  1.4936 -1.0161  0.3246
#> 
#> (3,.,.) = 
#>   2.8547 -0.1060  4.0940 -0.1179 -1.2840
#>  -1.4719  0.8413 -0.9135  0.7308  1.5090
#>   1.1487 -0.4273  0.7517 -0.2701 -0.3500
#> 
#> (4,.,.) = 
#>   2.5605 -0.3309  3.9806 -0.7140 -2.2081
#>   0.5292 -1.3793 -1.2622 -0.3457 -3.8844
#>   1.3015  0.2754  1.6564  0.2836  1.5119
#> 
#> (5,.,.) = 
#>  -2.5831  0.4967 -2.3522  0.2419  0.0933
#>  -1.0223  0.6403 -0.5734  1.4696 -1.0765
#>   1.2865  0.4111  4.6089 -1.1717 -1.9566
#> 
#> (6,.,.) = 
#>   1.9581 -0.3917  1.2661  0.5849 -0.6734
#>  -2.7924  0.2780 -4.2534  0.5348  2.1664
#>   0.7017 -0.8647 -0.8597  0.4090 -2.4589
#> 
#> (7,.,.) = 
#>   0.2941  0.1213  1.8838 -0.8240 -1.1247
#>   0.6212 -0.0402  0.3500  0.1743  0.5660
#>   1.9979  0.1886  2.3622  0.8589  0.0604
#> 
#> (8,.,.) = 
#>  -0.8254  0.4913  1.3677 -1.4624  0.5111
#>  -0.8439  0.3921 -0.9820  0.3531  1.6087
#>  -1.3474  0.9299  3.2713 -2.6460 -0.4057
#> 
#> (9,.,.) = 
#>   3.5851 -1.8490  1.9296 -0.1223 -5.4952
#>   0.9599 -0.0710  1.6307 -0.2846 -0.7737
#>   0.2721  1.6978  3.4028  1.1835  1.2378
#> 
#> (10,.,.) = 
#>  -0.7557 -0.0979  0.4091 -0.7022 -2.8106
#>   1.1422  0.6919  2.3702  0.4531  1.6270
#>  -1.1847  0.1346 -2.2479  1.1503 -0.0321
#> [ CPUFloatType{10,3,5} ]