Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.50749
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -2.2513
#>  0.5098
#>  0.1148
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  1.0286 -0.4028  0.0924
#>  0.0414 -0.3837 -1.4385
#>  0.9574  0.8048  0.0295
#>  1.8138 -1.1227 -0.0402
#>  3.1904  0.1680 -0.0726
#> -0.2193 -0.5651  0.0945
#> -0.3475  1.2953  1.2206
#> -0.8399 -0.3748 -0.4062
#> -1.3206 -0.9181 -0.0169
#>  0.1667 -0.7254 -2.6082
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.6890  3.1857 -1.4676 -0.6015 -2.5448
#>  -0.8249 -1.3450  0.9520  0.5075  1.0533
#>  -0.7866 -0.2661  0.3585  1.7718 -3.3150
#> 
#> (2,.,.) = 
#>   1.1996  0.5810 -2.2220 -0.0843  0.3338
#>   0.0830  0.1344 -0.6991 -0.0468 -0.7524
#>  -1.5329 -0.8460  0.0440  0.0854 -0.7971
#> 
#> (3,.,.) = 
#>   2.9022 -0.9911  1.2175  2.6550  3.8427
#>   6.3013 -2.1741 -2.9205 -2.2880 -1.4644
#>   1.9610 -1.4235  0.5403 -3.9419  2.6712
#> 
#> (4,.,.) = 
#>  -4.3915  3.0565  1.3077  3.6165 -1.6016
#>   0.2767 -1.1290 -0.6393  0.4823 -2.2246
#>   3.9748 -0.9367  2.4328 -2.2618  0.2641
#> 
#> (5,.,.) = 
#>   4.1878 -2.5544  0.5455 -0.0983 -0.4372
#>  -0.0344 -1.9874 -1.1618  0.0014 -1.5318
#>   2.9588 -0.4974  0.0389  1.8713 -0.7472
#> 
#> (6,.,.) = 
#>   1.2090 -1.8599 -0.7590  2.7833  0.1866
#>   2.2150 -0.5538  0.0411  0.8569 -0.0081
#>  -0.5589  0.8568 -0.3304 -1.8755 -0.3196
#> 
#> (7,.,.) = 
#>  -0.1067 -0.5220 -0.6230  0.8438 -0.2801
#>  -1.0319  1.4681 -1.6079  1.9506 -2.6697
#>   0.4412 -0.5091 -1.5899  2.6843 -1.1000
#> 
#> (8,.,.) = 
#>  -0.0031 -0.4256  1.5254 -2.4231 -1.6136
#>  -0.6094 -0.0177 -0.2351  0.5108 -0.1122
#>   0.9840 -0.4675 -0.4258  0.9908  0.6850
#> 
#> (9,.,.) = 
#>   0.3497  1.3102  1.0016 -1.1315  1.7131
#>  -1.6865  0.8903 -2.7759 -1.0032 -5.0079
#>  -1.5851  2.4589 -1.0574  0.4422 -2.0452
#> 
#> (10,.,.) = 
#>  -2.3431  1.7447 -3.2192 -2.7166  1.1542
#>   0.5949  0.9223  1.7355  0.7507  2.7887
#>  -0.4395  1.0786 -0.7277 -0.7243  3.3317
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.8010 -1.8328 -4.1684 -3.7757 -4.6271
#>   2.7331  2.4914 -4.8130  1.4013 -0.3222
#>  -3.9964 -0.1736  3.6125 -4.7730 -5.7277
#> 
#> (2,.,.) = 
#>   1.6165  1.2490  0.1760  2.6998  2.3725
#>   3.1566  3.7868 -3.1579  2.8777  0.3714
#>  -1.5354  0.3048 -3.4741 -4.7966 -5.9526
#> 
#> (3,.,.) = 
#>   0.0336  0.2713  1.9594  1.5186  3.8630
#>   1.4358  3.6110 -4.7211 -1.1829 -5.4787
#>  -1.6821 -3.9148  4.6126  0.5836  3.9147
#> 
#> (4,.,.) = 
#>   0.3789 -1.1199  4.4921  3.6128  5.6109
#>  -2.2689 -0.9232  4.3784 -1.1684 -0.7807
#>   0.7187 -0.4030 -1.4203  0.4804  0.9334
#> 
#> (5,.,.) = 
#>  -0.4401  3.4832 -1.7573 -2.6207 -6.9317
#>   1.9388  0.5142  3.0573  5.2846  7.1489
#>  -0.1258 -0.1792  5.3948  2.9439  3.0237
#> 
#> (6,.,.) = 
#>   1.6663  1.1870 -3.4700  0.4496 -1.3928
#>   2.9069  4.0666 -0.2282  4.2773  2.7181
#>  -0.0780  5.0012 -0.5542 -1.3342 -5.1184
#> 
#> (7,.,.) = 
#>  -3.2665  0.8438  3.5817 -3.4375 -3.0979
#>   0.5360 -0.5040  1.1249  1.4869  0.7754
#>   0.7431 -6.3413 -1.4957  1.4688  6.0393
#> 
#> (8,.,.) = 
#>   3.2860  4.4188 -6.7460  0.5394 -4.9253
#>   2.2806 -2.7680 -1.0362  3.7108  5.6480
#>  -1.3221 -1.1074  1.1612 -1.4213 -0.9945
#> 
#> (9,.,.) = 
#>   0.7802 -2.2014 -0.3908  1.4049  2.5757
#>  -1.2875  2.7655  0.2433 -2.4739 -4.2635
#>   1.5058 -2.5667 -5.5210 -0.1677  2.0873
#> 
#> (10,.,.) = 
#>  -0.2165 -2.1312 -0.3290  0.0659  3.3712
#>   2.1005  4.2103 -1.5485  2.0086 -0.4557
#>  -0.6516 -0.8008 -1.8464 -2.0909 -1.8979
#> [ CPUFloatType{10,3,5} ]