Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.232442
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.2237
#> -2.9953
#> -0.8358
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.6835  0.4267 -1.8562
#>  0.8844 -3.7448  1.9345
#> -1.6713 -2.0017 -0.7220
#> -1.2376  4.1470 -1.0089
#>  1.2994 -1.6806 -3.4683
#>  0.9436  1.0673 -2.8371
#>  0.3857 -2.7765  0.9283
#> -1.5557 -4.1785  2.2004
#> -1.1715 -2.4012  1.4064
#> -2.5698 -3.9059  0.8745
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.5208 -1.0210  1.4440 -1.1323 -1.5809
#>  -1.9276  0.0470  2.0304  0.2850 -0.2787
#>   0.9006 -3.2111  2.8771  0.2986  4.3910
#> 
#> (2,.,.) = 
#>  -1.1961  0.1640  3.1611 -2.2739 -0.5912
#>  -0.5623 -1.2097  1.8852 -0.8949  0.6509
#>   0.3563 -1.1920 -1.6917  1.6592  1.0230
#> 
#> (3,.,.) = 
#>  -1.2853  1.2756  2.1208 -0.3184  0.6107
#>   1.2433  2.7608  1.4293 -2.0099  2.4489
#>   1.0337  2.4018  1.2494 -2.4562  2.7469
#> 
#> (4,.,.) = 
#>   0.9416  0.3872 -2.2158  0.2433 -0.5509
#>  -0.4397  1.3644 -1.3951  0.3417 -1.8035
#>   1.1337  1.8931 -4.8248  0.2355 -1.9738
#> 
#> (5,.,.) = 
#>   1.4881  0.1827  1.8847 -0.8898  0.9136
#>   0.9291  0.1223  0.3533 -0.4422  0.6491
#>   1.6929 -0.6815 -0.0299 -1.0432  1.4475
#> 
#> (6,.,.) = 
#>  -1.1351  0.2071  0.0880  0.5002  2.8513
#>  -0.3019  0.6562  0.4778 -0.4224 -0.2899
#>  -0.1076 -0.2914  0.0220 -0.3119  0.8185
#> 
#> (7,.,.) = 
#>  -2.4225  0.7948  0.0107 -0.6351 -4.0140
#>   2.0854  0.4524 -2.5536  0.3603 -0.1928
#>  -0.5203  0.3311  1.7098 -0.4098  0.3872
#> 
#> (8,.,.) = 
#>  -1.0754 -0.2828  0.0684 -0.8131 -0.2559
#>  -0.8701  3.6946  1.8604 -0.6374 -4.2431
#>   0.3108  0.9496 -1.5124  1.2651 -0.6497
#> 
#> (9,.,.) = 
#>   1.5130 -3.1901  0.9035 -1.8355  0.4123
#>   0.8323  6.2052 -0.5009  2.6429 -2.1082
#>   0.4077  6.1197  0.6296  4.1313 -4.4506
#> 
#> (10,.,.) = 
#>  -1.1674  0.6608  0.3663 -0.6180 -1.3858
#>   2.7294 -0.6450 -1.0543  1.2703  2.2824
#>   0.2300  0.8249  0.4252 -0.3121  2.0438
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.2465 -0.7317 -6.1592  0.1165  3.0408
#>   1.7198  0.9098  4.3989  2.3645 -4.6224
#>   0.9263 -0.4436 -0.9452  0.3285 -0.2435
#> 
#> (2,.,.) = 
#>   0.7029  1.1310  1.4717  2.1434 -2.5816
#>  -1.1222 -1.0195 -1.4106 -1.8651  2.6364
#>   1.0800  0.2634 -0.9865 -0.1326 -0.2230
#> 
#> (3,.,.) = 
#>  -1.7586 -0.9603 -2.1332 -1.2772  3.0194
#>  -2.1240 -1.7551  0.2441 -1.6303  2.5581
#>   0.5193  1.2736  4.3125 -0.7928 -2.1202
#> 
#> (4,.,.) = 
#>   0.6744  1.2291 -1.1318 -0.0243 -0.1829
#>   3.2921 -0.1010 -3.4756  2.3762 -1.8344
#>  -1.1218 -0.8880 -1.6677 -3.5217  3.6596
#> 
#> (5,.,.) = 
#>  -0.2542 -0.2702  4.6503  0.2976 -2.0471
#>   0.0031  0.2743  1.9030  2.4010 -2.2834
#>   0.9274  1.0157 -0.4147  0.3958 -0.8559
#> 
#> (6,.,.) = 
#>   1.9486  1.4145  1.1684  1.6473 -3.0170
#>  -0.9599  0.0842 -3.7328 -0.4574  2.5338
#>  -1.6722  0.6554  3.1510  0.6834 -0.9023
#> 
#> (7,.,.) = 
#>  -2.2783 -0.5337  1.0942  0.9014  0.5594
#>   0.0795 -0.6311  4.0856  1.5403 -2.6237
#>   0.4374  0.8439  0.2168  1.6235 -1.4866
#> 
#> (8,.,.) = 
#>  -0.6924  0.9558  3.5336  5.4383 -4.4547
#>   0.3545  2.3607  3.2474  0.0880 -2.2818
#>   1.7587 -0.5408 -5.5995 -1.7531  2.5331
#> 
#> (9,.,.) = 
#>  -1.5376 -0.4598  0.2475 -1.9827  2.0884
#>  -3.1436 -1.4960  5.6240 -2.4731  1.1899
#>   1.3906 -1.2609 -0.5799 -3.8071  1.8249
#> 
#> (10,.,.) = 
#>   3.8131  1.4291 -1.0293  0.1365 -2.3523
#>   0.9537 -1.0316 -0.2314 -1.1112  0.3659
#>  -1.5475 -1.2844 -0.2574 -1.2936  2.1216
#> [ CPUFloatType{10,3,5} ]