Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 1.65408
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.7412
#> -0.4754
#>  2.1440
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  2.4078 -1.4631  1.6778
#>  0.9080 -0.8773  0.4081
#> -0.6084  0.0151 -0.6859
#>  0.2938  0.1423 -2.1831
#> -0.6499 -0.2103  0.4219
#> -0.9469 -0.0371 -0.2475
#> -0.2640 -1.1722  0.3363
#> -0.0799 -0.9441 -1.8749
#> -2.5565 -1.1661 -1.1119
#> -0.5346  0.0330 -1.9139
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   4.2148  0.9165 -3.3858  0.0374 -7.5605
#>   0.2765  0.5747  0.9320 -0.8669 -0.9248
#>   0.1396  0.3395  1.9167 -0.6386 -3.8885
#> 
#> (2,.,.) = 
#>  -1.2896  0.5766 -2.2115  0.3911 -1.2323
#>  -2.0264  4.4705 -1.9177  0.8919 -4.6889
#>  -1.8784  2.4743 -2.3512  1.7021 -2.2544
#> 
#> (3,.,.) = 
#>  -3.0732 -2.5451 -0.1846  3.6274  1.5504
#>  -1.5853  0.6540  0.7031  1.8094  1.6766
#>  -2.2542 -3.5808 -0.2393  1.9513  1.6656
#> 
#> (4,.,.) = 
#>  -1.5460 -0.5264  0.7458  0.0404 -0.1476
#>   2.0932 -2.2042 -0.8800 -2.9558  0.6045
#>   1.5685  2.1077 -0.9794  0.6160 -0.1464
#> 
#> (5,.,.) = 
#>  -0.0743  0.5233  0.6395 -0.7793  1.2141
#>  -0.2005  1.6640 -2.4981  2.4730 -2.7383
#>  -2.3412 -5.5923  2.5747 -1.5687  0.5222
#> 
#> (6,.,.) = 
#>   1.3014  1.6833 -0.4022 -0.4962  1.7381
#>  -0.9334  2.1251  2.2349  1.2584 -0.7964
#>  -0.2754  3.1264  0.1234  2.0647 -1.9639
#> 
#> (7,.,.) = 
#>  -0.1511  0.6919 -4.1180 -5.9451  1.0199
#>   1.5932 -0.9434  2.0457  4.1119 -3.0513
#>  -0.8204 -0.3315  0.9010  2.2344 -1.0457
#> 
#> (8,.,.) = 
#>   0.1677 -1.7644 -0.5936 -0.8907 -0.8310
#>  -3.1382 -4.8261  1.2292 -4.2141 -5.3266
#>  -2.6910 -8.5136  1.4680 -6.6501 -6.5235
#> 
#> (9,.,.) = 
#>   1.5157 -0.8286 -2.9910 -3.1857  1.3384
#>  -1.5302 -6.0779 -3.2748  0.8932  1.4766
#>  -0.4905 -4.5402 -0.6314  3.0450  0.9941
#> 
#> (10,.,.) = 
#>  -4.1568 -0.4221  1.7947 -3.2934 -1.6669
#>  -1.4859  0.8024 -3.5180  0.0355 -1.0195
#>  -0.7089 -0.6855 -0.6825 -0.3562 -0.7248
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.1327 -0.2009  0.2706  0.4552 -0.4736
#>  -4.6876 -1.0628  1.2062 -0.6572 -0.5606
#>   0.1968  0.2367 -1.3026 -0.2134 -0.8629
#> 
#> (2,.,.) = 
#>   0.3101 -0.7509  3.2206  0.7370  1.9112
#>   1.1869  0.3313 -1.2433  0.4703 -1.2352
#>   5.9353  0.4819 -0.8408  1.5604 -0.4942
#> 
#> (3,.,.) = 
#>  -1.1558 -0.0191 -1.2704  0.0760 -1.9276
#>   0.9610 -0.6747  0.7893  0.6253 -0.4248
#>   9.2090  0.8397 -1.4454  1.7416  0.0449
#> 
#> (4,.,.) = 
#>  -0.9243  0.2729 -0.6711 -0.4823 -0.1060
#>  -1.5246 -0.0026 -1.1414 -0.7803 -0.7921
#>  -1.8639 -0.1432  0.1823 -0.1889 -0.3085
#> 
#> (5,.,.) = 
#>  -2.5947 -0.6732  0.6015 -0.3122 -0.5558
#>  -2.8733 -0.9596  2.0892  0.0298  0.2688
#>   5.2931  0.7795 -1.0993  1.4319 -0.4587
#> 
#> (6,.,.) = 
#>   1.0069  0.8488 -2.1797  0.3291 -1.5420
#>   3.7164 -0.5693  2.1807  0.8099  1.9146
#>   4.4279 -0.2952  1.7053  1.0024  1.6882
#> 
#> (7,.,.) = 
#>  -4.6175  0.1304 -1.8071 -1.2308 -1.7778
#>   9.6220  0.6905 -0.0929  1.4729  1.8716
#>  -6.1160 -0.3459 -0.7165 -1.3116 -1.4848
#> 
#> (8,.,.) = 
#>   1.0055  1.1592 -2.8022 -0.4826 -0.8431
#>  -7.9203 -0.8978  2.1137 -1.0496  0.1709
#>  -1.4824  1.6734 -6.3267 -2.2306 -2.7913
#> 
#> (9,.,.) = 
#>   5.4531 -0.6142  3.6566  1.9317  2.6637
#>   1.8219 -0.4826  3.0864  0.8463  2.4206
#>   0.5669  0.9236 -2.6608 -0.1770 -1.4952
#> 
#> (10,.,.) = 
#>  -4.9802  0.1032 -1.3982 -1.8872 -0.6279
#>   2.9438 -0.3908  2.8522  0.9575  2.4745
#>   3.3101 -0.3615  2.2994  1.0941  1.8211
#> [ CPUFloatType{10,3,5} ]