Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 0.880765
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -7.4092
#> -2.5797
#> -1.8387
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.9300  2.3464 -1.9809
#>  0.3231 -0.1355 -0.6076
#> -0.8288  0.2507  0.2619
#> -1.3245 -1.6832  1.8771
#> -1.1285  0.5819  0.4634
#>  0.0193  4.0313 -0.2679
#> -0.6611  0.5647 -2.8146
#>  5.1640 -2.3105 -0.0244
#>  1.6952  0.3951  2.0038
#>  1.3336 -1.1271 -2.1812
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.1757  1.4763 -1.0495 -1.5309  2.2183
#>   1.7919 -3.0040  2.5996 -2.9468 -1.7465
#>  -1.2431  0.2431 -1.6952  1.1487  1.8336
#> 
#> (2,.,.) = 
#>  -3.6283  3.1082  0.0765  1.3054 -0.9046
#>   1.0282  0.0547 -0.7301 -0.0303  0.2010
#>   1.4399 -0.5060  0.4642 -1.1334  2.0396
#> 
#> (3,.,.) = 
#>   1.5343  4.1856  3.2244  0.0107 -0.6631
#>   0.6125 -0.2906  3.6732 -0.2242 -0.2325
#>  -0.1886 -3.2406  2.7016 -0.5825  1.8732
#> 
#> (4,.,.) = 
#>  -0.9215 -0.2297 -0.1561 -0.2672 -0.5145
#>   2.4735  1.9600  2.9534 -0.4771 -2.5971
#>  -0.1502  2.9300  2.2574  1.1226 -2.5359
#> 
#> (5,.,.) = 
#>  -1.1106 -0.1002 -0.1361 -2.1186  0.0906
#>  -1.8519 -0.1516  0.7286  1.0840  3.1020
#>  -0.6966  0.5780  0.7682 -1.5865  2.1596
#> 
#> (6,.,.) = 
#>  -0.7161  0.2356  0.2205 -0.9249  1.3295
#>   2.1876 -3.1749  0.3156 -0.3486 -3.3328
#>  -2.5457 -3.2595  1.1906 -6.1339  2.0814
#> 
#> (7,.,.) = 
#>   0.0108 -1.4819  0.9815  0.8182  1.0623
#>   1.3853 -3.4932  1.9267 -1.8766  0.4285
#>  -0.9835  0.1829  1.8418  0.4070  3.1734
#> 
#> (8,.,.) = 
#>  -0.3665 -0.9531  1.8861 -2.0899  1.5686
#>  -4.4965 -1.5176  0.0654  2.4530  0.0353
#>   3.3707  0.6998  2.5693  1.0684  2.9161
#> 
#> (9,.,.) = 
#>  -3.5502  3.2697 -1.3678  3.1811  2.3358
#>   3.0576 -0.6718 -0.8753  0.8674 -1.5285
#>  -2.4606  2.3320 -2.6664  2.5697  2.5373
#> 
#> (10,.,.) = 
#>   0.1583 -3.9403  1.4469 -1.2367  5.2625
#>   1.8211  4.4170 -3.1008 -2.0418 -3.8713
#>   0.7075  6.1567 -1.2387 -3.8671 -4.9850
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.0488 -2.3002 -0.5586 -1.6584 -0.9212
#>   1.6702  0.2012 -0.8558  0.1456  0.4148
#>  -2.3588 -0.7589  0.2186 -0.0711  0.1977
#> 
#> (2,.,.) = 
#>   0.1105 -1.1317  0.1727 -1.3223 -0.8435
#>  -4.0212  3.1223  2.4841  1.6767  1.6112
#>   0.2188 -1.8396 -1.5112 -1.4808  0.8412
#> 
#> (3,.,.) = 
#>  -0.3938 -2.0825  0.4602 -0.8978 -2.5385
#>  -1.5972  0.2055  0.2269  2.4850 -0.9232
#>  -1.5984  3.3333  1.8897  1.4223  1.4652
#> 
#> (4,.,.) = 
#>   2.7654 -1.9825 -2.7939 -2.5823  1.8314
#>  -1.9887  1.5992  1.1316  2.8944 -0.5212
#>  -1.1643 -1.7989 -1.0259 -0.2833  0.1306
#> 
#> (5,.,.) = 
#>   0.5413  1.8860  0.0564  1.3163  1.0859
#>   0.6768 -4.8377 -1.7803 -2.6726 -2.1102
#>   0.7965 -3.0951 -0.9124 -2.3461 -1.4503
#> 
#> (6,.,.) = 
#>  -1.4388  0.7798  0.6267  0.0699  0.9256
#>   1.5862 -2.6658 -1.9796 -1.8087 -0.1502
#>   1.5646  0.9965  0.3028 -0.2341  0.0465
#> 
#> (7,.,.) = 
#>  -2.4573  8.7204  5.7257  6.3762 -0.4216
#>   1.9215 -0.1003  0.2286 -0.8151 -1.0810
#>  -1.7681 -1.4820 -0.2275  0.2916 -0.7183
#> 
#> (8,.,.) = 
#>  -0.5798  1.2619  0.7368 -0.8265  1.5232
#>  -3.9959  4.4514  1.1483  4.9440  2.9565
#>   0.7919 -1.7026 -1.8104 -0.3159  0.2511
#> 
#> (9,.,.) = 
#>   0.6590 -0.3035 -0.2272 -0.1474 -0.3583
#>   3.2049 -1.8969 -0.9648 -0.8311 -2.4280
#>  -2.1035 -0.6607 -0.9978  0.7173  1.4364
#> 
#> (10,.,.) = 
#>   0.4784  0.9940  0.6655  2.0601 -1.5374
#>  -2.3666 -0.6816  1.2327  0.2078 -1.4510
#>   4.2429 -4.0675 -2.3618 -5.1726 -0.5817
#> [ CPUFloatType{10,3,5} ]