Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 0.22311
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.0440
#> -0.8390
#>  1.2847
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.8813  0.2298  1.8640
#> -0.4326  1.4312 -0.6332
#>  2.0178 -1.0015  0.8866
#> -1.6055 -2.1697  0.6260
#> -0.0361 -0.1602  1.0186
#>  1.4284  1.6517 -1.0083
#>  0.1500 -0.3586  0.8782
#> -0.5429  0.7422 -2.2328
#>  1.6992  0.7365 -1.1526
#>  0.1443  0.0558  0.2637
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   2.1820  0.8707  1.1459  1.7157 -0.5177
#>  -4.3039 -0.5954 -1.4933  0.0809 -0.1863
#>   0.4035 -0.1013  1.4801  1.2497  0.2445
#> 
#> (2,.,.) = 
#>   0.3473  1.7446 -0.0039 -2.9293  1.9113
#>  -1.6784  0.2257  4.9303 -1.8283  2.4174
#>  -1.6877 -1.5180  0.6054  2.1935 -0.4534
#> 
#> (3,.,.) = 
#>   0.8347 -6.2734  0.2001 -0.7657  1.7708
#>  -0.7236  4.9853 -1.3482  0.1136 -1.0608
#>  -0.8549  0.6387 -5.5486 -0.2949 -2.9051
#> 
#> (4,.,.) = 
#>  -2.4468 -1.6119 -1.6398 -2.4124 -0.3956
#>  -2.9289  0.6097  0.4420  0.2850  0.4777
#>   0.6472  0.1565 -0.3492  3.5526  0.1408
#> 
#> (5,.,.) = 
#>   1.5793 -1.2658  0.5779 -2.0517  0.1442
#>   2.7837 -0.6966 -0.1814 -2.0660 -0.3937
#>   1.3862 -0.8706  2.9352 -2.6612  1.6791
#> 
#> (6,.,.) = 
#>   0.3021 -2.5631  0.1351  1.5730  0.2564
#>   5.4143 -0.2371 -0.2215  2.5249  2.1169
#>   4.1984 -1.8136 -1.6667  1.2807  1.1211
#> 
#> (7,.,.) = 
#>  -0.7146 -0.2806 -0.4436 -0.2739  1.5857
#>  -1.1191  3.7727  0.9833 -5.3116 -1.9434
#>   0.9988 -3.0761 -1.8449  4.2335 -2.3168
#> 
#> (8,.,.) = 
#>  -0.6493  0.1418  4.2923  0.6988 -2.0286
#>   3.9196  0.9027 -1.0427 -0.7675  0.2015
#>   0.1430  0.6931  0.9398 -2.1359 -0.0474
#> 
#> (9,.,.) = 
#>  -1.1877 -1.3411  1.1199  1.5149  3.5105
#>   0.9798 -1.3947  3.6796  1.9072 -4.9262
#>   2.3068  3.2709  0.3444  2.7156 -3.9753
#> 
#> (10,.,.) = 
#>  -1.3090  0.7529 -1.0970 -0.2405  0.6831
#>  -0.6798 -2.1347 -0.7649  3.0702  1.3522
#>  -2.0239 -3.2276 -2.4083  5.4901  2.9768
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -1.1760  0.6225 -0.4287 -0.8911 -1.1332
#>  -1.9360  0.6631 -0.9472 -0.7976  2.6440
#>  -3.4536  5.0837  0.2088  4.5133  0.9242
#> 
#> (2,.,.) = 
#>  -0.2362 -0.0399 -0.3736 -1.6955 -0.8777
#>  -0.2538 -1.2092 -0.4619 -1.7898 -1.3430
#>   2.3165  1.1142  1.5764  5.2953  5.0827
#> 
#> (3,.,.) = 
#>  -0.7541  1.1137  0.2242  1.8689  0.3395
#>   0.4514  1.0453  0.7419  3.4991  3.0727
#>   2.3049  1.5718  1.8886  6.1099  3.7733
#> 
#> (4,.,.) = 
#>   2.2976 -1.2337  0.8892  2.4286  3.4115
#>   1.6367 -1.3659  0.5335  1.0485  1.3199
#>   0.2201  0.3192 -0.0449 -0.7478 -0.2177
#> 
#> (5,.,.) = 
#>   0.1364 -3.2884 -1.5017 -7.7241 -4.9133
#>  -2.3979  6.5681  1.2521  7.5717  1.9143
#>   0.5450  2.3262  1.2429  5.5739  4.1134
#> 
#> (6,.,.) = 
#>  -1.1072  0.9509 -0.1567  0.3379 -0.6104
#>   0.4660  0.3636  0.4348  1.2842  0.2706
#>   1.2140 -3.1069 -0.4475 -3.7123 -2.7787
#> 
#> (7,.,.) = 
#>  -0.2686  2.9896  1.0215  4.2546  0.7059
#>   1.6757 -1.5326  0.2982 -0.1532  1.0222
#>   1.3749  0.6912  0.8891  1.8421  0.0962
#> 
#> (8,.,.) = 
#>   1.7974 -3.1070 -0.3499 -3.6949 -1.5754
#>   2.0510 -6.0731 -1.4222 -8.6077 -3.4650
#>  -1.3050  1.7007  0.1952  3.2918  2.9553
#> 
#> (9,.,.) = 
#>   1.9542 -0.8065  0.7397  1.2122  1.0492
#>   2.7132  2.0940  2.6849  9.9967  6.4687
#>  -0.2929  2.0399  0.5925  2.2917 -0.3039
#> 
#> (10,.,.) = 
#>   0.6715  2.4531  1.4826  6.1799  3.3686
#>  -2.0479  1.8895 -0.3781  0.0768 -1.3579
#>   0.8799 -1.4383 -0.0791 -1.2511 -0.3364
#> [ CPUFloatType{10,3,5} ]