Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 2.06759
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.2848
#>  0.8719
#> -0.0396
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.4768 -1.7006  1.1715
#> -1.9424 -5.6618 -1.0390
#>  0.1627 -0.6561 -2.6345
#>  0.2978 -1.7096  1.4634
#> -1.5827 -1.5826  2.4791
#> -1.6526  3.2332  0.4622
#>  2.1358  1.1276 -3.5155
#> -4.5856 -0.6102  0.0870
#>  3.0304  6.4178  2.2323
#> -3.0041  0.6452  2.7746
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.0539 -0.2866  0.0295 -0.0396  0.0103
#>  -0.6292  1.4872 -2.6785 -1.2843 -2.2077
#>  -0.8611  2.1514 -0.3936 -0.5721 -1.7132
#> 
#> (2,.,.) = 
#>  -0.3513 -2.2904 -0.1822  2.8763 -0.8499
#>   0.7189  1.9953 -0.8196 -0.6737  0.9592
#>   0.0890 -8.7816 -0.2492  9.6243 -2.6810
#> 
#> (3,.,.) = 
#>   2.4766  0.1759 -2.4997  0.0629 -1.6155
#>   2.7999  1.3079 -0.6778  2.1292 -2.4193
#>  -2.9410  1.1882  0.0512 -3.0437  3.0630
#> 
#> (4,.,.) = 
#>   3.8823  2.4800 -5.4583 -3.2447 -4.5206
#>  -1.1845 -0.3538  1.4496  0.6629  0.4697
#>  -0.0416 -1.4690  2.7877  1.2304 -4.8752
#> 
#> (5,.,.) = 
#>   0.1331 -2.2795 -0.1547  2.1540  0.2320
#>   0.2741  1.9374  0.3223  0.4836 -0.4837
#>   0.5879  4.3922  2.4263 -7.2309 -0.8230
#> 
#> (6,.,.) = 
#>   3.8461  5.2160 -0.5015  0.7308  2.1615
#>   0.7155 -3.4053  1.4277  0.5774  3.9904
#>  -2.0589 -3.9885  1.7712 -0.7630 -1.6973
#> 
#> (7,.,.) = 
#>  -2.5465 -0.4045  2.3311 -0.4129 -1.6999
#>   1.0029  2.7592 -2.5042 -1.4726 -0.3807
#>  -0.9501  0.3773  2.1664  0.6311 -0.9165
#> 
#> (8,.,.) = 
#>   0.3106  1.0825 -0.4134 -0.8231  0.8917
#>  -3.3255 -0.3793  0.5299 -1.4969  0.2216
#>   2.1510 -0.6342 -0.1332  1.0758 -0.6575
#> 
#> (9,.,.) = 
#>   0.5586  0.2560  0.5475 -0.3507 -0.7999
#>   0.9423  0.4434  1.0499 -2.1670 -1.3237
#>  -2.9847 -1.8735 -2.0485  3.6034  6.6753
#> 
#> (10,.,.) = 
#>   1.8594 -1.7428 -1.7508  3.2527 -1.1294
#>   0.9329  0.2012  1.6563  2.2632  1.3570
#>  -1.1695 -1.5044  0.2581 -2.5555 -1.2004
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -0.0986 -2.7771  0.4812 -1.7455  1.7574
#>   0.2745  0.5832 -1.4949 -3.1257  4.0068
#>   0.6302 -1.3836 -0.0548 -0.3557  1.3324
#> 
#> (2,.,.) = 
#>  -0.2069  1.7264  1.5385  1.8599 -0.4851
#>  -0.1145  3.7590 -0.7873  0.8731 -0.7697
#>  -1.6039  3.6536  6.0994  4.7015 -1.2596
#> 
#> (3,.,.) = 
#>   0.6779  1.4599 -0.2695  0.1581  1.8489
#>  -0.1786  0.6953  2.6484  1.4565  1.2168
#>   1.3883 -3.4545 -2.0996 -1.2917  0.8130
#> 
#> (4,.,.) = 
#>   1.5357 -0.5205 -1.5583 -0.1359  1.6278
#>  -0.2653 -0.3733 -0.8223 -1.3026  0.1159
#>   0.2296 -0.0721 -0.5383 -1.5488  2.3043
#> 
#> (5,.,.) = 
#>   0.0077  0.6587 -1.0501 -1.5151  1.3819
#>   1.3082 -1.2052 -0.7373 -0.3200  2.1227
#>   1.0470  0.9274 -2.2215 -0.8925  1.4970
#> 
#> (6,.,.) = 
#>  -0.6002 -1.0286  2.1541 -1.1602  2.8005
#>  -0.2366 -0.8338  0.6076  0.0178 -0.2297
#>  -2.8943  6.4717  8.7468  7.8544 -4.4328
#> 
#> (7,.,.) = 
#>  -0.8129 -0.2057  1.1431 -0.2854 -0.1016
#>   0.9622 -1.5345 -0.3707  0.3918  0.4329
#>   0.2507  0.6210  3.2208  4.7310 -2.2725
#> 
#> (8,.,.) = 
#>   0.9698  3.5743 -0.3689  2.7375 -0.6076
#>   0.4724 -3.1863  1.5586  0.5882  0.6644
#>  -0.5195  1.7220 -0.7452 -0.6437 -0.3567
#> 
#> (9,.,.) = 
#>  -1.1456  1.7803  0.7671  1.2554 -2.8838
#>  -0.8974  1.9692 -1.6969 -3.4238  2.0287
#>  -0.7244  3.1028  0.6513  1.7162 -2.0288
#> 
#> (10,.,.) = 
#>  -0.1139  2.8504 -0.8858  1.2851 -2.0598
#>  -0.5184  0.5568  1.2536  0.2297  0.3602
#>  -0.8788  1.3710  4.2493  2.9018 -0.3061
#> [ CPUFloatType{10,3,5} ]