Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -2.62571
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -0.1207
#>  0.1279
#> -1.1572
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.8952  3.4039  1.2125
#> -0.6867 -0.2336  0.8576
#> -1.2697  1.3498  0.3836
#>  2.9348  1.2356  0.4581
#> -0.9789 -0.6226  0.6313
#>  0.1534 -1.4413  1.8485
#> -0.8953  0.9535  1.3598
#> -1.6942  0.3796 -0.3817
#>  2.3217  1.2971 -1.5568
#> -2.4917 -0.6430 -1.5462
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.6045 -0.1054  1.1283 -3.1805 -5.1953
#>  -0.9326 -0.1358  0.3358  0.6043  0.0242
#>   1.0985  1.4012  0.0167 -0.7036 -0.4516
#> 
#> (2,.,.) = 
#>   0.6107 -1.8649 -1.1641  2.5378 -6.2422
#>   1.3437  1.3239  1.1525 -2.6677  6.7791
#>  -0.8055 -0.1036  1.9862 -1.8690  1.2423
#> 
#> (3,.,.) = 
#>  -1.1780  1.4065  0.4994  1.1855 -0.8005
#>   3.2373 -1.7488  2.1491 -1.1241  1.7923
#>  -1.2876  1.6321 -2.5498 -0.5707 -1.0112
#> 
#> (4,.,.) = 
#>  -0.5591 -0.3348  1.6107 -1.7421  0.6428
#>   2.6247  3.6589 -0.0818  3.2463  0.2822
#>  -0.2827  1.7873  0.9064  1.0344 -0.2768
#> 
#> (5,.,.) = 
#>  -1.3754  0.1413  1.4892  0.4109  0.8156
#>  -1.4186  0.1705  1.6027  0.1202  1.1788
#>   1.2356 -0.0465 -1.4835  0.9591 -1.2089
#> 
#> (6,.,.) = 
#>   0.6538  1.6152  0.1520  0.0596  3.0978
#>   0.8460 -0.9068  0.7023  1.7364  2.4896
#>  -0.2318  1.2969 -1.6292  1.9012  0.8973
#> 
#> (7,.,.) = 
#>  -0.1634  0.1724 -0.7327  1.1595 -1.9923
#>  -0.4786  0.4306 -0.1128 -2.5454  0.0288
#>  -1.8896  3.2717 -1.7228  0.3071  0.6750
#> 
#> (8,.,.) = 
#>   5.1517  2.7278  1.8985  1.5803 -6.0893
#>  -0.6200  1.5674 -1.8419  2.7221  4.0059
#>  -2.4522 -0.9122 -1.4661  0.6882  3.2389
#> 
#> (9,.,.) = 
#>  -1.6802  2.5560 -0.5219 -2.2781 -1.7557
#>  -1.2454  1.6475 -1.0472 -2.2255 -2.2989
#>   0.1268 -0.0380  0.7302  1.2396  2.0606
#> 
#> (10,.,.) = 
#>  -3.2564 -0.2378 -2.1333 -2.5627  1.7778
#>   1.5476 -0.6952 -0.8552  0.1398  1.2307
#>   1.2315  0.2015 -0.1805  0.4855 -0.3327
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -3.6953 -0.3724 -2.6328 -1.2967 -5.6964
#>  -1.3311 -0.3637 -1.3775 -0.9073 -3.5334
#>   0.9588  0.2622 -2.8397  0.9454  0.9809
#> 
#> (2,.,.) = 
#>   4.9251  0.1830  3.3294  2.6317  3.4501
#>  -0.3933  0.2507 -3.5790 -0.3887  1.3355
#>  -1.6763 -0.3618 -4.9047 -2.3539 -2.7403
#> 
#> (3,.,.) = 
#>   2.1481  0.5387  2.8550  1.4684  5.4611
#>  -2.3193 -0.0255 -0.6040 -0.5477 -1.8384
#>  -0.9951  0.0748 -1.6722 -0.7458  0.3052
#> 
#> (4,.,.) = 
#>   0.3260  0.3157  0.7233  0.4383  2.6424
#>  -1.9282 -0.4253 -0.7227 -1.3729 -3.7662
#>  -4.0462 -0.7673 -3.1122 -2.0560 -8.5062
#> 
#> (5,.,.) = 
#>  -2.0710 -0.2678  1.8686 -0.7034 -2.7571
#>  -2.0854  0.0609  1.9348 -0.1746 -0.6598
#>   1.0001  0.5342  1.8278  0.9861  4.7210
#> 
#> (6,.,.) = 
#>  -1.2038 -0.2238 -2.1981 -1.7588 -1.1597
#>   0.5001 -0.0816 -0.0002  0.6714 -1.2286
#>   2.8301  0.4616  0.8849  1.1980  5.2681
#> 
#> (7,.,.) = 
#>  -1.1943  0.0200 -2.8150 -0.4470 -1.1502
#>  -1.2177 -0.3866 -2.5951 -0.4379 -4.6826
#>  -1.2391 -0.3359  1.1064 -0.9573 -2.3932
#> 
#> (8,.,.) = 
#>   0.4282  0.5375 -0.6067  0.6360  3.9971
#>   2.0814 -0.2276 -0.5529  1.2366 -1.9765
#>  -0.0440  0.1171  0.9140  1.1852 -0.5119
#> 
#> (9,.,.) = 
#>  -3.5188 -0.1104 -5.8383 -2.0843 -3.0050
#>   2.1135  0.0827  0.1943  1.0108  1.3292
#>   0.0914 -0.6402  0.5549  0.0742 -5.4642
#> 
#> (10,.,.) = 
#>   2.1886  0.2266 -0.1836  1.4918  1.8344
#>   1.7545  0.4085  2.2214  1.7614  3.3148
#>  -5.4891 -0.4227 -4.6647 -2.2210 -7.1021
#> [ CPUFloatType{10,3,5} ]