Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> 7.03295
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.5741
#> -0.3255
#> -0.1159
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  2.6223  3.7030  0.7400
#>  1.8945  1.7939  0.5778
#>  0.6156  1.6131  0.2881
#>  1.5101 -2.1566  1.8846
#> -1.1521  1.6962  0.0883
#>  2.8449 -1.6660  0.4277
#> -0.3381 -0.0316  2.4233
#>  0.9262 -1.8355 -0.1359
#>  1.2348  2.5609 -0.0260
#>  1.2292 -0.4373 -3.6828
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   0.4088  0.1532  0.5610  0.2375  2.3070
#>   5.0677 -3.8798 -1.4565  2.8532  2.1777
#>   1.1134 -1.6993  0.6576  0.6649  0.5714
#> 
#> (2,.,.) = 
#>  -0.7629 -0.0140  0.7441  1.8525  3.2699
#>  -2.0508 -1.1491 -0.2896  2.2856  2.2808
#>  -0.2248 -0.9501 -0.8352 -0.6542 -0.9856
#> 
#> (3,.,.) = 
#>  -1.7365 -1.5915 -2.8468  2.8660  1.6813
#>   2.2818 -1.9416  1.3468 -0.6600 -1.3307
#>   2.5837 -3.5961 -0.4850 -0.5974 -0.9778
#> 
#> (4,.,.) = 
#>  -2.2726 -3.4006 -0.8556  2.8663  0.5622
#>  -2.1224 -2.5102 -0.7348  2.3526  0.2281
#>  -0.2618  1.1910  1.3918 -0.1959 -0.6392
#> 
#> (5,.,.) = 
#>  -4.2210  0.2733 -4.0233 -0.0566 -1.0297
#>   0.5414 -1.1539  0.2356  1.9339 -1.0352
#>  -0.0858  5.2520 -0.6881 -3.3668  4.5160
#> 
#> (6,.,.) = 
#>   2.0427 -4.6711 -0.9279 -3.5877  1.0178
#>  -0.2579 -1.2817 -0.6335 -0.1488 -0.1205
#>  -0.1640 -1.9865  0.4093 -0.7915 -0.3587
#> 
#> (7,.,.) = 
#>   0.0914  2.6302 -0.2983 -3.6634  0.0232
#>   0.3353 -0.6393  0.2954  1.6553 -0.6759
#>   2.5724  5.1952  3.7466 -0.4029  1.6230
#> 
#> (8,.,.) = 
#>  -0.4845  0.9014 -1.2108 -1.3779  1.1286
#>   1.2553 -0.0257  2.8264  1.4358 -0.4614
#>   1.2884 -5.6040  5.1507 -3.9004 -1.3168
#> 
#> (9,.,.) = 
#>   0.2361  0.1529  0.7093  1.5741 -0.5132
#>   0.3392  0.0696  0.2525 -0.6644 -0.9361
#>   2.7876 -1.9153 -0.1039 -1.1424  4.5677
#> 
#> (10,.,.) = 
#>  -4.0625  0.0353  1.6240 -0.5578  2.2609
#>  -3.1660 -0.7539 -0.7659 -3.1803  1.1917
#>   4.2351  0.1397  0.8556  1.2357  0.6936
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>  -2.5098  0.0634  1.0491 -3.9151  0.6521
#>   0.4394  4.3108 -0.7996 -4.4151 -0.8135
#>   2.2842  1.1117  0.5981  1.2899 -1.0483
#> 
#> (2,.,.) = 
#>   3.2917 -6.6021  1.3978  9.8952  1.2203
#>   1.3501 -0.6079  1.0953  0.5297  0.4463
#>   0.1352 -1.0133 -1.3991  2.8333 -0.0774
#> 
#> (3,.,.) = 
#>   1.0817 -1.7340  1.1073  1.5152  0.7851
#>  -3.4260  2.9031  0.8958 -7.1271 -0.5174
#>  -0.0789  2.7523 -0.2887 -3.7215 -0.2265
#> 
#> (4,.,.) = 
#>  -1.1913  0.6287  0.3415 -1.7521 -0.2892
#>   1.1896  2.4506 -1.8225 -0.2677 -0.7583
#>  -0.6840  1.0205 -0.7866 -0.9287 -0.3960
#> 
#> (5,.,.) = 
#>   1.8299 -1.7174 -0.7828  2.9720  1.1485
#>   1.0890 -0.9325  0.1591  0.7800  0.9352
#>  -1.6227  2.5735  0.1445 -4.2486 -0.7427
#> 
#> (6,.,.) = 
#>  -0.5949  1.2682 -0.4909 -2.7802  0.4777
#>  -0.9049  2.1152 -1.1571 -3.1277  0.0145
#>   1.6571 -2.1690 -0.1425  4.2408  0.3556
#> 
#> (7,.,.) = 
#>   2.7925  1.7064 -1.6260  2.5458 -1.1003
#>   2.0847  2.9820 -1.5031 -0.8216 -0.6081
#>   0.3319 -2.9279  0.7940  3.4804  0.4482
#> 
#> (8,.,.) = 
#>   0.2108 -3.4509  0.8332  3.5125  0.8721
#>   1.6069 -1.8168  0.9553  2.1406  0.8137
#>   3.7405  4.0781 -1.7291  0.5118 -1.5970
#> 
#> (9,.,.) = 
#>  -1.0046  2.0872 -1.6321 -2.4205 -0.2418
#>   1.9755 -2.4887 -0.4929  4.9841  0.5337
#>  -1.2647  2.2357 -0.9394 -4.2819  0.3535
#> 
#> (10,.,.) = 
#>  -2.5783  3.5085 -0.1261 -6.3896 -0.6535
#>  -1.2248  0.0398  1.7299 -2.8264  0.3083
#>   3.4728 -2.1352 -0.6205  6.3201  0.1611
#> [ CPUFloatType{10,3,5} ]