Matmul

Matmul

Arguments

input
(Tensor) the first tensor to be multiplied

other
(Tensor) the second tensor to be multiplied

out
(Tensor, optional) the output tensor.

Note

The 1-dimensional dot product version of this function does not support an `out` parameter.

matmul(input, other, out=None) -> Tensor

Matrix product of two tensors.

The behavior depends on the dimensionality of the tensors as follows:

If both tensors are 1-dimensional, the dot product (scalar) is returned.

If both arguments are 2-dimensional, the matrix-matrix product is returned.

If the first argument is 1-dimensional and the second argument is 2-dimensional, a 1 is prepended to its dimension for the purpose of the matrix multiply. After the matrix multiply, the prepended dimension is removed.

If the first argument is 2-dimensional and the second argument is 1-dimensional, the matrix-vector product is returned.

If both arguments are at least 1-dimensional and at least one argument is N-dimensional (where N > 2), then a batched matrix multiply is returned. If the first argument is 1-dimensional, a 1 is prepended to its dimension for the purpose of the batched matrix multiply and removed after. If the second argument is 1-dimensional, a 1 is appended to its dimension for the purpose of the batched matrix multiple and removed after. The non-matrix (i.e. batch) dimensions are broadcasted (and thus must be broadcastable). For example, if input is a \((j \times 1 \times n \times m)\) tensor and other is a \((k \times m \times p)\) tensor, out will be an \((j \times k \times n \times p)\) tensor.

Examples

# vector x vector
tensor1 = torch_randn(c(3))
tensor2 = torch_randn(c(3))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.5543
#> [ CPUFloatType{} ]

# matrix x vector
tensor1 = torch_randn(c(3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> -1.0748
#>  0.0295
#> -1.3756
#> [ CPUFloatType{3} ]

# batched matrix x broadcasted vector
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#>  0.5941  0.8701 -0.2910
#>  0.4488  0.0754  1.3375
#> -1.2592  1.8409  0.5066
#>  0.6234 -0.2534  0.8752
#> -0.5142  1.2403  1.7472
#> -1.2152  0.1999  0.2109
#>  1.0808 -0.0185  0.0017
#>  0.1728  0.9394 -1.1047
#>  0.6742  0.0913  1.2810
#> -0.0093 -0.2898  0.8128
#> [ CPUFloatType{10,3} ]

# batched matrix x batched matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(10, 4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   -4.9010   1.2752  -1.9103  -0.7308  -1.6892
#>   10.0931  -0.5024   1.8872   3.2587  -0.4530
#>   -3.1845  -0.3745  -2.1631   0.8955   0.9524
#> 
#> (2,.,.) = 
#>  -0.0195  0.2336 -0.1648 -1.1423 -0.1834
#>  -0.0492 -1.1049 -0.4440 -1.5038  0.6879
#>   0.3193  1.6756 -0.6901 -0.7054  0.0836
#> 
#> (3,.,.) = 
#>   2.9058 -4.0333 -4.1481  2.0654  0.7059
#>  -4.0311  7.1441  0.9818 -4.7148  0.7912
#>  -2.1133  2.4464  0.9383  2.0209 -0.3702
#> 
#> (4,.,.) = 
#>   2.8855  0.6784 -3.8542  0.4534  2.8369
#>   3.8723 -0.1521 -0.2855  1.6627  4.4765
#>  -1.9402 -0.1551  2.6832 -1.0236 -0.1508
#> 
#> (5,.,.) = 
#>  -0.8675 -1.0538  1.7580 -0.8748  0.2218
#>   1.0755  0.5115  2.3009 -3.2440  2.9478
#>  -1.5155 -1.2206 -1.6970  2.6380 -2.9842
#> 
#> (6,.,.) = 
#>   0.7153  2.8359 -0.4342 -1.5048 -1.6542
#>   0.3879  2.5230 -0.1644 -1.5833 -1.5019
#>   0.7879  2.5858  0.6381 -0.8508  0.2870
#> 
#> (7,.,.) = 
#>   0.7594  0.4043 -2.7816  1.1153 -1.0090
#>   3.1190 -1.0716 -2.6517  1.6220 -3.4661
#>  -4.6712  2.7996 -4.1126  0.8684  4.4905
#> 
#> (8,.,.) = 
#>  -0.8368  0.7409  1.2285 -0.0344 -0.1216
#>  -2.5213  1.4233  5.0653  0.2900 -2.7408
#>  -1.2304 -0.8387  0.9760 -0.1214 -0.2982
#> 
#> (9,.,.) = 
#>   0.6302  1.9668  2.1260 -3.5661  3.1188
#>  -0.6115 -0.9723 -2.1853  1.5044 -2.4648
#>   0.2309 -0.6932 -1.0314  1.0734 -0.9127
#> 
#> (10,.,.) = 
#>  -0.6085 -1.4543  0.9416 -0.9075  1.1657
#>  -0.1576  2.4780 -1.1478  0.0162 -2.6088
#>   1.8548  1.8830  0.3853  1.0931 -1.1246
#> [ CPUFloatType{10,3,5} ]

# batched matrix x broadcasted matrix
tensor1 = torch_randn(c(10, 3, 4))
tensor2 = torch_randn(c(4, 5))
torch_matmul(tensor1, tensor2)

#> torch_tensor 
#> (1,.,.) = 
#>   3.4214 -1.2106 -0.6298  4.1462 -1.6306
#>   1.2793  0.1481 -1.0392  2.2621 -2.3936
#>  -2.2713 -0.2334 -2.1545  2.4741  0.3863
#> 
#> (2,.,.) = 
#>  -0.8221 -1.3350  1.8479 -2.9218  5.1041
#>   0.2587 -1.2589 -0.8096  1.1778  1.8011
#>   0.5955  0.5769  1.6347  0.1286 -1.2819
#> 
#> (3,.,.) = 
#>  -0.7440 -0.2037 -0.7547 -1.5726  1.4278
#>   1.7038 -0.6690  2.0003  0.7086  0.4786
#>  -1.8517 -0.1515  0.1199 -1.2977  2.1332
#> 
#> (4,.,.) = 
#>   1.9669  0.5846  4.5887 -2.0836 -0.5198
#>   2.3897  0.2687 -2.0213  3.4715 -4.1437
#>  -2.2119 -1.0653 -8.9694  9.0768 -2.4080
#> 
#> (5,.,.) = 
#>   1.7037  0.4159  1.0114  0.6847 -2.0591
#>  -2.8276 -0.3166 -1.5908 -0.4944  2.3620
#>   2.6359 -0.6674  1.1492  1.8629 -0.8457
#> 
#> (6,.,.) = 
#>   1.7514  0.3400  1.1350  0.3517 -1.7532
#>  -3.6329 -1.0829  0.8851 -1.8361  5.7042
#>   0.0827  0.5083  0.8891 -4.8145  1.1828
#> 
#> (7,.,.) = 
#>   2.3415  0.4653  3.3203 -1.8986 -0.8973
#>  -0.6047  0.7703 -0.4558 -0.8566 -0.9788
#>  -2.0909  0.4261 -0.0260 -3.4066  1.9474
#> 
#> (8,.,.) = 
#>  -1.6201 -1.0116 -4.6657  5.1869 -0.1742
#>   0.9228 -0.4858  0.9489  1.1449  0.1569
#>  -3.1509  0.2783  2.3634 -6.6384  4.9453
#> 
#> (9,.,.) = 
#>   2.6801  0.1688 -2.3664  4.4283 -4.6229
#>  -0.6608 -0.1312  1.5208 -0.6322  1.3551
#>  -0.5575  1.1899 -1.7771  0.4593 -2.7964
#> 
#> (10,.,.) = 
#>   0.9173 -0.4096  0.9089 -1.1496  0.9921
#>  -0.1213  0.2654 -0.4653  1.7980 -1.3837
#>  -0.4864 -1.1666 -1.8596  3.4988  0.8258
#> [ CPUFloatType{10,3,5} ]